% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyDOID.R
\name{access_DOrepo}
\alias{access_DOrepo}
\title{Access DOrepo (INTERNAL)}
\usage{
access_DOrepo(x)
}
\arguments{
\item{x}{A 'pyDOID.repo.DOrepo' object or the path to a local copy of the
HumanDiseaseOntology repository.}
}
\description{
Access \code{x} as a 'pyDOID.repo.DOrepo' by confirming it's class, if already
instantiated, or instantiating the object from a given path.
}
\seealso{
DOrepo
}
\keyword{internal}
