% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyDOID.R
\name{access_owl_xml}
\alias{access_owl_xml}
\title{Access OWL/RDF XML (INTERNAL)}
\usage{
access_owl_xml(x)
}
\arguments{
\item{x}{A 'pyDOID.owl.xml' object or the path to an OWL/RDF XML file that
can be instantiated as such an object by \code{\link[=owl_xml]{owl_xml()}}.}
}
\description{
Access \code{x} as a 'pyDOID.owl.xml' by confirming it's class, if already
instantiated, or instantiating the object from a given path.
}
\keyword{internal}
