% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_empty_col}
\alias{append_empty_col}
\title{Append Empty Columns}
\usage{
append_empty_col(df, col, order = FALSE)
}
\arguments{
\item{df}{A data.frame.}

\item{col}{The name(s) of one or more columns desired in the final
data.frame, as a character vector. Any names not currently in the
data.frame will be added as empty columns; those present will remain
unchanged.}

\item{order}{Whether to reorder the data.frame to match the order of \code{col},
as a boolean (default: \code{FALSE}). If \code{FALSE} empty columns are added to the
right. When \code{order = TRUE}, \code{col} is used to specify column order. Thus,
in addition to the names of the empty columns to append, \code{col} must
include all column names in \code{df}. \code{append_empty_col()} will not
subset/select columns.}
}
\description{
Appends empty columns (value = \code{NA}) to a data.frame.
}
