% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{append_to_url}
\alias{append_to_url}
\title{Append to URL}
\usage{
append_to_url(url, x)
}
\arguments{
\item{url}{A URL or the internal name of a URL used in this package, as a
string. \emph{See \code{\link[=get_url]{get_url()}} for the names of available internal URLs.}}

\item{x}{Value to append, as a character vector.}
}
\description{
Append a value to a URL.
}
\section{Note}{

No URL validation is performed, but \code{append_to_url} will ensure that \code{url}
and \code{x} are concatenated with a single \code{/}.
}

