% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\encoding{UTF-8}
\name{as_html_img}
\alias{as_html_img}
\title{Construct HTML \code{img} Tag(s)}
\usage{
as_html_img(src, alt, ..., quote = "\\"")
}
\arguments{
\item{src}{The image source(s), as a character vector.}

\item{alt}{The alternate text, as a character vector.}

\item{...}{Named character vector(s) or name-value pairs of HTML attributes.
Attributes with values of \code{NULL} or \code{NA} will be dropped. Names should
correspond exactly to the desired HTML attributes.

For historical reasons, input may also be a single character vector of
name-value pairs.}

\item{quote}{The quote with which to surround the attribute values. Use \code{""}
if quotes are not desired (default: \verb{"\\""}, double quote).}
}
\value{
A character vector of HTML \code{img} tags.
}
\description{
Vectorized construction of one or more HTML \code{img} tags with the specified
attributes. \code{src} & \code{alt} are required. Where they are missing, no \code{img}
tag will be created. Additional attributes should be the same length as \code{src}
or length <= 1. Length‑1 attributes are recycled.
}
