% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble-methods.R
\name{as_tibble.esummary_list}
\alias{as_tibble.esummary_list}
\alias{as_tibble.esummary_list_nested}
\title{Convert \code{esummary} Object into Tibble}
\usage{
\method{as_tibble}{esummary_list}(x, ...)

\method{as_tibble}{esummary_list_nested}(x, ...)
}
\arguments{
\item{x}{An \code{esummary} object (\code{esummary_list} or \code{esummary_list_nested})}

\item{...}{Ignored; included for extensibility.}
}
\value{
An untidy, where \code{esummary_id} is the input identifier. \code{esummary_list_nested}
objects will have an additional \code{cites} column.
}
\description{
Converts an \code{esummary} object into a \link[tibble:tibble]{tibble}.
}
\section{Note}{

For single inputs to \code{\link[rentrez:entrez_summary]{rentrez::entrez_summary()}}, \code{always_return_list} must
be \code{TRUE}.
}

\seealso{
citedby_pubmed, pubmed_summary
}
