% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_AGR_entity.R
\name{assign_entity_to_mod}
\alias{assign_entity_to_mod}
\title{Assign AGR Entities to MODs}
\usage{
assign_entity_to_mod(AGR_df)
}
\arguments{
\item{AGR_df}{a dataframe derived from AGR data (usually a
\href{https://www.alliancegenome.org/downloads}{downloaded .tsv file})}
}
\value{
The input dataframe with an additional column (\code{mod_assigned}) indicating the
assignment.
}
\description{
Assigns Alliance of Genome Resource (AGR) entities to the appropriate model
organism database (MOD) using a step-wise approach. AGR entities may be
"sourced" from the "Alliance" obscuring which MOD the annotations were
originally made in.
}
\details{
The step-wise approach used to assign entities is as follows:
\enumerate{
\item \code{Source} is used where it pertains to a MOD (and not the "Alliance");
\strong{Note:} "OMIM Via RGD" is assigned to RGD
\item \code{DBObjectID} is used where it comes from a MOD (i.e. not "HGNC")
\item \code{WithOrthologs} is used where the ID comes from a MOD
\item Any remaining entities are left assigned to the "Alliance" because their
original source can not be identified
}
}
