% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alliance.R
\name{assign_record_to_mod}
\alias{assign_record_to_mod}
\title{Assign Alliance Records to MODs}
\usage{
assign_record_to_mod(alliance_tbl)
}
\arguments{
\item{alliance_tbl}{a dataframe derived from Alliance data (usually a
\href{https://www.alliancegenome.org/downloads}{downloaded .tsv file})}
}
\value{
The input dataframe with an additional column (\code{mod_assigned}) indicating the
assignment.
}
\description{
Assigns Alliance of Genome Resource records to the appropriate model
organism database (MOD) using a step-wise approach. Alliance records may be
"sourced" from the "Alliance" obscuring which MOD the records were
originally made in.
}
\details{
The step-wise approach used to assign records is as follows:
\enumerate{
\item \code{Source} is used where it pertains to a MOD (and not the "Alliance");
\strong{Note:} "OMIM Via RGD" is assigned to RGD
\item \code{DBObjectID} is used where it comes from a MOD (i.e. not "HGNC")
\item \code{WithOrthologs} is used where the ID comes from a MOD
\item Any remaining records are left assigned to the "Alliance" because their
original source can not be identified
}
}
