% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_id_converter.R
\name{batch_id_converter}
\alias{batch_id_converter}
\title{Convert IDs in Batches}
\usage{
batch_id_converter(x, type = NULL, ...)
}
\arguments{
\item{x}{(character) One or more of: doi, pmid, pmcid, or manuscript id,
see examples. required.}

\item{type}{(character) one of doi, pmid, pmcid, or manuscript id}

\item{...}{Curl args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
Batches inputs to convert any number of IDs using PubMed Centrals ID
converter API, via \link[rcrossref:id_converter]{rcrossref::id_converter}.
}
