% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{build_html_element}
\alias{build_html_element}
\title{Build HTML Element(s)}
\usage{
build_html_element(
  tag,
  ...,
  content = NULL,
  close_empty = TRUE,
  quote = "\\"",
  include = "optional"
)
}
\arguments{
\item{tag}{The tag(s) to include, as a character vector. Length‑1 character
vectors will be recycled to the length of the}

\item{...}{Named character vector(s) or name-value pairs of HTML attributes.
Attributes with values of \code{NULL} or \code{NA} will be dropped. Names should
correspond exactly to the desired HTML attributes.

For historical reasons, input may also be a single character vector of
name-value pairs.}

\item{content}{The content to include within the tag(s), as a character
vector. The default, \code{NULL}, means no content is intended. For tags that are
required to be empty, content will be ignored. Length‑1 content will be
recycled.}

\item{close_empty}{Whether to close empty tags (e.g. \verb{<img>}) with a space
and slash (default: \code{TRUE} for compatibility with XHTML). For a reference,
see https://stackoverflow.com/questions/1946426/html-5-is-it-br-br-or-br.}

\item{quote}{The quote with which to surround the attribute values. Use \code{""}
if quotes are not desired (default: \verb{"\\""}, double quote).}

\item{include}{The type(s) of tag(s) to include from the output; one of
"optional" (default, including both optional & required tags) or "required"
(including \emph{ONLY} required tags).

Take extra care when using \code{include = "required"}. There are many cases where
a tag considered "optional" is actually required. \code{build_html_element()}
only handles two possible scenarios: (1) both tags are included when
\code{content} has a value, and (2) start tags are included when there are
corresponding attributes. To understand tag requirements, review this
\href{https://www.w3.org/TR/html401/index/elements.html}{table of HTML tags} and
the \href{https://www.w3.org/TR/2011/WD-html5-20110405/syntax.html}{HTML5 syntax}.}
}
\value{
A character vector of HTML elements, including start tags along with
attributes, content, and/or end tags as appropriate.
}
\description{
Vectorized construction of one or more HTML elements including the specified
tag(s), attribute(s) and content. Length‑1 inputs will be recycled to the
length of the longest input. All inputs with length > 1, must be of the same
length.
}
