% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{build_hyperlink}
\alias{build_hyperlink}
\title{Build Hyperlinks}
\usage{
build_hyperlink(x, url, as, ..., txt = x, preserve_NA = TRUE)
}
\arguments{
\item{x}{Value(s) to append, as a character vector.}

\item{url}{A URL or a URL name recognized by this package, as a string;
see \link{get_url} for recognized names.}

\item{as}{What format to use for the hyperlink, as a string; one of "gs"
(Google Sheet), "xlsx" (Excel), or "html".}

\item{...}{One or more name-value pairs of html \verb{<a>}
\href{https://www.w3schools.com/tags/tag_a.asp}{attributes}.}

\item{txt}{\emph{(OPTIONAL)} The text to display for each link, as a character
vector. The default uses \code{x} as the text. If \code{NULL}, the full URL will serve
as the text.  If a string, the value will be used for the text of each
hyperlink.}

\item{preserve_NA}{Whether to preserve \code{NA} in output, as a boolean. \code{FALSE}
will result in \code{NA} being appended to the end of \code{url} (almost certainly not
desired).}
}
\description{
Builds hyperlinks for Google Sheets, Excel, or HTML by appending a value to
the end of a base url.
}
\examples{
build_hyperlink(
    x = "allenbaron",
    url = "github",
    as = "html",
    txt = "A hyperlink!"
)

}
