% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{build_hyperlink_html}
\alias{build_hyperlink_html}
\title{Build an HTML Hyperlink (INTERNAL)}
\usage{
build_hyperlink_html(x, url, text, target = "_blank", preserve_NA = TRUE)
}
\arguments{
\item{x}{Value(s) to append, as a character vector.}

\item{url}{A URL or a URL name recognized by this package, as a string;
see \link{get_url} for recognized names.}

\item{text}{The text to display for the link, as a character vector.}

\item{target}{An html \verb{<a>}
\href{https://www.w3schools.com/tags/att_a_target.asp}{target} attribute.}

\item{preserve_NA}{Whether to preserve \code{NA} in output, as a boolean. \code{FALSE}
will result in \code{NA} being appended to the end of \code{url} (almost certainly not
desired).}
}
\description{
Builds a hyperlink formatted as html.
}
\examples{
build_html_hyperlink(x = "allenbaron", url = "github", text = "A hyperlink!")

}
\keyword{internal}
