% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast_to_robot_template}
\alias{cast_to_robot_template}
\title{Cast to ROBOT template}
\usage{
cast_to_robot_template(.df, ...)
}
\description{
Extracts new, curated data from a data frame and casts it into a ROBOT
template. This function is intended for an OBO Foundry ontology but could be
used more widely.
}
