% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_tld}
\alias{cc_tld}
\title{Country-code Web Domains}
\format{
A data frame with 247 rows and 3 columns:
\describe{
\item{domain}{DNS name of the two-letter country-code top-level domain. They follow ISO 3166-1 alpha-2, with some exceptions such as ".ac" for Ascension Island, ".eu" for the European Union, or ".uk" for United Kingdom of Great Britain and Northern Ireland instead of ".gb". ISO codes bv, bl, mf, sj, gb, and um are not used for country code top-level domains. }
\item{entity}{Country, dependency, or region }
\item{generic}{Whether \code{domain} is a technically "non-restricted ccTLD" used like traditional generic TLDs; ie. it may be used outside of \code{entity}}
}
}
\source{
Compiled by J. Allen Baron; last updated 2022-12-01.
}
\usage{
cc_tld
}
\description{
Country-code Top-level domain information from Wikipedia
\href{https://en.wikipedia.org/wiki/Country_code_top-level_domain}{Country code top-level domain}
article.
}
\keyword{datasets}
