% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{check_dl_status}
\alias{check_dl_status}
\title{Check Exit Status of Downloads (Internal)}
\usage{
check_dl_status(exit_code, dest_file, abort = FALSE)
}
\arguments{
\item{exit_code}{An integer vector of exit code(s) produced by
\code{\link[utils:download.file]{utils::download.file()}}.}

\item{dest_file}{A character vector of relative or absolute paths to
downloaded files. Used to make output and messages more informative; the
existence of these files is not tested.}

\item{abort}{A logical scalar indicating whether to abort when non-zero
exit codes are produced. When \code{FALSE}, (default) failed downloads will
produce warnings. \code{TRUE} will likely only be useful for single file
downloads.}
}
\value{
A character vector of paths to destination files that downloaded
successfully, with a warning indicating which fail when \code{abort = FALSE}
(default).
}
\description{
Check the return value when \code{\link[utils:download.file]{utils::download.file()}} is used, as recommended
in the "Good practice" section of it's documentation.
}
