% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{check_html_attr}
\alias{check_html_attr}
\title{Check HTML Attribute Inputs}
\usage{
check_html_attr(..., max_length = NULL)
}
\arguments{
\item{...}{Named character vector(s) or name-value pairs of HTML attributes.
Attributes with values of \code{NULL} or \code{NA} will be dropped. Names should
correspond exactly to the desired HTML attributes.

For historical reasons, input may also be a single character vector of
name-value pairs.}

\item{max_length}{The maximum expected length of attributes, as an integer. If
\code{NULL} (default), the maximum length will be the length of the longest
attribute.}
}
\description{
Checks that all HTML attribute inputs are valid, meaning (1) either
length <= 1 or the same length as the longest attribute, (2) that all
attributes are named, and (3) that no attributes are duplicated.
}
\keyword{internal}
