% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citedby.R
\name{citedby_pmid}
\alias{citedby_pmid}
\title{Get Cited By PMIDs from PubMed}
\usage{
citedby_pmid(
  id = NULL,
  web_history = NULL,
  by_id = FALSE,
  config = NULL,
  save_raw = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{id}{PubMed IDs for which to retrieve "cited by", as a character vector.}

\item{web_history}{An Entrez History object to be used instead of \code{id}, as
generated by the modified \code{\link[rentrez:entrez_post]{rentrez::entrez_post()}} fork at
https://github.com/allenbaron/rentrez (see NOTE); should be used when
"cited by" information is desired for >200 IDs.}

\item{by_id}{Whether to split "cited by" publications by the ID they
cite, as a boolean (default: \code{TRUE}); \code{FALSE} returns a unified list of
\emph{unique} cited by publications \emph{without} reference to the ID they cite.}

\item{config}{Configuration options passed to \code{\link[httr:GET]{httr::GET()}}.}

\item{...}{Additional terms to add to the request, see NCBI documentation on
\href{https://www.ncbi.nlm.nih.gov/books/NBK25499/#_chapter4_ESearch_}{ESearch}
for a complete list.}
}
\description{
List PubMed IDs of publications citing those specified by \code{web_history} or
\code{id}, potentially split \code{by_id}. \code{citedby_pmid()} can handle cases where the
number of input IDs is >200 automatically (unlike \link[rentrez:entrez_link]{rentrez::entrez_link}).
}
\section{NOTE}{

The rentrez fork at https://github.com/allenbaron/rentrez implements pull
request \href{https://github.com/ropensci/rentrez/pull/174}{#174}, which is not
yet implemented in the CRAN rentrez and is required for
rentrez::entrez_post() to function fully.
}

