% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{concat_pm_citation}
\alias{concat_pm_citation}
\title{Concatenate PubMed Citations (from txt file)}
\usage{
concat_pm_citation(file)
}
\arguments{
\item{file}{path to .txt file; or another possible input to \code{readLines()}}
}
\description{
Concatenates citations spanning multiple lines, usually obtained
by downloading a text file from PubMed
}
