% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confine_list.R
\name{confine_list}
\alias{confine_list}
\alias{release_list}
\title{Convert (Nested) List to/from Character Vector}
\usage{
confine_list(.list)

release_list(.json_chr)
}
\arguments{
\item{.list}{A (nested) list.}

\item{.json_chr}{A character vector containing the JSON representation of
(nested) list elements below the top level.}
}
\description{
These functions convert a (nested) list to or from a character vector of the
same length, preserving and restoring the list structure and values using
JSON strings. They do not perform \emph{any} flattening or simplification and the
JSON strings are verbose.
}
