% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot_wrappers.R
\name{convert_to_ofn}
\alias{convert_to_ofn}
\title{Convert Ontology Files to OFN}
\usage{
convert_to_ofn(path, out_path = NULL, gzip = FALSE, .robot_path = NULL)
}
\arguments{
\item{path}{The path to the ontology file, as a string.}

\item{out_path}{The path of the desired output OFN output file, as a string.
If \code{NULL} (default), a temporary file will be created.}

\item{gzip}{Whether output should be gzipped, as a boolean
(default: \code{FALSE}). If \code{TRUE}, the output file will end with a '.gz' file
extension.}

\item{.robot_path}{The path to a ROBOT executable or .jar file, as a string.
When \code{NULL} (default), if a system ROBOT executable is available it will
be used, otherwise an error will be signaled.

\strong{NOTE:} \code{DO.utils} caches the last ROBOT used for future use.}
}
\value{
The output path of the OFN file, invisibly.
}
\description{
Converts ontology files to OWL functional syntax (OFN) using
ROBOT (\url{http://robot.obolibrary.org/}).
}
\keyword{internal}
