% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_AGR_entity.R
\name{count_AGR_entity}
\alias{count_AGR_entity}
\title{Count AGR Entities}
\usage{
count_AGR_entity(AGR_df, by_type = TRUE, pivot = TRUE)
}
\arguments{
\item{AGR_df}{a dataframe derived from AGR data (usually a
\href{https://www.alliancegenome.org/downloads}{downloaded .tsv file})}

\item{by_type}{logical indicating whether to count by entity type}

\item{pivot}{logical indicating whether to pivot type to columns; ignored if
type = FALSE}
}
\value{
A summary tibble.
}
\description{
Counts entities in data from the Alliance of Genome Resources (AGR) by model
organism database (MOD) and, optionally, by entity type.
}
