% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_AGR_records.R
\name{count_AGR_records}
\alias{count_AGR_records}
\title{Count AGR Records}
\usage{
count_AGR_records(AGR_df, by_type = TRUE, pivot = TRUE)
}
\arguments{
\item{AGR_df}{a dataframe derived from AGR data (usually a
\href{https://www.alliancegenome.org/downloads}{downloaded .tsv file})}

\item{by_type}{logical indicating whether to count by object type}

\item{pivot}{logical indicating whether to pivot values to type columns;
ignored if type = FALSE}
}
\value{
A summary tibble.
}
\description{
Counts records in data from the Alliance of Genome Resources (AGR) by model
organism database (MOD) and, optionally, by object type. A record, as used
here, is a row in the data (after removing likely duplicates, see NOTE).
}
