% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alliance.R
\name{count_alliance_records}
\alias{count_alliance_records}
\title{Count Alliance Records}
\usage{
count_alliance_records(
  alliance_tbl,
  term_subset = NULL,
  by_type = TRUE,
  pivot = TRUE,
  record_lvl = "disease-object",
  assign_to = c("species", "curator")
)
}
\arguments{
\item{alliance_tbl}{a dataframe derived from Alliance data (usually a
\href{https://www.alliancegenome.org/downloads}{downloaded .tsv file})}

\item{term_subset}{character vector of DOIDs to limit counts to}

\item{by_type}{logical indicating whether to count by Alliance object type
(i.e. gene, allele, model)}

\item{pivot}{logical indicating whether to pivot values to type columns;
ignored if by_type = FALSE.}

\item{record_lvl}{a string indicating the desired specificity of records.}

\item{assign_to}{how to assign records when counting; one of "species" or
"curator" (i.e. the organization responsible for curating the record)}
}
\value{
A summary tibble with the count of unique object annotations defined by
\code{record_lvl}, aggregated according to species/curator (\code{assign_to}) and,
optionally, object type (\code{by_type}).
}
\description{
Counts records in data from the Alliance of Genome Resources. Counts can
be ascribed to the species the record is associated with or the Model
Organism Database (MOD) that curated it, optionally by object type.
\code{count_alliance_records()} was primarily designed to count records in the
Alliance Disease Associations File. \emph{There is no guarantee that any/all}
\emph{options will work for other files.}
}
\details{
The type of record information to use in counting should be specified with
\code{record_lvl} which accepts the following values:
\itemize{
\item "full_record" counts full non-duplicate records
\item "disease-object" counts unique disease-object combinations
\item "disease" counts unique diseases
\item "object" counts unique MOD objects (i.e. gene, allele, model identifiers)
}
}
\section{NOTE}{

For disease-related data, some exact duplicates (reason unknown) and records
that differ by seemingly unimportant information (e.g. only the date differs)
have existed. These types of duplicates are removed prior to record counts.
}

\seealso{
Other Alliance functions:
\code{\link[=download_alliance_tsv]{download_alliance_tsv()}},
\code{\link[=read_alliance]{read_alliance()}},
\code{\link[=save_alliance_counts]{save_alliance_counts()}}
}
\concept{Alliance functions}
