% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website-deprecated.R
\name{deprecated-website}
\alias{deprecated-website}
\alias{make_use_case_html}
\alias{make_user_list_html}
\title{Deprecated Website Functions}
\usage{
make_use_case_html(out_dir = "graphics/website", group = "all")

make_user_list_html(file)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=update_website_use_cases]{update_website_use_cases()}} instead of \code{make_use_case_html()} and
\code{make_user_list_html()}.
}
\keyword{internal}
