% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_AGR_entity.R
\name{dl_read_AGR}
\alias{dl_read_AGR}
\title{Download and Read AGR .tsv.gz files}
\usage{
dl_read_AGR(url, dest_dir)
}
\arguments{
\item{url}{URL to AGR file}

\item{dest_dir}{path to directory where file will be saved}
}
\value{
A dataframe.
}
\description{
Downloads a URL-specified .tsv.gz file from the Alliance of Genome
Resources (AGR) and reads that file into R as a tibble. Files can be found at
\url{https://www.alliancegenome.org/downloads}. Right-click on the "tsv" link
of a desired file and select "Copy Link" to get the file URL.
}
\details{
The file is downloaded for the purpose of reproducibility and future
reference. A date stamp will be added to the base file name.
}
