% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.utils-package.R
\docType{package}
\name{DO.utils-package}
\alias{DO.utils}
\alias{DO.utils-package}
\title{DO.utils: Support Management of the Disease Ontology (DO)}
\description{
This package is used to automate management of the Human Disease Ontology (DO). Functionality is currently limited to tracking citations but will eventually include automating preparation of terms and definitions to speed the curation process.
}
\keyword{internal}
