% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.utils.R
\docType{package}
\name{DO.utils}
\alias{DO.utils}
\title{Human Disease Ontology (DO) support package}
\description{
The DO.utils packages provides functions that assist in three main areas:
\enumerate{
\item Managing DO tools/builds
\item Assessing DO's impact on science (bibliometrics/scientometrics)
\item Making R programming simpler/faster
}
}
