% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DO_colors}
\alias{DO_colors}
\title{Human Disease Ontology Official Colors}
\format{
A named, vector of colors in hexadecimal format, including most
recognizable teal-ish colors (default, mid, light, website, & websafe) and
less used oranges (orange, orange_mid, orange_light).
}
\source{
Updated by J. Allen Baron on 2021-11-23.
}
\usage{
DO_colors
}
\description{
Named, hexadecimal colors for official DO use.
}
\keyword{datasets}
