% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DO_pubs}
\alias{DO_pubs}
\title{Human Disease Ontology (DO) Publication Info}
\format{
A data frame with 9 variables and 1 row for each official DO
publication:
\describe{
\item{internal_id}{short-version identifier used by the DO team for DO
publications}
\item{pmid}{PubMed ID}
\item{pmcid}{PubMed Central ID}
\item{doi}{DOI}
\item{scopus_eid}{Scopus Electronic Identification, not the same as a
Scopus ID}
\item{semantic_scholar_id}{Semantic Scholar Corpus ID}
\item{first_author}{Publication First Author}
\item{title}{Publication Title}
\item{citation_nlm}{full NLM-formatted citation}
}
}
\source{
Compiled by J. Allen Baron; last updated 2021-11-11
}
\usage{
DO_pubs
}
\description{
A dataset of identifiers for official DO publications, along with their
title and NLM-formatted citations.
}
\keyword{datasets}
