% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_AGR_records.R
\name{download_AGR}
\alias{download_AGR}
\title{Download AGR .tsv.gz File}
\usage{
download_AGR(url, dest_dir)
}
\arguments{
\item{url}{URL to AGR file}

\item{dest_dir}{path to directory where file will be saved}
}
\value{
Likely, an (invisible) integer code, 0 for success and non-zero for failure.
See Value section of \code{\link[utils:download.file]{utils::download.file()}} for more details.
}
\description{
Downloads a URL-specified .tsv.gz file from the Alliance of Genome
Resources (AGR). Files can be found at
\url{https://www.alliancegenome.org/downloads}. Right-click on the "tsv" link
of a desired file and select "Copy Link" to get the file URL.
}
\details{
A date stamp indicating download date is added to the base file name.
}
\section{Recommendation}{

Although it's possbile to directly read a file from the URL, downloading it
promotes reproducibility and ensures future access if needed.
}

