% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_alliance_tsv}
\alias{download_alliance_tsv}
\title{Download Alliance .tsv.gz File}
\usage{
download_alliance_tsv(
  dest_dir,
  url = "https://fms.alliancegenome.org/download/DISEASE-ALLIANCE_COMBINED.tsv.gz",
  ...
)
}
\arguments{
\item{dest_dir}{path to directory where file will be saved}

\item{url}{URL to Alliance file; the default is the Alliance's complete
disease-linked data file}

\item{...}{Additional arguments passed on to \code{\link[utils:download.file]{utils::download.file()}}.}
}
\value{
Path to saved file.
}
\description{
Downloads a URL-specified .tsv.gz file from the Alliance of Genome
Resources. Files can be found at
\url{https://www.alliancegenome.org/downloads}. Right-click on the "tsv" link
of a desired file and select "Copy Link" to get the file URL.
}
\details{
A date stamp indicating download date is added to the base file name.
}
\section{Recommendation}{

Although it's possbile to directly read a file from the URL, downloading it
promotes reproducibility and ensures future access if needed.
}

\seealso{
Other Alliance functions:
\code{\link[=count_alliance_records]{count_alliance_records()}},
\code{\link[=read_alliance]{read_alliance()}},
\code{\link[=save_alliance_counts]{save_alliance_counts()}}
}
\concept{Alliance functions}
