% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_obo_ontology}
\alias{download_obo_ontology}
\title{Download OBO Foundry Ontology File}
\usage{
download_obo_ontology(ontology_id, dest_dir, on_failure = "warn", ...)
}
\arguments{
\item{ontology_id}{A character vector of OBO Foundry ontology identifier(s)
(lowercase, as found on http://www.obofoundry.org/). For reference,
ontology identifiers are also provided in \link{obofoundry_metadata} within
this package.}

\item{dest_dir}{Path to directory where files will be saved.}

\item{on_failure}{A string indicating how to handle download failure:
\itemize{
\item "warn" - produce a warning; includes exit codes for debugging
\item "abort" - abort execution
\item "list_failed" - list URLs that failed (\emph{output format differs}, see
\code{Value})
\item "warn-list_failed" - combination of "warn" and "list_failed"
\item "skip" - do nothing
}}

\item{...}{Additional arguments passed on to \code{\link[=download.file]{download.file()}}.}
}
\value{
Unless \code{on_failure} includes "list_failed", the successfully downloaded
\code{dest_file}(s); otherwise, a 2-vector list where \code{successful} =
\code{dest_file}(s) and \code{failed} = \code{url}(s).
}
\description{
Downloads the current version of one or more ontologies from the OBO Foundry.
}
