% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_omim}
\alias{download_omim}
\title{Download OMIM Files}
\usage{
download_omim(omim_file, dest_dir, api_key = NULL, on_failure = "abort", ...)
}
\arguments{
\item{omim_file}{A character vector of one or more OMIM file name(s).}

\item{dest_dir}{Path to directory where files will be saved; all OMIM files
are saved in .txt format and consist of tab-separated values with a
copyright header.}

\item{api_key}{Your API key from OMIM, as a string; required if downloading
any file other than 'mim2gene.txt'.

\href{https://www.omim.org/downloads}{Register for downloads} at OMIM.}

\item{on_failure}{A string indicating how to handle download failure:
\itemize{
\item "warn" - produce a warning; includes exit codes for debugging
\item "abort" - abort execution
\item "list_failed" - list URLs that failed (\emph{output format differs}, see
\code{Value})
\item "warn-list_failed" - combination of "warn" and "list_failed"
\item "skip" - do nothing
}}

\item{...}{Additional arguments passed on to \code{\link[utils:download.file]{utils::download.file()}}.}
}
\value{
Unless \code{on_failure} includes "list_failed", the successfully downloaded
\code{dest_file}(s); otherwise, a 2-vector list where \code{successful} =
\code{dest_file}(s) and \code{failed} = \code{url}(s).
}
\description{
Downloads one or more specified files from \href{https://www.omim.org/}{OMIM}
(the Online catalog of Mendelian Inheritance in Man). \code{mim2gene.txt} is the
only file accessible without an API key from OMIM.
}
\seealso{
\code{\link[=read_omim]{read_omim()}} to read downloaded files as tibble/data.frames.
}
