% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_blank.R
\name{drop_blank}
\alias{drop_blank}
\title{Drop Blanks}
\usage{
drop_blank(x)
}
\arguments{
\item{x}{A character vector or list.}
}
\value{
An object of the same class and length as \code{x}.
}
\description{
Drops blank values from a vector or list. See \code{\link[=is_blank]{is_blank()}} for what
constitutes a blank value. \code{drop_blank()} is a generic function.
}
\examples{
drop_blank(c("", "A"))
drop_blank(
    list(
        c("", "A"),
        c("A", "B"),
        c("C", "D", "", "E", "")
    )
)

}
