% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyDOID.R
\name{establish_owl_xml}
\alias{establish_owl_xml}
\title{Establish OWL/RDF XML}
\usage{
establish_owl_xml(x)
}
\arguments{
\item{x}{A pyDOID.owl.xml object or the path to an OWL/RDF XML file that
can be instantiated as such an object.}
}
\description{
Establish that \code{x} is a pyDOID.owl.xml by confirming it's class, if already
instantiated, or instantiating the object from a given path.
}
\keyword{internal}
