% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_as_tidygraph}
\alias{extract_as_tidygraph}
\title{Extract OWL/RDF as a tidygraph}
\usage{
extract_as_tidygraph(x, query = NULL, collapse_method = "first", debug = FALSE)
}
\arguments{
\item{x}{A 'pyDOID.owl.xml' object or the path to an OWL/RDF XML file that
can be instantiated as such an object by \code{\link[=owl_xml]{owl_xml()}}.}

\item{query}{A SPARQL 1.1 query, as a string or the path to a .sparql/.rq
file. If \code{NULL}, input is assumed to be an OBO Foundry ontology and the
\code{oboInOwl:id} and \code{rdfs:label} of all OBO classes will be extracted. See
"Query Requirements" section for more information.}

\item{collapse_method}{The method to use when collapsing extra data columns,
as a string. See "Query Requirements" section for  more information.}

\item{debug}{\emph{[Intended for developers only]} Whether to debug execution.
Returns all intermediate objects as part of output.}
}
\description{
Extract 'nodes' and their 'parents' defined by a SPARQL \code{query} from OWL/RDF
as a tidygraph. Utilizes \code{\link[=robot]{robot()}}.
}
\section{Query Requirements}{

\link[tidygraph:tidygraph-package]{tidygraph} expects \emph{unique} child-parent
relationships, so at a minimum the SPARQL \code{query} should include \code{?id} (some
identifier for a 'child') and \code{?parent} (some identifier for each child's
parent(s)). All additional output variables specified in the SPARQL query
will be treated as 'node' annotations and collapsed using a method from
\code{\link[=collapse_col_flex]{collapse_col_flex()}} to prevent duplication of records.
}

