% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_class_axiom}
\alias{extract_class_axiom}
\title{Extract Class Axioms}
\usage{
extract_class_axiom(DO_repo)
}
\arguments{
\item{DO_repo}{The local path to the \code{HumanDiseaseOntology} repo, as a
string.}
}
\value{
A list of two character vectors (\code{eq} and \code{subclass}) containing axioms in
OWL functional format.
}
\description{
Extract \code{owl:equivalentClass} and \code{owl:subClassOf} axioms from the
doid-edit.owl file.
}
\seealso{
Other \verb{extract_*_axiom} functions:
\code{\link[=extract_eq_axiom]{extract_eq_axiom()}},
\code{\link[=extract_subclass_axiom]{extract_subclass_axiom()}}
}
\concept{\verb{extract_*_axiom} functions}
