% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_obo_anon}
\alias{extract_obo_anon}
\title{Extract Anonymous Relationships from OBO Foundry Ontology}
\usage{
extract_obo_anon(
  obo_ont,
  prefix = NULL,
  id = NULL,
  render = "label",
  .robot_path = NULL
)
}
\arguments{
\item{obo_ont}{The path to an ontology file, as a string.}

\item{prefix}{A character vector of OBO prefixes (aka ID spaces) to filter
results to, or \code{NULL} (default) to return all axioms. \emph{Ignored if \code{id} is}
\emph{provided.}}

\item{id}{A character vector of OBO IDs (CURIEs) to filter results to or
\code{NULL} (default) to return all entities with logical relations.}

\item{render}{The format for rendering classes & properties, as a string.
One of:
\itemize{
\item \code{"label"} (default): Use labels, quoting as needed.
\item \code{"id"}: Use OBO IDs (CURIEs).
}}

\item{.robot_path}{The path to a ROBOT executable or .jar file, as a string.
When \code{NULL} (default), if a system ROBOT executable is available it will
be used, otherwise an error will be signaled.

\strong{NOTE:} \code{DO.utils} caches the last ROBOT used for future use.}
}
\value{
A tibble with the columns: \code{id}, \code{data_type}, and \code{value}, where \code{value}
is the axiom in Manchester syntax rendered according to \code{format}.
}
\description{
Extracts all anonymous relations (logical/complex) from any OBO Foundry
Ontology in Manchester format, including equivalent classes, subclasses, and
disjoint classes. \code{extract_obo_anon()} is designed to supplement SPARQL
queries that generally cannot return anonymous relationships.
}
\section{NOTES}{

Uses \href{https://robot.obolibrary.org/export}{ROBOT export} internally.
}

