% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_obo_class}
\alias{extract_obo_class}
\title{Extract OBO Classes with Parents}
\usage{
extract_obo_class(
  input,
  class,
  method = "descendants",
  .robot_path = NULL,
  tidy_what = c("header", "uri_to_curie"),
  ...
)
}
\arguments{
\item{input}{The path to an RDF/OWL file recognized by ROBOT, as a string.}

\item{class}{One or more valid OBO classes (see \code{\link[=is_valid_obo]{is_valid_obo()}} for
valid input formats). Accepted values may depend on \code{method}.}

\item{method}{The method to use when extracting classes. Options include:
\itemize{
\item \code{"self"}: Extract the class(es) only.
\item \code{"parents"}: Extract the class(es) and \emph{asserted} parent(s).
\item \code{"descendants"} (default): Extract the class(es) and all descendants.
\item \code{"ancestors"}: Extract the class(es) and all ancestors.
\item \code{"common_ancestor"}: Extract the class(es) and all ancestors up to a common
ancestor. The common ancestor could be in the set.
}}

\item{.robot_path}{The path to a ROBOT executable or .jar file, as a string.
When \code{NULL} (default), if a system ROBOT executable is available it will
be used, otherwise an error will be signaled.

\strong{NOTE:} \code{DO.utils} caches the last ROBOT used for future use.}

\item{tidy_what}{The elements of a SPARQL-created data.frame to tidy, as a
character vector. One or more of the following:
\itemize{
\item \code{"everything"} to apply all tidy operations (has precedence over
\code{"nothing"}).
\item \code{"header"} to remove leading \verb{?} from header labels.
\item \code{"unnest"} to unnest list columns with \code{\link[=unnest_cross]{unnest_cross()}}.
\item \code{"uri_to_curie"} to convert all URIs recognized by DO.utils to CURIEs with
\code{\link[=to_curie]{to_curie()}}.
\item \code{"lgl_NA_FALSE"} to replace \code{NA} in logical columns with \code{FALSE}.
\item \code{"as_tibble"} to make the output a \link[tibble:tibble]{tibble}.
\item \code{"rm_lang_tag"} to remove language tags. Tags will only be removed from
\code{character} class columns, and then only if there is one unique language tag
in the given column.
\item \code{"nothing"} to prevent all tidying.
}}

\item{...}{
  Arguments passed on to \code{\link[=is_valid_obo]{is_valid_obo}}
  \describe{
    \item{\code{allow}}{The OBO ID format(s) to consider valid; as a character vector.
One or more of: \code{"curie"}, \code{"obo_curie"} (e.g. \code{obo:DOID_4}), \code{"uri"},
\code{"<uri>"} (URI in angle brackets), or \code{"ns.lui"} (i.e. an OBO CURIE without
\verb{obo:}).

\code{"standard"} (default) is a generic grouping that represents syntactically
correct formats (\code{"curie"}, \code{"obo_curie"}, \code{"uri"} and \code{"<uri>"}).

\code{"ns.lui"} will only be included if explicitly specified.}
    \item{\code{ns_type}}{The type of OBO namespaces to consider valid; as a character
vector. One of: \code{"obo"} (default), \code{"ont"} (ontology primary namespaces
only), or \code{"prop"} (ontology property namespaces only; may not be exhaustive).}
  }}
}
\value{
A \link[tibble:tibble]{tibble} with the columns: \code{id}, \code{label}, \code{parent_id},
and \code{parent_label}.
}
\description{
Extracts the classes and their \code{rdfs:subClassOf} \emph{asserted} parents from an
OBO ontology.
}
\examples{
\dontrun{
# Extract all descendants of 'diabetes mellitus'
doid_path <- "<path to doid.owl>"
extract_obo_class(doid_path, "DOID:9351")

# Extract all ancestors of 'diabetes mellitus'
extract_obo_class(doid_path, "DOID:9351", method = "ancestors")

# Extract all classes of appendicitis (DOID:8337), alcoholic gastritis
(DOID:8680), and viral esophagitis (DOID:6297) up to a common ancestor
extract_obo_class(
    doid_path,
    c("DOID:8337", "DOID:8680", "DOID:6297"),
    method = "common_ancestor"
)
}

}
\seealso{
\code{\link[=format_subtree]{format_subtree()}} to arrange data in a tree structure similar to
ontology browsers.
}
