% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_obo_data}
\alias{extract_obo_data}
\title{Extract OBO Foundry Ontology Data}
\usage{
extract_obo_data(
  obo_ont,
  prefix = NULL,
  id = NULL,
  include_anon = TRUE,
  ...,
  .robot_path = NULL
)
}
\arguments{
\item{obo_ont}{The path to an ontology file, as a string.}

\item{prefix}{A character vector of OBO prefixes (aka ID spaces) to filter
results to, or \code{NULL} (default) to return all axioms. \emph{Ignored if \code{id} is}
\emph{provided.}}

\item{id}{A character vector of OBO IDs (CURIEs) to filter results to or
\code{NULL} (default) to return all entities with logical relations.}

\item{include_anon}{Whether to include anonymous relationships
(logical/complex) in the output, as a boolean (default: \code{TRUE}). See
\code{\link[=extract_obo_anon]{extract_obo_anon()}}.}

\item{...}{
  Arguments passed on to \code{\link[=extract_obo_anon]{extract_obo_anon}}
  \describe{
    \item{\code{render}}{The format for rendering classes & properties, as a string.
One of:
\itemize{
\item \code{"label"} (default): Use labels, quoting as needed.
\item \code{"id"}: Use OBO IDs (CURIEs).
}}
  }}

\item{.robot_path}{The path to a ROBOT executable or .jar file, as a string.
When \code{NULL} (default), if a system ROBOT executable is available it will
be used, otherwise an error will be signaled.

\strong{NOTE:} \code{DO.utils} caches the last ROBOT used for future use.}
}
\value{
A tibble of class \code{obo_data} with the columns: \code{id}, \code{predicate},
\code{value}, \verb{axiom predicate}, and \verb{axiom value}. An additional class is added
to indicate the file the data came from (without extension or directories).
}
\description{
Extracts data from an OBO Foundry ontology.
}
\section{NOTES}{

Uses \href{https://robot.obolibrary.org/query}{ROBOT query} internally.
}

