% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_obo_def_xref}
\alias{extract_obo_def_xref}
\title{Extract OBO Definition Cross-References}
\usage{
extract_obo_def_xref(
  onto_path,
  output = NULL,
  tidy_what = c("header", "uri_to_curie")
)
}
\arguments{
\item{onto_path}{The path to an ontology file recognizable by
\href{https://robot.obolibrary.org/}{ROBOT}, as a string.}

\item{output}{The path where output will be written, as a string, or \code{NULL}
(default) to load data directly.}

\item{tidy_what}{The elements of a SPARQL-created data.frame to tidy, as a
character vector. One or more of the following:
\itemize{
\item \code{"everything"} to apply all tidy operations (has precedence over
\code{"nothing"}).
\item \code{"header"} to remove leading \verb{?} from header labels.
\item \code{"unnest"} to unnest list columns with \code{\link[=unnest_cross]{unnest_cross()}}.
\item \code{"uri_to_curie"} to convert all URIs recognized by DO.utils to CURIEs with
\code{\link[=to_curie]{to_curie()}}.
\item \code{"lgl_NA_FALSE"} to replace \code{NA} in logical columns with \code{FALSE}.
\item \code{"as_tibble"} to make the output a \link[tibble:tibble]{tibble}.
\item \code{"rm_lang_tag"} to remove language tags. Tags will only be removed from
\code{character} class columns, and then only if there is one unique language tag
in the given column.
\item \code{"nothing"} to prevent all tidying.
}}
}
\value{
A 2 column \code{tibble} (cols: \code{id}, \code{def_xref}).
}
\description{
Extracts cross-reference (xref) annotations from OBO ontology class
definitions. These are typically URLs of references supporting the definition
or curator IDs.
}
