% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_obo_mappings}
\alias{extract_obo_mappings}
\title{Extract Mappings from an OBO Foundry Ontology}
\usage{
extract_obo_mappings(
  onto_path,
  id = NULL,
  version_as = "release",
  output = NULL,
  tidy_what = c("header", "unnest")
)
}
\arguments{
\item{onto_path}{The path to an ontology file recognizable by
\href{https://robot.obolibrary.org/}{ROBOT}, as a string.}

\item{id}{A character vector of IRIs or CURIEs to filter results to or
\code{NULL} (default) to return all mappings.}

\item{output}{The path where output will be written, as a string, or \code{NULL}
(default) to load data directly.}

\item{tidy_what}{The elements of a SPARQL-created data.frame to tidy, as a
character vector. One or more of the following:
\itemize{
\item \code{"everything"} to apply all tidy operations (has precedence over
\code{"nothing"}).
\item \code{"header"} to remove leading \verb{?} from header labels.
\item \code{"unnest"} to unnest list columns with \code{\link[=unnest_cross]{unnest_cross()}}.
\item \code{"uri_to_curie"} to convert all URIs recognized by DO.utils to CURIEs with
\code{\link[=to_curie]{to_curie()}}.
\item \code{"lgl_NA_FALSE"} to replace \code{NA} in logical columns with \code{FALSE}.
\item \code{"as_tibble"} to make the output a \link[tibble:tibble]{tibble}.
\item \code{"rm_lang_tag"} to remove language tags. Tags will only be removed from
\code{character} class columns, and then only if there is one unique language tag
in the given column.
\item \code{"nothing"} to prevent all tidying.
}}

\item{versionIRI_as}{The format for \code{subject_source_version}, if a
\code{versionIRI} is found, as a string. One of:
\itemize{
\item \code{"release"} (default): Use the ontology's release version as a string.
Assumes that the release version in the \code{versionIRI} follows \verb{release/} or
\verb{releases/}. Will return full \code{versionIRI} as fallback.
\item \code{"iri"}: Use the ontology's full \code{versionIRI}.
}}
}
\value{
If \code{output} is specified, the path to the output file with the data.
Otherwise, the data as a \link[tibble:tibble]{tibble}.

Mappings data will be formatted according to the
\href{https://github.com/mapping-commons/sssom}{SSSOM} specification,
with an additional \code{status} column indicating the status (active, deprecated,
etc.) of each \code{subject_id}.

\code{subject_source_version} will be the ontology \code{versionIRI} (formatted
according to \code{version_as} or today's date if \code{versionIRI} is not defined.
}
\description{
Extract mappings from the OBO Foundry Ontology as SSSOM. .
}
