% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_ordo_mappings}
\alias{extract_ordo_mappings}
\title{Extract mappings from ORDO}
\usage{
extract_ordo_mappings(ordo_path, as_skos = TRUE, output = NULL, ...)
}
\arguments{
\item{ordo_path}{The path to the ORDO file, as a string.}

\item{as_skos}{Whether to convert ORDO's annotated \code{oboInOwl:hasDbXref}
mappings to their
\href{https://www.w3.org/TR/2009/REC-skos-reference-20090818/}{Simple Knowledge Organization System (SKOS)}
equivalents, as a boolean (default: \code{TRUE}).

The ORDO-skos equivalent predicates are as follows:
\itemize{
\item \code{"BTNT"} - \code{skos:narrowMatch}
\item \code{"NTBT"} - \code{skos:broadMatch}
\item \code{"E"} - \code{skos:exactMatch}
\item \code{"ND"} - \code{doid:undefinedMatch} (supplements SKOS)
\item \code{"W"} - \code{doid:notMatch} (supplements SKOS)
}}

\item{output}{The path where output will be written, as a string, or \code{NULL}
(default) to load data directly.}

\item{...}{
  Arguments passed on to \code{\link[=tidy_sparql]{tidy_sparql}}
  \describe{
    \item{\code{as_curies}}{Whether to convert IRIs to CURIEs, as a boolean
(default: \code{TRUE}).}
    \item{\code{lgl_NA_false}}{Whether to replace \code{NA} values with \code{FALSE} in logical
outputs, as a boolean (default: \code{TRUE}).}
  }}
}
\value{
If \code{output} is specified, the path to the output file with the data.
Otherwise, the data as a \href{tibble::tibble}{tibble}.

ORDO mappings data will be formatted according to the
\href{https://github.com/mapping-commons/sssom}{SSSOM} specification,
with an additional \code{status} column indicating the status (active, deprecated,
etc.) of each ORPHA term.

If \code{as_skos = FALSE}, ORDO's text-based \code{oboInOwl:hasDbXref} annotations
denoting the type of relationship the Xref represents (simple text code only)
will be included in the \code{predicate_modifier} column.
}
\description{
Extract mappings from the Orphanet Rare Disease Ontology (ORDO). ORDO uses
\code{oboInOwl:hasDbXref} for mapping with annotations to indicate
exact/broad/narrow-ness. Utilizes \code{\link[=robot]{robot()}}.
}
