% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_pmid.elink}
\alias{extract_pmid.elink}
\title{Extract PubMed ID from elink object}
\usage{
\method{extract_pmid}{elink}(x, linkname = NULL, quietly = FALSE, no_result = "error", ...)
}
\arguments{
\item{x}{An \code{elink} object, generated by \link[rentrez:entrez_link]{rentrez::entrez_link} and it's
derivatives (e.g. \link{citedby_pmid}).}

\item{linkname}{linkname as specified for \href{https://eutils.ncbi.nlm.nih.gov/entrez/query/static/entrezlinks.html}{Entrez API};
only necessary if more than one \code{linkname} is returned from PubMed.}

\item{quietly}{Suppress PubMed linkname message when \code{linkname} is not
specified and multiple results exist in an \code{elink} object.}

\item{no_result}{The type of condition that should be signaled when no PubMed
results exist in a response; one of "error", "warning" (default),
"message" or "none".}

\item{...}{Unused, included for generic consistency only.}
}
\description{
Extract PubMed ID from an \code{elink} object.
}
\seealso{
Other extract_pmid documentation:
\code{\link[=extract_pmid]{extract_pmid()}},
\code{\link[=extract_pmid.elink_list]{extract_pmid.elink_list()}}
}
\concept{extract_pmid documentation}
