% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_retry_after}
\alias{extract_retry_after}
\title{Extract Retry-After (INTERNAL)}
\usage{
extract_retry_after(x, format = "sec")
}
\arguments{
\item{x}{An \link[httr:response]{httr::response} object.}

\item{format}{The format to return the header in, as a string. Either
"sec", meaning in seconds, or "datetime", meaning the datetime after
which the request can be retried.}
}
\value{
For "sec", an integer. For "datetime", a POSIXct scalar.
}
\description{
Extract "Retry-After" header from an \link[httr:response]{httr::response} object.
}
\section{Provenance}{

modified from httr2:::resp_retry_after (MIT license, copyright httr2 authors,
https://github.com/r-lib/httr2/blob/main/LICENSE.md)
}

\section{Note}{

httr converts headers to lowercase (no idea why), but it applies the
"insensitive" class and has indexing methods so standard headers can be
extracted.
}

\keyword{internal}
