% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{extract_ScoredMatch}
\alias{extract_ScoredMatch}
\title{Extract ScoredMatch Objects (INTERNAL)}
\usage{
extract_ScoredMatch(py_ScoredMatch, prefix = NULL, prefix_sep = ":")
}
\arguments{
\item{py_ScoredMatch}{A GILDA ScoredMatch (python) object.}

\item{prefix}{\emph{Optional} prefix to add to namespace local unique identifiers
(LUI; e.g. 4, the LUI for "disease" in DO), as a string; preferably to
create a complete namespace ID (e.g. "DOID:4").}

\item{prefix_sep}{\emph{Optional} separator placed between \code{prefix} and
namespace LUIs, as a string. Ignored if \code{prefix = NULL}.}
}
\description{
Extracts mapping results from specialized \code{ScoredMatch} python objects
(as defined by GILDA) AT THE LEVEL of individual mappings.
}
\keyword{internal}
