% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_subclass_axiom}
\alias{extract_subclass_axiom}
\title{Extract 'Subclass Of' Axioms}
\usage{
extract_subclass_axiom(DO_repo)
}
\arguments{
\item{DO_repo}{The local path to the \code{HumanDiseaseOntology} repo, as a
string.}
}
\value{
'Subclass Of' axioms in OWL functional format, as a character vector.
}
\description{
Extract \code{owl:subClassOf} axioms from the doid-edit.owl file.
}
\seealso{
Other \verb{extract_*_axiom} functions:
\code{\link[=extract_class_axiom]{extract_class_axiom()}},
\code{\link[=extract_eq_axiom]{extract_eq_axiom()}}
}
\concept{\verb{extract_*_axiom} functions}
