% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_doid}
\alias{format_doid}
\title{Format DOIDs}
\usage{
format_doid(x, as = "CURIE", allow_bare = FALSE)
}
\arguments{
\item{x}{A set of DOIDs, as a character vector.}

\item{as}{The format to convert the DOIDs to, as a string. All valid formats
are possible options: "CURIE" (default), "URI", "obo_CURIE", "basename".}

\item{allow_bare}{Whether bare numbers should be allowed as input, \code{TRUE}
or \code{FALSE} (default).}
}
\description{
Convert valid DOIDs and/or bare numbers to a specified DOID format. No
attempt is made to confirm bare numbers or DOIDs match existing diseases in
the ontology.
}
\examples{
x <- c(
    "http://purl.obolibrary.org/obo/DOID_0001816",
    "DOID:4",
    "obo:DOID_14566",
    "DOID_0040001"
)

format_doid(x, as = "URI")
format_doid(x, as = "CURIE")
format_doid(x, as = "obo_CURIE")
format_doid(x, as = "basename")

w_bare <- c(x, "0001816")
format_doid(w_bare, allow_bare = TRUE)

}
