% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_hyperlink}
\alias{format_hyperlink}
\title{Format URLs as Hyperlinks}
\usage{
format_hyperlink(url, as, ..., text = NULL, preserve = "url")
}
\arguments{
\item{url}{One or more URLs, as a character vector.}

\item{as}{What format to use for the hyperlink, as a string; one of "gs"
(Google Sheet), "xlsx" (Excel), or "html".}

\item{...}{\emph{(Only for \code{as = "html"})} One or more name-value pairs of html
\verb{<a>} \href{https://www.w3schools.com/tags/tag_a.asp}{attributes}.}

\item{text}{\emph{(Optional)} The text to display for each link, as a character
vector. If \code{NULL} (default), the URL itself will serve as the text. If only
one value is provided, it will be recycled.}

\item{preserve}{The value to return when \code{url} is \code{NA}, as a string. One of
"url" (default) or "text".}
}
\description{
Formats URLs as hyperlinks for Google Sheets, Excel, or html.
}
\section{Excel Note}{

Use the \code{openxlsx} pkg to write data with hyperlinks to Excel.
}

\examples{
format_hyperlink("https://www.google.com", "gs")
format_hyperlink("https://www.google.com", "xlsx")
format_hyperlink("https://www.google.com", "html")

# with 'text' (argument must be named)
format_hyperlink("https://www.google.com", "gs", text = "Google")
format_hyperlink("https://www.google.com", "xlsx", text = "Google")
format_hyperlink("https://www.google.com", "html", text = "Google")

# html with <a> attributes
format_hyperlink(
    "https://www.google.com",
    "html",
    text = "Google",
    target = "_blank",
    rel = "external"
)

# NA values in 'url' are passed through without modification by default. If
# 'text' is provided and preferred when 'url' is NA use preserve = "text".
format_hyperlink(c("https://www.google.com", NA), "gs")
format_hyperlink(
    c("https://www.google.com", NA),
    "gs",
    text = c("Google", "placeholder")
)

# 'url' is always preserved when 'text' is NA but 'url' is not
format_hyperlink(
    c("https://www.google.com/", "https://madeup.url.com/fakeID"),
    "html",
    text = c("google", NA),
    preserve = "text"
)

}
