% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_helpers.R
\name{generify_obo}
\alias{generify_obo}
\title{Generify OBO Classes/Properties (INTERNAL)}
\usage{
generify_obo(x)
}
\arguments{
\item{x}{Complete logical axioms in OWL functional syntax, as a character
vector.}
}
\description{
\code{\link[=format_axiom]{format_axiom()}} internal helper that replaces URI/CURIEs defined by OBO
ontologies with namespace-prefixed descriptions/types, where defined
internally, or just the namespace. Will not modify non-OBO class URI/CURIEs.
}
\section{Internally-defined OBO namespace-types}{

The following namespaces have 'namespace:types' defined by this function:
\itemize{
\item CL: 'CL:cell'
\item CHEBI: 'CHEBI:chemical'
\item DISDRIV: 'DISDRIV:disease_driver'
\item DOID: 'DOID:disease'
\item FOODON: 'FOODON:food_material'
\item ECO: 'ECO:evidence'
\item GENO: 'GENO:inheritance_pattern'
\item NCBITaxon: 'NCBITaxon:organism'
\item OMIM: 'OMIM:susceptibility'
\item UBERON: 'UBERON:anatomy'
\item UPHENO: 'UPHENO:phenotype'
\item SO: 'SO:sequence'
\item SYMP: 'SYMP:symptom'
\item TRANS: 'TRANS:pathogen_transmission'
}

Ontologies with multiple object types used by the DO (e.g. HP, onset or
phenotype), that are imported via DISDRIV (e.g. ExO), or used for their
properties (e.g. RO) are not defined internally and will be formatted
namespace only.
.
}

\concept{format_axiom() helpers}
\keyword{internal}
