% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{get_html_indent}
\alias{get_html_indent}
\title{Get HTML Indentation Data}
\usage{
get_html_indent(html)
}
\arguments{
\item{html}{HTML code as a character vector or a string with newline
characters.}
}
\value{
A list of the form \code{list(type = " ", base = "  ", unit = "  ")},
where \code{type} is the preferred indent type (space or tab), \code{base} is the
base indent (the minimum indent across all lines), and \code{unit} is the indent
increment (the minimum increment above the base indent across all lines).
}
\description{
Get indentation data from HTML code, including the preferred indent type
(space or tab), the base indent (the minimum indent across all lines), and
the indent increment (the minimum increment above the base indent across all
lines).

This function is useful for formatting HTML code to improve readability and
maintain consistent indentation.
}
