% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\name{gs_col2rgb}
\alias{gs_col2rgb}
\title{Convert Google Sheets Colors to RGB}
\usage{
gs_col2rgb(colors)
}
\arguments{
\item{colors}{A character vector of hex codes or one of the following color
names from Google Sheets red, orange, yellow, green, cyan, blue, purple, magenta, light red 3, light orange 3, light yellow 3, light green 3, light cyan 3, light blue 3, light purple 3, light magenta 3, white, black, dark grey 1, light grey 2.}
}
\value{
A 3 x \code{length(colors)} matrix with \code{Red}, \code{Green}, & \code{Blue} rows in
1 column for each color in \code{colors}.
}
\description{
Converts one or more Google Sheets color names or hex codes to RGB values.
}
\seealso{
Other Google Sheets Formatting functions: 
\code{\link{set_gs_fill}()}
}
\concept{Google Sheets Formatting functions}
\keyword{internal}
