% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{guess_delim}
\alias{guess_delim}
\title{Guess Data Delimiter}
\usage{
guess_delim(x, strict = TRUE)
}
\arguments{
\item{x}{String or character data for which to guess delimiter.}

\item{strict}{Whether delimiter should be found in all lines (default:
\code{TRUE}), or just a majority of lines. If \code{FALSE}, an attempt is made to guess
the delimiter for data that may not be entirely delimited.}
}
\description{
Identifies the delimiter of data as being the greater of "," or "\\t" present
in all or most lines.
}
\keyword{internal}
