% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{hoist_ArticleIds}
\alias{hoist_ArticleIds}
\title{Hoists IDs from Pubmed/PMC Data (Internal)}
\usage{
hoist_ArticleIds(pubmed_df, id = NULL)
}
\arguments{
\item{pubmed_df}{A data frame, as produced by \code{\link[=as_tibble]{as_tibble()}} on a PubMed/PMC
\code{esummary_list}.}

\item{id}{One or more IDs to hoist, as a character vector. If \code{NULL}
(default), all IDs will be hoisted. Available IDs may include "doi",
"eid", "mid", "pii", "pmcid", "pmcid_long", "pmid", or "rid".}
}
\description{
Hoists IDs from Pubmed/PMC Data. \emph{Does not use \code{\link[tidyr:hoist]{tidyr::hoist()}} because}
\emph{\code{IdType} identifier is not a parent and ID positions in list can be}
\emph{variable.}
}
\section{Note}{

In direct results from PubMed/PMC the \code{ArticleIds} results are not
equivalent. The ID type names from PMC correspond with those in this package
but the PubMed ID type names do not. For PubMed \code{pmcid} is named \code{pmc} and
\code{pmid} is named \code{pubmed}.
}

