% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_obsolete}
\alias{identify_obsolete}
\title{Identify Obsolete (INTERNAL)}
\usage{
identify_obsolete(x, ...)
}
\arguments{
\item{x}{An object with information about obsolete terms (use \code{\link[=methods]{methods()}} for
available classes).}

\item{...}{Not used; included for extensibility.}
}
\description{
Identify obsolete terms of the Human Disease Ontology.
}
\keyword{internal}
