% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{init_domain_repo}
\alias{init_domain_repo}
\title{Initiate a Domain Repository}
\usage{
init_domain_repo(domain, check_robots, delay)
}
\arguments{
\item{domain}{The domain(s) for which to instantiate repositories, as a
character vector.}

\item{check_robots}{Whether to check the robots.txt for the specified
domain(s), as a boolean vector with one value (to be applied to all
domains) or one value per domain.}

\item{delay}{The default delay to use, as a positive integer vector with
one value (to be applied to all domains) or one value per domain. If
\code{check_robots = TRUE} this delay will be overridden by a domain's
robots.txt delay (if present).}
}
\value{
For a single domain, a \code{domain_repo}; otherwise a list of \code{domain_repo}
objects.
}
\description{
Create a \code{domain_repo} list to store, update, and use HTTP information for
specified domain(s). \emph{Designed to be used with URL audits.}
}
