% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot.R
\name{install_robot}
\alias{install_robot}
\title{Install OBO Foundry ROBOT Tool (Mac/Linux ONLY)}
\usage{
install_robot(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link[utils:download.file]{utils::download.file()}}; should not be
needed}
}
\value{
The version of robot.jar, if successful, otherwise a warning indicating where
failures occurred along with their non-zero integer codes. See documentation
of \code{\link[utils:download.file]{utils::download.file()}} (for download of robot/robot.jar) and \code{\link[=system2]{system2()}}
(for making robot executable; uses system \code{chmod}) under "Value" for more
details.
}
\description{
Installs latest ROBOT and robot.jar files to default system path
(/usr/local/bin). Only tested for Macs but also will likely work for Linux.
}
