% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_mapping}
\alias{inventory_mapping}
\title{Assess if Mappings Exist in DO}
\usage{
inventory_mapping(
  onto_path,
  input,
  include_pred = c("skos:exactMatch", "skos:closeMatch", "oboInOwl:hasDbXref"),
  when_pred_NA = "error"
)
}
\arguments{
\item{onto_path}{The path to an ontology file, as a string.}

\item{input}{A data.frame or the path to a .tsv or .csv file (possibly
compressed) that includes identifiers to compare against the mappings in the
ontology in one and ONLY one of the following columns: \code{id} or \code{mapping}.}

\item{include_pred}{The predicates to include when testing for one-to-multiple
mappings, as a character vector (default: \code{skos:exactMatch},
\code{skos:closeMatch}, and \code{oboInOwl:hasDbXref}). All other predicates are
ignored.}

\item{when_pred_NA}{What to do when missing predicates are detected, as a
string; one of "error" (default), "warn", or NULL (do nothing). \code{NA}
predicates are \emph{always} ignored when no mapping exists (i.e. one or both
corresponding values of \code{x} or \code{y} is/are also \code{NA}).}
}
\value{
The \code{input} with the identifer column (re)named \code{mapping} and 5 additional
columns:
\itemize{
\item \code{exists}: Logical indicating whether an ID exists as a mapping in the DO.
\item \code{mapping_type}: The mapping predicate(s) of this ID to a disease, if
present. Multiple predicate(s) between the same ID and DOID will be pipe
delimited.
\item \code{doid}: The DOID of the disease mapped to this ID, if present.
\item \code{do_label}: The label of the disease mapped to this ID, if present.
\item \code{do_dep}: Logical indicating whether a disease is deprecated or not, if
present.
\item \code{multimaps}: The direction in which an ID or DO term maps to multiple
terms in the other resource, as "input_to_doid", "doid_to_input", "both_ways"
or \code{NA}.
}

Output will have the class \code{mapping_inventory}.
}
\description{
Assesses whether identifiers are present in the Human Disease Ontology as
mappings (either xrefs or skos mappings). Utilizes \code{\link[=robot]{robot()}} for comparison.
}
\examples{
\dontrun{
# manually copy or download data ... UPDATE EXAMPLE!!!
inventory_mapping(
    onto_path = "~/Ontologies/HumanDiseaseOntology/src/ontology/doid-edit.owl",
    input = "omimps.csv",
)
}

}
