% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_omim_susc}
\alias{inventory_omim_susc}
\title{Assess whether OMIM susceptibilities are in the DO}
\usage{
inventory_omim_susc(
  susc_path,
  omim_input,
  do_path = NULL,
  keep_mim = c("#", "\%")
)
}
\arguments{
\item{susc_path}{The path to the \code{omim_susc_import.owl} file, as a string.}

\item{omim_input}{An \code{omim_tbl} created by \code{\link[=read_omim]{read_omim()}} or the path to a
.tsv or .csv file (possibly compressed) that can be read by \code{\link[=read_omim]{read_omim()}} and
includes OMIM data to compare against the susceptibility classes in the
ontology.

NOTE: If an \code{omim_tbl} is provided, \code{keep_mim} will be ignored.}

\item{do_path}{The path to a Human Disease Ontology file, as a string, or
\code{NULL} (default). If provided, additional information about the DOIDs
(labels, deprecated status) that are related to the susceptibilities will
be included in the output.}

\item{keep_mim}{[\strong{OMIM search data only}] The MIM symbols representing
the data types to keep, as a character vector, or \code{NULL} to retain all
(default: \code{"#"} and \code{"\%"}).

The \href{https://www.omim.org/help/faq#1_3}{OMIM} defined MIM symbols are:\tabular{ll}{
   MIM symbol \tab MIM type \cr
   \code{*} \tab gene \cr
   \code{+} \tab gene, includes phenotype \cr
   \verb{#} \tab descriptive entry, not unique locus; usually a phenotype \cr
   \verb{\%} \tab phenotype / phenotypic locus, unknown molecular basis \cr
   \code{^} \tab deprecated \cr
   \code{none} \tab phenotype (usually), suspected or possibly overlapping \cr
}}
}
\value{
The \code{omim_input} with 4 additional columns:
\itemize{
\item \code{exists}: Logical indicating whether an OMIM ID is present in the DO as a
susceptibility.
\item \code{susc_label}: The label of the susceptibility.
\item \code{susc_dep}: Logical indicating whether the susceptibility is deprecated or
not.
\item \code{related_doid}: All disease(s) related to a given OMIM susceptibility
(delimited by " | "). If \code{do_path} is provided, the data will be formatted as
"label (DOID; deprecated)" for each related disease; otherwise, only the
DOID(s) will be included.
}

Output will have the class \code{omim_susc_inventory}.
}
\description{
Assesses whether OMIM entries are present in the Human Disease Ontology as
susceptibilities (in the \code{omim_susc_import.owl} file). Utilizes \code{\link[=robot]{robot()}} for
comparison.
}
\examples{
\dontrun{
# execute within the HumanDiseaseOntology repository and download data from
# https://www.omim.org/phenotypicSeries/PS145600 to omimps.tsv
inventory_omim_susc(
  susc_path = "src/ontology/omim_susc_import.owl",
  omim_input = "omimps.tsv",
  do_path = "src/ontology/doid-edit.owl"
)
}

}
