% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_report}
\alias{inventory_report}
\alias{inventory_report.omim_inventory}
\title{Report Inventory Results}
\usage{
inventory_report(inventory_df, verbose = TRUE)

\method{inventory_report}{omim_inventory}(inventory_df, verbose = TRUE, include_xrefs = TRUE)
}
\arguments{
\item{inventory_df}{A \verb{*_inventory} data.frame created by one of DO.utils'
\verb{inventory_*()} functions.}

\item{verbose}{Whether to report results on screen (default: \code{TRUE}).}

\item{include_xrefs}{Whether \code{oboInOwl:hasDbXrefs} should be included in
tests for "one-to-many" mappings, as a boolean (default: \code{TRUE}).
\code{skos:exactMatch} & \code{skos:closeMatch} mappings are always included.}
}
\value{
A list of data.frames including statistical \code{counts} and results for
review, invisibly.

Specific \code{omim_inventory} method data.frames:
\itemize{
\item \code{doid_deprecated},
\item \code{omim_to_many}: Results where an OMIM ID corresponds to multiple
DOIDs, excluding skos broad/narrow/related matches.
\item \code{doid_to_many}): Results where a DOID corresponds to multiple
OMIM IDs, excluding skos broad/narrow/related matches.
}
}
\description{
Reports on results of an \verb{inventory_*()} run.
}
