% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert_nm}
\alias{invert_nm}
\title{Invert Names and Values (INTERNAL)}
\usage{
invert_nm(x)
}
\arguments{
\item{x}{A named object.}
}
\description{
Currently only supports character vectors.
}
\keyword{internal}
