% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_curie}
\alias{is_curie}
\title{CURIE (RDF ID) Predicate}
\usage{
is_curie(x, def = "obo_generic")
}
\arguments{
\item{x}{A set of CURIEs, as a character vector.}

\item{def}{The definition of CURIEs to test against; as a string. One of
\code{"obo"}, \code{"obo_generic"} (default), \code{"w3c"}, or \code{"w3c_safe"} (see \code{Notes}
section for details).}
}
\description{
This predicate is designed to validate CURIEs, compact URIs conforming to the
\href{https://www.w3.org/TR/2010/NOTE-curie-20101216/}{W3C CURIE Syntax 1.0}
standard (or stricter as allowed by that standard).
}
\section{Notes}{

The \href{https://www.w3.org/TR/2010/NOTE-curie-20101216/}{W3C CURIE Syntax 1.0}
standard defines the most general CURIE syntax allowed (\code{def = "w3c"}; not a
perfect implementation, e.g. U+3000 non-breaking space is not accepted). Note
that by definition URIs will be identified as CURIEs. If it is desirable to
distinguish these use \code{def = "w3c_safe"} and wrap CURIEs in square brackets
(as defined in the standard). Alternatively, use the stricter OBO
Foundry-based standards (as stated in that standard, more strict definitions
for CURIEs can be defined).

\code{"obo"} corresponds to the official OBO Foundry definition of a CURIE as
stated in the \href{http://obofoundry.org/id-policy.html}{OBO Foundry ID Policy}.
This should work for \emph{most} class CURIEs from OBO Foundry ontologies but will
not recognize \code{obo:IDSPACE_LOCALID} formatted CURIEs or CURIEs with letters
or symbols in their local unique identifier. To allow these in the set while
still restricting to CURIE patterns found in OBO Foundry ontologies, use
\code{"obo_generic"}.
}

\examples{
id <- c(
    #### pass all ####
    "DOID:0001816", "CL:0000066",
    #### pass "obo_generic" & "w3c" ####
    "obo:DOID_4", "obo:so#has_origin", # obo prefixed CURIEs
    "oboInOwl:hasDbXref", "skos:exactMatch", # obo object properties
    "alfred:LO362836C", # not OBO but conforms to `"obo_generic"` pattern
    #### pass only "w3c" ####
    "_4dn.biosource:4DNSR73BT2A2", "aceview.worm:aap-1",
    #### always fail ####
    "4dn.biosource:4DNSR73BT2A2", # starts with a number
    "0001816", # bare number without prefix
    " obo:HP_0000001", # must have NO `[:space:]` characters
    "http://purl.obolibrary.org/obo/DOID_0001816" # URI
)

}
\seealso{
Other ID predicates:
\code{\link[=is_uri]{is_uri()}},
\code{\link{obo_ID_predicates}}

Other predicates:
\code{\link[=all_duplicated]{all_duplicated()}},
\code{\link{char_val_predicates}},
\code{\link[=iff_all_vals]{iff_all_vals()}},
\code{\link[=is_invariant]{is_invariant()}},
\code{\link[=is_uri]{is_uri()}},
\code{\link{lgl_predicates}},
\code{\link{num_val_predicates}},
\code{\link{obo_ID_predicates}}
}
\concept{ID predicates}
\concept{predicates}
