% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_hex_color}
\alias{is_hex_color}
\title{Test for Hexadecimal Colors}
\usage{
is_hex_color(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Tests whether values in a character vector are valid hexadecimal color codes.
Will \emph{NOT} recognize \href{https://en.wikipedia.org/wiki/Web_colors#Shorthand_hexadecimal_form}{abbreviated hex codes}
(e.g. \verb{#fff}).
}
\seealso{
Other predicates: 
\code{\link{char_val_predicates}},
\code{\link{is_curie}()},
\code{\link{is_invariant}()},
\code{\link{lgl_predicates}},
\code{\link{num_val_predicates}},
\code{\link{obo_ID_predicates}}
}
\concept{predicates}
