% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_predicates.R
\name{is_valid_doid}
\alias{is_valid_doid}
\title{DO-specific predicate}
\usage{
is_valid_doid(x)
}
\arguments{
\item{x}{A set of DOIDs, as a character vector.}
}
\description{
This predicate is designed to identify and validate common formats for DOIDs
derived from the Human Disease Ontology (DO). Note that bare numbers are NOT
recognized as valid DOIDs, a prefix is \emph{always} required.
}
\examples{
# formats considered valid
is_valid_doid("http://purl.obolibrary.org/obo/DOID_0001816") # URI
is_valid_doid("DOID:4") # CURIE, standard version
is_valid_doid("obo:DOID_14566") # obo CURIE, less common
is_valid_doid("DOID_0040001") # basename (prefix removed)

# invalid formats
is_valid_doid(c("0001816", "4")) # bare numbers
is_valid_doid("obo:DOID:14566") # some specificity for separators is enforced

}
