% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_valid_obo_prop}
\alias{is_valid_obo_prop}
\title{OBO Property ID Predicate \emph{(DEPRECATED)}}
\usage{
is_valid_obo_prop(x)
}
\arguments{
\item{x}{A set of IDs, as a character vector.}
}
\description{
\strong{\emph{DEPRECATED}} -- Use \code{is_valid_obo(x, ns_type = 'prop')} instead.
}
