% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexiclean.R
\name{lexiclean}
\alias{lexiclean}
\title{Prepare Text for Matching}
\usage{
lexiclean(x, mod = "all")
}
\arguments{
\item{x}{A character vector.}

\item{mod}{Desired modifications, as a character string:
\itemize{
\item \code{"all"} : Use all modifications.
\item \code{"lc"} : Convert to lowercase.
\item \code{"roman"} : Convert all numbers to roman numerals.
\item \code{"rm_space"} : Remove all spaces, defined by \verb{[:space:]}, see \code{\link[=regex]{regex()}}.
\item \code{"rm_punct"} : Remove all punctuation, defined by \verb{[:punct:]}, see
\code{\link[=regex]{regex()}}.
}}
}
\value{
A character vector with specified modifications applied.
}
\description{
Process text to improve exact, quasi-exact and/or lexical text matching
(e.g. "lexical cleaning").
}
\examples{
x <- c("disease", "X-linked exudative vitreoretinopathy 2", "hepatitis D",
       "dilated cardiomyopathy 1FF", "chromosome 15q25 deletion syndrome",
       "spinocerebellar ataxia type 19/22", "Addison's disease",
       "disease type II", "B-lymphoblastic leukemia/lymphoma, BCR-ABL1–like")

lexiclean(x)
lexiclean(x, "roman")

}
