% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{lgl_predicates}
\alias{lgl_predicates}
\alias{is_boolean}
\title{Logical predicate}
\usage{
is_boolean(x)
}
\arguments{
\item{x}{vector to be tested}
}
\description{
This predicate is designed to identify boolean vectors (i.e. length 1 logical
vectors).
}
\seealso{
Other predicates:
\code{\link[=all_duplicated]{all_duplicated()}},
\code{\link{char_val_predicates}},
\code{\link[=iff_all_vals]{iff_all_vals()}},
\code{\link[=is_curie]{is_curie()}},
\code{\link[=is_invariant]{is_invariant()}},
\code{\link[=is_uri]{is_uri()}},
\code{\link{num_val_predicates}},
\code{\link{obo_ID_predicates}}
}
\concept{predicates}
\concept{type_predicates}
