% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{make_use_case_html}
\alias{make_use_case_html}
\title{Make HTML for DO Use Case Tables}
\usage{
make_use_case_html(out_dir = "graphics/website", group = "all")
}
\arguments{
\item{out_dir}{The path to the directory where output should be saved, as a
string.}

\item{group}{The group(s) to generate html for, as a character vector. One or
more of: "all" (default), "ontology", "resource", or "methodology".}
}
\value{
One "html" file in \code{out_dir} for each \code{group} named as
"DO_use_case-\{group\}.html" and the "User" data from the Google Sheet
invisibly.
}
\description{
Makes the row and cell html code for the various sections/tables of the
disease-ontology.org "Use Cases" page from the DO team's "DO_uses" google
sheet. This function explicitly avoids including the html code for defining
the table itself to provide for flexibility. The "html" output in the files
specified must be manually copied and pasted into the disease-ontology.org
"Use Cases" file in the appropriate section/table.
}
