% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{make_user_list_html}
\alias{make_user_list_html}
\title{Make HTML for DO User List}
\usage{
make_user_list_html(file)
}
\arguments{
\item{file}{The file path where the output should be saved, as a string.}
}
\description{
Makes the row and cell html code for the "Users of the Disease Ontology"
section of the collaborators page on disease-ontology.org from the DO team's
"Uses" google sheet. This function explicitly avoids including the html
code for defining the table itself to provide for flexibility.
}
