% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_arg_several}
\alias{match_arg_several}
\title{Match Length-2+ Vector Arguments}
\usage{
match_arg_several(arg, choices)
}
\arguments{
\item{arg}{Function argument, as a character vector, or \code{NULL}.}

\item{choices}{Candidate values, as a character vector.}
}
\description{
Matches arguments with several inputs allowed against a set of choices.
Similar to \code{\link[base:match.arg]{base::match.arg()}} with \code{several.ok = TRUE} \emph{EXCEPT}
\code{match_arg_several()} will signal an error for unmatching values in \code{arg} instead
of silently dropping them.
}
\keyword{internal}
