% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carefully.R
\name{match_carefully}
\alias{match_carefully}
\title{Matches Carefully}
\usage{
match_carefully(x, table, nomatch = NA_integer_, incomparables = NULL)
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}

\item{nomatch}{the value to be returned in the case when no match is
    found.  Note that it is coerced to \code{integer}.}

\item{incomparables}{a vector of values that cannot be matched.  Any
    value in \code{x} matching a value in this vector is assigned the
    \code{nomatch} value.  For historical reasons, \code{FALSE} is
    equivalent to \code{NULL}.}
}
\description{
Wrapper for \code{\link[base:match]{base::match()}} that will NOT match \code{NA} values. Uses
\code{\link[dplyr:if_else]{dplyr::if_else()}} to skip \code{NA} values.
}
\concept{carefully}
