% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr_to_scalar.R
\name{mold_string}
\alias{mold_string}
\title{Create a String from Inputs}
\usage{
mold_string(..., delim = "|", na.rm = FALSE, unique = FALSE)
}
\arguments{
\item{...}{One or more R objects (vector, list, data.frame, etc).}

\item{delim}{A delimiter to place between vector elements (default: "|").}

\item{na.rm}{A logical scalar indicating whether \code{NA} values should be
removed (default: \code{FALSE}).}

\item{unique}{Whether to include only unique values (across all inputs), as
a logical.}
}
\description{
Creates a single string from one or more inputs with each value separated by
\code{delim}.
}
\examples{
# collapses individual vectors
mold_string(1:10)

# input order is preserved
mold_string(1:2, letters[1:2])
mold_string(data.frame(x = 1:2, y = letters[1:2]))

# factor values are preserved ONLY if all inputs are factors
mold_string(factor(letters[1:2]), factor("c")) # values preserved
mold_string(factor(letters[1:2]), "c") # values lost

# unique applies across all inputs, order is determined by first appearance
mold_string(c(3, 1, 2), 1:4, unique = FALSE)
mold_string(c(3, 1, 2), 1:4, unique = TRUE)

}
