% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{multimaps}
\alias{multimaps}
\title{Identify One-to-Multiple Mappings}
\usage{
multimaps(
  x,
  pred,
  y,
  include_pred = c("skos:exactMatch", "skos:closeMatch", "oboInOwl:hasDbXref"),
  when_pred_NA = "error"
)
}
\arguments{
\item{x}{Vector with \code{subject} of mappings (i.e. those being tested; the
"one" in the "one-to-multiple" test).}

\item{pred}{Vector with predicate(s) of mappings. Predicate(s) should
be formatted as CURIEs but can include multiple delimited predicates.}

\item{y}{Vector with \code{object} of mappings (i.e. those being counted; the
"multiple" in the "one-to-multiple" test).}

\item{include_pred}{The predicates to include when testing for one-to-multiple
mappings, as a character vector (default: \code{skos:exactMatch},
\code{skos:closeMatch}, and \code{oboInOwl:hasDbXref}). All other predicates are
ignored.}

\item{when_pred_NA}{What to do when missing predicates are detected, as a
string; one of "error" (default), "warn", or NULL (do nothing). \code{NA}
predicates are \emph{always} ignored when no mapping exists (i.e. one or both
corresponding values of \code{x} or \code{y} is/are also \code{NA}).}
}
\value{
A logical vector specifying the positions in \code{x} that map to
multiple values in \code{y}. Incomplete mappings, where values of \code{x}, \code{y}, or
both are \code{NA}, are ignored and return \code{FALSE}.
}
\description{
Identifies values in \code{x} that map to multiple values in \code{y} for specified
mapping predicates.
}
\keyword{internal}
