% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obofoundry_metadata}
\alias{obofoundry_metadata}
\title{OBO Foundry Metadata}
\format{
A data frame with 22 variables:
\describe{
\item{id}{}
\item{title}{}
\item{domain}{}
\item{description}{}
\item{activity_status}{}
\item{is_obsolete}{}
\item{replaced_by}{}
\item{ontology_purl}{}
\item{preferred_prefix}{}
\item{homepage}{}
\item{contact_name}{}
\item{contact_email}{}
\item{contact_github}{}
\item{license_label}{}
\item{taxon_id}{}
\item{taxon_label}{}
\item{twitter}{}
\item{facebook}{}
\item{publications}{}
\item{in_foundry_order}{}
\item{in_foundry}{}
\item{build_infallible}{}
}
}
\source{
\url{http://www.obofoundry.org/registry/ontologies.jsonld}, last accessed 2021-11-05.
}
\usage{
obofoundry_metadata
}
\description{
Metadata about OBO Foundry ontologies.
}
\keyword{datasets}
