% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_html_table}
\alias{parse_html_table}
\title{Parse HTML (mat-)table to Dataframe}
\usage{
parse_html_table(html, type = "table", n = 1L)
}
\arguments{
\item{html}{A file path to, or a string of, HTML content containing a
(mat-)table.}

\item{type}{The type of table to parse, as a string. One of:
\itemize{
\item \code{"table"} (default): A standard HTML table.
\item \code{"mat-table"}: An Angular Material table.
\item \code{"unknown"}: attempt to parse any existing table of either type.
}}

\item{n}{The position of the table to parse, as a positive integer (default:
1).}
}
\value{
A \code{tibble} with the table content.
}
\description{
Parses an HTML table or mat-table to a data.frame.
}
