% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_url}
\alias{parse_url}
\title{Parse URLs}
\usage{
parse_url(url)
}
\description{
Parses URLs into protocol, domain, and path components. From
\code{robotstxt:::parse_url()} (MIT license, \href{https://github.com/ropensci/robotstxt/blob/master/LICENSE}{copyright Peter Meissner}), with
slight modification.
}
\keyword{internal}
