% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_na_rm}
\alias{paste_na_rm}
\title{Paste Omitting NA Values}
\usage{
paste_na_rm(..., sep = " ", collapse = NULL)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.  When \code{collapse} is a string,
    the result is always a string (\code{\link[base]{character}} of length 1).}
}
\description{
Pastes elements together while omitting \code{NA} values, returning \code{NA} only if
all values are \code{NA}.
}
\examples{
paste_na_rm(
  c(letters[1:2], NA, NA),
  c(1, NA, NA, 4),
  sep = ", "
)

paste_na_rm(
  c(letters[1:2], NA, NA),
  c(1, NA, NA, 4),
  sep = ", ",
  collapse = "; "
)

}
