% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{pivot_subtree}
\alias{pivot_subtree}
\title{Convert Subtree DF to tidygraph}
\usage{
pivot_subtree(subtree_tg, top_node)
}
\arguments{
\item{subtree_tg}{A subtree tidygraph from \code{\link[=as_subtree_tidygraph]{as_subtree_tidygraph()}}.}

\item{top_node}{The top node of the tree, as a valid DOID (see
\code{\link[=is_valid_doid]{is_valid_doid()}} for valid input formats).}
}
\description{
Converts a subtree tidygraph, from \code{\link[=as_subtree_tidygraph]{as_subtree_tidygraph()}}, to a
dataframe with the tree hierarchically organized. This is the second step
in formatting a subtree as a text-based ontology tree similar to the
presentation on \href{https://disease-ontology.org/}{disease-ontology.org}.
}
\keyword{internal}
