% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_citedby}
\alias{plot_citedby}
\title{Plot Publications Citing DO by Year}
\usage{
plot_citedby(
  data_file = "data/citedby/DO_citedby.csv",
  out_dir = "graphics/website",
  color_set = "accent1",
  w = 8,
  h = 5.6
)
}
\arguments{
\item{data_file}{The path to the file containing the list of publications
citing the DO, as a string.}

\item{out_dir}{The directory where the plot \code{"DO_cited_by_count.png"}
should be saved, as a string. If \code{NULL} the plot is not saved to disk.}

\item{color_set}{The prefix of the color set to use from \link{DO_colors}.
Available sets include: "sat", "accent1" (default), "accent2", and
"orange".  The default and light versions of the specified color set will
be used.}

\item{w}{The width of the plot in inches, as numeric.}

\item{h}{The height of the plot in inches, as numeric.}
}
\description{
Plots the count of publications that cite the Human Disease Ontology by
year.
}
\section{Data Preparation}{

To prepare data, execute \code{scripts/citedby_full_procedure.R}.
}

