% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{plot_def_src}
\alias{plot_def_src}
\title{Plot Definition Sources}
\usage{
plot_def_src(DO_repo, out_dir = "graphics/website", w = 8, h = 5.6)
}
\arguments{
\item{DO_repo}{The local path to the \code{HumanDiseaseOntology} repo, as a
string.}

\item{out_dir}{The directory where the plot "DO_def_src.png" should be saved,
as a string. If \code{NULL} the plot is not saved to disk.}

\item{w}{The width of the plot in inches, as numeric.}

\item{h}{The height of the plot in inches, as numeric.}
}
\description{
Plots the count of definition sources for \emph{non-obsolete} terms the Human
Disease Ontology.
}
\section{Data Preparation}{

If this plot will be added to disease-ontology.org, the latest release of the
\code{HumanDiseaseOntology} Github repo should be checked out prior to running
this function.
}

