% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{plot_term_def_counts}
\alias{plot_term_def_counts}
\title{Plot DO Term & Definition Counts}
\usage{
plot_term_def_counts(
  release_file = "data/DO_release/DO_release_details.csv",
  counts_file = "data/DO_release/DO_term_def_counts.csv",
  out_dir = "graphics/website",
  w = 8,
  h = 5.6
)
}
\arguments{
\item{release_file}{The path to the file containing DO release details, as a
string.}

\item{counts_file}{The path to the file containing the count of DO terms
and definitions by release, as a string.}

\item{out_dir}{The directory where the plot \code{"DO_term_def_count.png"}
should be saved, as a string.}

\item{w}{The width of the plot in inches, as an integer.}

\item{h}{The height of the plot in inches, as an integer.}
}
\description{
Plots the count of \emph{non-obsolete} terms and definitions in the Human Disease
Ontology over time (using data from each release).
}
\section{Data Preparation}{

To prepare data, execute:
\enumerate{
\item \code{scripts/DO_term_def_counts.R} - requires installation of a python virtual
environment using \code{scripts/install_reticulate_python.R}.
\item \code{scripts/DO_release_details.R}
}
}

