% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_xref_counts}
\alias{plot_xref_counts}
\title{Plot Xref Counts}
\usage{
plot_xref_counts(DO_repo, out_dir = "graphics/website", w = 8, h = 5.6)
}
\arguments{
\item{DO_repo}{The local path to the HumanDiseaseOntology repo, as a string,
or a \link{DOrepo} object.}

\item{out_dir}{The directory where the plot \code{"DO_xref_count.png"}
should be saved, as a string. If \code{NULL} the plot is not saved to disk.}

\item{w}{The width of the plot in inches, as numeric.}

\item{h}{The height of the plot in inches, as numeric.}
}
\description{
Plots the count of cross-references by source in the Human Disease Ontology.
}
