% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc_summary.R
\name{pmc_summary}
\alias{pmc_summary}
\title{Get PubMed Central Summary}
\usage{
pmc_summary(input, config = NULL, version = "2.0", retmode = "xml", ...)
}
\arguments{
\item{input}{One of the following:
\enumerate{
\item A vector with unique PubMed Central IDs.
\item A list of vectors with unique PubMed Central IDs.
\item A \code{web_history} object (see NCBI Entrez API documentation for
information).
}}

\item{config}{vector configuration options passed to \code{httr::GET}}

\item{version}{Argument included here for flexibility,
but best left with the defaults as set; see \code{\link[rentrez:entrez_summary]{rentrez::entrez_summary()}}
for details.}

\item{retmode}{"xml" (default) or "json"; "xml" is preferred because of a
higher response limit. This default is opposite the
\code{\link[rentrez:entrez_summary]{rentrez::entrez_summary()}} default.}

\item{...}{character Additional terms to add to the request, see NCBI
documentation linked to in references for a complete list}
}
\description{
Retrieves PubMed Central Summary information for specified publications.
Acts in exactly the same manner as \code{\link[=pubmed_summary]{pubmed_summary()}}. \emph{Consult}
\emph{documentation there for more details.}
}
