% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{preprocess_omim_dl}
\alias{preprocess_omim_dl}
\title{Preprocesses omim.org Data (INTERNAL)}
\usage{
preprocess_omim_dl(file, ...)
}
\arguments{
\item{file}{The path to a .tsv or .csv file (possibly compressed) with OMIMPS
data.}

\item{...}{
  Arguments passed on to \code{\link[=read_delim_auto]{read_delim_auto}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A tibble with data corresponding to the OMIM PS table, with the
PS phenotype included \emph{ONLY} when source was the "Download as" button;
copy and pasting the PS table with the PS header is not allowed.
}
\description{
Determines if OMIM data is from "Download as" button on a Phenotypic Series
site or copy-pasted, and pre-processes file to fix headers and retain only
data.
}
\keyword{internal}
