% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priority_sort.R
\name{priority_sort}
\alias{priority_sort}
\title{Sort Vectors by Priority}
\usage{
priority_sort(x, levels, exclude = NA)
}
\arguments{
\item{x}{a vector of data, usually taking a small number of distinct
    values.}

\item{levels}{an optional vector of the unique values (as character strings)
    that \code{x} might have taken.  The default is the unique set of
    values taken by \code{\link[base]{as.character}(x)}, sorted into
    increasing order \emph{of \code{x}}.  Note that this set can be
    specified as smaller than \code{sort(unique(x))}.}

\item{exclude}{a vector of values to be excluded when forming the
    set of levels.  This may be factor with the same level set as \code{x}
    or should be a \code{character}.}
}
\description{
Sort vectors, prioritizing according to levels specified. Uses
\code{\link[base:factor]{base::factor()}} internally with unspecified arguments disabled.
}
