% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{class}
\name{progress_bar-class}
\alias{progress_bar-class}
\alias{progress_bar}
\title{Progress Bar (RefClass)}
\description{
Instantiate a base R-style progress bar and advances it by one 'tick' at a
time.
}
\section{Fields}{

\describe{
\item{\code{bar}}{A \code{progress_bar} object created by \code{\link[utils:txtProgressBar]{utils::txtProgressBar()}}.}

\item{\code{tick}}{An integer to track progress.}
}}

\section{Methods}{

\describe{
\item{\code{advance(...)}}{Advances progress bar one tick.}

\item{\code{initialize(n, ...)}}{Initializes a progress bar when called with \code{progress_bar$new(n, ...)}.
\itemize{
\item{\code{n}}{: Total number of ticks.}
\item{\code{...}}{: Arguments passed on to \code{\link[utils:txtProgressBar]{utils::txtProgressBar()}}.}
}}
}}

