% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pub_id_types}
\alias{pub_id_types}
\title{Prioritized List of Publication IDs for Matching}
\format{
A length-4 character vector:
\describe{ pmid > pmcid > doi > scopus_eid }
}
\usage{
pub_id_types
}
\description{
Prioritized List of Publication IDs for Matching
}
\keyword{datasets}
