% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rentrez_helpers.R
\name{pubmed_links}
\alias{pubmed_links}
\title{List Available PubMed-to-PubMed Links}
\usage{
pubmed_links()
}
\description{
List link types available for PubMed in the Entrez API that could be
specified in \code{\link[rentrez:entrez_link]{rentrez::entrez_link()}}.
}
