% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pubs}
\alias{pubs}
\alias{DO_pubs}
\alias{ST_pubs}
\title{Publication Info}
\format{
Each data frame consists of 10 variables:
\describe{
\item{internal_id}{short-version identifier used by the DO team}
\item{pmid}{PubMed ID}
\item{pmcid}{PubMed Central ID}
\item{doi}{DOI}
\item{scopus_eid}{Scopus Electronic Identification, not the same as a
Scopus ID}
\item{lens_id}{Lens.org ID}
\item{semantic_scholar_id}{Semantic Scholar Corpus ID}
\item{first_author}{Publication First Author}
\item{title}{Publication Title}
\item{citation_nlm}{Full NLM-formatted citation}
}

\code{DO_pubs} includes 10 publications.

\code{ST_pubs} includes 1 publications.
}
\source{
Compiled by J. Allen Baron; last updated 2022-07-22.
}
\usage{
DO_pubs

ST_pubs
}
\description{
Datasets describing official publications of the Human Disease Ontology
project, grouped by ontology. Data includes
publication identifiers from various sources, along with publication titles
and NLM-formatted citations.
\itemize{
\item DO_pubs: Human Disease Ontology publications only
\item ST_pubs: Symptom Ontology & Pathogen Transmission Ontology publications
}
}
\keyword{datasets}
