% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyDOID.R
\name{pyDOID}
\alias{pyDOID}
\alias{DOrepo}
\title{Instantiate a DO Repository}
\usage{
DOrepo(path)
}
\arguments{
\item{path}{The local path to the
\href{https://github.com/DiseaseOntology/HumanDiseaseOntology}{HumanDiseaseOntology}
repository, as a string.}
}
\description{
Instantiates a \code{DOrepo} python object from the
\href{https://github.com/allenbaron/pyDOID}{pyDOID} package (powered by
\href{https://github.com/rstudio/reticulate}{reticulate}).
}
\details{
The \code{DOrepo} class inherits the methods & subclasses (accessible via \code{$}
notation) of the
\href{https://gitpython.readthedocs.io/en/stable/tutorial.html#meet-the-repo-type}{GitPython.Repo}
class, and adds the following:
\itemize{
\item \verb{DOrepo$tag_iterate(function, start, end)}: Iterates through DO releases
(tags) from \code{start} to \code{end} executing a \code{function} at each. If \code{start} or
\code{end} are not specified the first or last release/tag of the repo will be
assumed.
\item \verb{DOrepo$\{doid|doid_merged\}$query(query, reload)}: Executes a SPARQL
\code{query}, input as a string or file path, against the specified DO file. Will
optionally \code{reload} the file into memory, useful in conjunction with
\code{tag_iterate()}.
\item \code{DOrepo$capture_head()}: Captures the current \code{head} of the repo, useful
for restoring state after git checkout(s). Use: If assigned to variable
\code{x}, restore state with \code{x$checkout()}.
}
}
