% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{pyobo_map}
\alias{pyobo_map}
\title{Map Terms to Terms of Specified Namespace}
\usage{
pyobo_map(terms, namespace)
}
\arguments{
\item{terms}{Term(s) to map, as a character vector.}

\item{namespace}{Namespace to map terms to (\emph{according to bioregistry}), as a
string.}
}
\value{
List of results for each term. See \code{\link[=parse_mapping]{parse_mapping()}} and dependencies for
details on format of results.
}
\description{
Map \code{term} to terms of a specified namespace (e.g. "DOID") using
\href{https://github.com/pyobo/pyobo}{PyOBO} (and by dependency other
\href{https://biopragmatics.github.io/}{biopragmatics} and
\href{https://indralab.github.io/}{INDRA labs} programs; e.g.
\href{https://github.com/indralab/gilda}{GILDA},
\href{http://bioregistry.io/}{bioregistry}). In INDRA labs
terminology, GILDA grounds (predicts matches of) \code{term} to \code{namespace} terms.
}
