% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_xref_split}
\alias{queue_xref_split}
\title{Curation Queue: DOIDs to split?}
\usage{
queue_xref_split(.DOrepo, src = "all")
}
\arguments{
\item{.DOrepo}{The path to the local HumanDiseaseOntology repository or a
\code{pyDOID.repo.DOrepo} object created by \code{\link[=DOrepo]{DOrepo()}}.}

\item{src}{The xref sources to include in the output identified by prefix
as a character vector or "all" (default) to include everything.}
}
\description{
Create a curation queue of DOIDs that may need to be split into multiple
diseases based on the fact that they have multiple xrefs from the same
source.
}
