% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\name{range_add_validation}
\alias{range_add_validation}
\alias{range_add_checkbox}
\alias{range_add_dropdown}
\title{Add Data Validation to Google Sheet Range}
\usage{
range_add_checkbox(
  ss,
  sheet = NULL,
  range,
  msg = "Value must be TRUE or FALSE",
  quiet = TRUE
)

range_add_dropdown(
  ss,
  sheet = NULL,
  range,
  values,
  msg = "Choose a valid value",
  reject_input = TRUE,
  display_arrow = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:drive_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to write into, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. Ignored if the sheet is specified via \code{range}. If neither argument specifies the sheet, defaults to the first visible sheet.}

\item{range}{Cells to apply data validation to. This \code{range} argument has
important similarities and differences to \code{range} elsewhere (e.g.
\code{\link[googlesheets4:range_read]{googlesheets4::range_read()}}):
\itemize{
\item Similarities: Can be a cell range, using A1 notation ("A1:D3") or using
the helpers in \link[googlesheets4:cell-specification]{googlesheets4::cell-specification}. Can combine sheet
name and cell range ("Sheet1!A5:A") or refer to a sheet by name
(\code{range = "Sheet1"}, although \code{sheet = "Sheet1"} is preferred for clarity).
\item Difference: Can NOT be a named range.
}}

\item{msg}{The message to display when the user types in a value that
violates the data validation rule. For \code{range_add_dropdown()}, only displayed
if \code{reject_input} is \code{TRUE}.}

\item{values}{The values to use for the dropdown list, as a character vector.}

\item{reject_input}{Whether to "Reject the input" (default: \code{TRUE}) if a
value violates the data validation rule or "Show a warning" (\code{FALSE}).}

\item{display_arrow}{Whether to display a dropdown arrow next to the cell
(default: \code{TRUE}) or not (\code{FALSE}).}
}
\description{
Add data validation to a Google Sheet range.
}
\section{Limitations of the Google Sheets API/\code{googlesheets4}}{

\itemize{
\item The API does not support chipset multi-selection in dropdowns:
https://stackoverflow.com/questions/79653536/how-to-enable-multiple-selection-in-data-validation-dropdown-using-google-sheets
}
}

\keyword{internal}
