% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_AGR_records.R
\name{read_AGR}
\alias{read_AGR}
\title{Read AGR .tsv.gz File}
\usage{
read_AGR(AGR_tsv)
}
\arguments{
\item{AGR_tsv}{path to AGR .tsv or .tsv.gz file}
}
\value{
A dataframe.
}
\description{
Reads in a .tsv or .tsv.gz file from the Alliance of Genome Resources (AGR)
as a tibble. It is recommended that AGR files be downloaded using
\code{\link[=download_AGR]{download_AGR()}}.
}
