% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_citedby}
\alias{read_citedby}
\title{Read 'Cited by' Data (INTERNAL)}
\usage{
read_citedby(resource = "DO")
}
\arguments{
\item{resource}{The resource to get 'cited by' information for, as a string.
One of: \code{"DO"}, \code{"SYMP"}, or \code{"TRANS"}.}
}
\value{
A \code{citedby_tbl} classed tibble.
}
\description{
Reads in the 'cited by' data for the specified resource from the official
\href{https://docs.google.com/spreadsheets/d/1wG-d0wt-9YbwhQTaelxqRzbm4qnu11WDM2rv3THy5mY/}{DO_uses}
Google Sheet.
}
\keyword{internal}
