% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_doid_edit}
\alias{read_doid_edit}
\title{Read doid-edit.owl (INTERNAL)}
\usage{
read_doid_edit(DO_repo)
}
\arguments{
\item{DO_repo}{The local path to the \code{HumanDiseaseOntology} repo, as a
string.}
}
\description{
Read the doid-edit.owl file from a local copy of the Human Disease Ontology
Github repo.
}
