% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat_doid}
\alias{reformat_doid}
\title{Reformat DOIDs}
\usage{
reformat_doid(x, to = "CURIE")
}
\arguments{
\item{x}{A set of DOIDs, as a character vector.}

\item{to}{The format to convert the DOIDs to, as a string. All valid formats
are possible options: "CURIE" (default), "URI", "obo_CURIE", "basename".}
}
\description{
Convert valid DOIDs to a different format.
}
\examples{
x <- c(
    "http://purl.obolibrary.org/obo/DOID_0001816",
    "DOID:4",
    "obo:DOID_14566",
    "DOID_0040001"
)

reformat_doid(x, to = "URI")
reformat_doid(x, to = "CURIE")
reformat_doid(x, to = "obo_CURIE")
reformat_doid(x, to = "basename")

}
