% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-methods.R
\name{replace_blank}
\alias{replace_blank}
\title{Replace Blanks with Specified Value}
\usage{
replace_blank(data, replace = NA_character_, ...)
}
\arguments{
\item{data}{A data object.}

\item{replace}{A string to use for replacement.}

\item{...}{Additional arguments passed on to methods. Not currently used.}
}
\value{
\itemize{
\item If \code{data} is a vector, \code{replace_blank()} returns a vector of the same class
as \code{data} (only blanks in character vectors are modified).
\item If \code{data} is a list, \code{replace_blank()} will recurse into the list (as
necessary) and replace all blank values in character/list elements but will
skip other internal components (e.g. data.frames, matrices, etc).
}
}
\description{
Replace blanks with specified value.
}
