% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website_helpers.R
\name{replace_html_counts}
\alias{replace_html_counts}
\title{Replace Counts in HTML File}
\usage{
replace_html_counts(DO_repo, svn_repo, page, reload = NULL)
}
\arguments{
\item{DO_repo}{The local path to the HumanDiseaseOntology repository, as a
string.}

\item{svn_repo}{The local path to the DO website svn trunk, as a string.}

\item{page}{The disease-ontology.org page in which to replace counts, as a
string; either "imports" or "slims".}

\item{reload}{\emph{DEPRECATED}. This argument is now ignored and will be
removed in a future release.}
}
\value{
Updated counts directly in the html of the svn repo for the specified page,
as well as the old and new counts for comparison (invisibly, as a tibble).
}
\description{
Replaces counts in the specified page of disease-ontology.org.
\code{replace_html_counts()} is the primary workhorse supporting the more
generalized \code{\link[=update_website_count_tables]{update_website_count_tables()}}.
}
\concept{update_website_count_tables() helpers}
\keyword{internal}
