% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary.R
\name{restore_names}
\alias{restore_names}
\title{Restore Element Names}
\usage{
restore_names(x, names_from)
}
\arguments{
\item{x}{object needing names retored}

\item{names_from}{object to restore names from}
}
\description{
Restores names to elements within objects after name-removing events.
}
\details{
Necessary for the following scenarios:
\itemize{
\item After string operations with \code{stringr} or \code{stringi}, until \code{stringi} fixes
\href{https://github.com/gagolews/stringi/issues/59}{issue #59}
}
}
