% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot.R
\name{robot}
\alias{robot}
\title{Execute Robot Commands}
\usage{
robot(params, ...)
}
\arguments{
\item{params}{string specifying subcommand and all arguments}

\item{...}{arguments passed on to \link[base:system2]{base::system2}}
}
\description{
Light wrapper for system ROBOT program (OBO Foundry). See
\href{http://robot.obolibrary.org/}{ROBOT documentation} for information about
subcommands and arguments.
}
\section{NOTE}{

Requires installation of the OBO Foundry ROBOT tool in the system path. This
can be achieved with \code{\link[=install_robot]{install_robot()}}.
}

\section{Citation}{

R.C. Jackson, J.P. Balhoff, E. Douglass, N.L. Harris, C.J. Mungall, and
J.A. Overton. ROBOT: A tool for automating ontology workflows.
BMC Bioinformatics, vol. 20, July 2019.
}

