% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot.R
\name{robot_query}
\alias{robot_query}
\title{Execute Robot Query}
\usage{
robot_query(input, rq, save)
}
\arguments{
\item{input}{path to owl file to be queried}

\item{rq}{file with SPARQL query to execute}

\item{save}{path to save result to}
}
\description{
Wrapper for system ROBOT program (OBO Foundry) that executes the query
subcommand. See \href{http://robot.obolibrary.org/}{ROBOT documentation} for
details.
}
\section{NOTE}{

Requires installation of the OBO Foundry ROBOT tool in the system path. This
can be achieved with \code{\link[=install_robot]{install_robot()}}.
}

\section{Citation}{

R.C. Jackson, J.P. Balhoff, E. Douglass, N.L. Harris, C.J. Mungall, and
J.A. Overton. ROBOT: A tool for automating ontology workflows.
BMC Bioinformatics, vol. 20, July 2019.
}

