% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roll_middle}
\alias{roll_middle}
\title{Calculate a Rolling/Windowed Middle (Mean) Value}
\usage{
roll_middle(x, limit, limit_type = "min")
}
\arguments{
\item{x}{A numeric vector.}

\item{limit}{A value that will be added to the start or end of \code{x}, as a
numeric scalar.}

\item{limit_type}{Specify whether \code{limit} should be used as the \code{"min"}
(default) or \code{"max"} value.}
}
\value{
A numeric vector of middle values of the same length as \code{x}.
}
\description{
Calculates a rolling/windowed mean between two values using an additional
\code{limit} value for the first or last calculation.
}
\section{Notes}{

An alternative function that produces the same result when the \code{limit} is
added at the start or end of the input \code{x} (e.g. \code{c(limit, x)} or
\code{c(x, limit)}) is \code{roll_mean(x, n = 2)} from the \code{RccpRoll} package.
}

\examples{
\dontrun{
x <- c(7, 14, 21, 25)

roll_middle(x, 0, "min")
roll_middle(x, 30, "max")
}

}
\keyword{internal}
