% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{round_down}
\alias{round_down}
\title{Round a number down}
\usage{
round_down(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{integer indicating the number of decimal places. Negative
values are allowed (see 'Details').}
}
\description{
Round a number down
}
\details{
Rounding to a negative number of digits means rounding to a power of ten, so
for example round_down(x, digits = -2) rounds to the nearest hundred.
}
