% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{round_zero}
\alias{round_zero}
\title{Round Number to Zero}
\usage{
round_zero(x, digits)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{The number of decimal places, as an integer. Negative
values are allowed (see 'Details').}
}
\description{
Round Number to Zero
}
\details{
Rounding to a negative number of digits means rounding to a power of ten, so
for example round_zero(x, digits = -2) rounds to the nearest hundred.
}
