% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alliance.R
\name{save_alliance_counts}
\alias{save_alliance_counts}
\title{Writes Alliance Counts to File}
\usage{
save_alliance_counts(counts_tbl, file, ...)
}
\arguments{
\item{counts_tbl}{record counts as \code{alliance_tbl} (e.g. output from
\code{\link[=count_alliance_records]{count_alliance_records()}})}

\item{file}{file to write to}

\item{...}{arguments to pass on to \code{\link[readr:write_csv]{readr::write_csv()}}}
}
\value{
Returns the \code{counts_tbl} with its version info footer invisibly.
}
\description{
Writes Alliance record counts to a .csv file with the version info of the
Alliance file it came from as a footer.
}
\seealso{
Other Alliance functions:
\code{\link[=count_alliance_records]{count_alliance_records()}},
\code{\link[=download_alliance_tsv]{download_alliance_tsv()}},
\code{\link[=read_alliance]{read_alliance()}}
}
\concept{Alliance functions}
