% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_robot_template.R
\name{save_data}
\alias{save_data}
\title{Save Data}
\usage{
save_data(data, save_to, ...)
}
\arguments{
\item{data}{A single data object to save.}

\item{save_to}{The path/location to save to, as a string or something
recognizable by \code{\link[googlesheets4:sheets_id]{googlesheets4::as_sheets_id()}}.}

\item{...}{Arguments passed on to methods.}
}
\description{
Saves data generated by \code{DO.utils} as Google Sheets, .csv, or .tsv file(s).
This is a generic function. Generally, data.frames will be saved directly to
one of these file types, while lists of data.frames will be saved as either
multiple sheets (Google Sheets) or multiple files (.csv/.tsv). For the latter
case, if data.frames are named the name will be appended before the
extension and, if not, a number will be added.
}
