% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_literature.R
\name{search_pmc}
\alias{search_pmc}
\title{Search PubMed Central}
\usage{
search_pmc(
  term,
  config = NULL,
  retmode = "xml",
  use_history = FALSE,
  pmid = FALSE,
  retmax = NULL,
  ...
)
}
\arguments{
\item{term}{the search term (character); see reference or
\link[rentrez:entrez_search]{rentrez::entrez_search} for syntax details}

\item{config}{vector configuration options passed to httr::GET}

\item{retmode}{character, one of json (default) or xml. This will make no
difference in most cases.}

\item{use_history}{logical. If TRUE return a web_history object for use in 
later calls to the NCBI}

\item{pmid}{whether to return PMID as well as PMCID (default = FALSE);
when TRUE, PMIDs are obtained via \link[rcrossref:id_converter]{rcrossref::id_converter})}

\item{retmax}{maximum number of PubMed IDs to return (integer; default = 20,
max = 100,000); see reference under 'Optional Parameters – Retrieval'}

\item{...}{character, additional terms to add to the request, see NCBI
documentation linked to in references for a complete list}
}
\description{
Searches PubMed Central using Entrez Utilities API via
\code{\link[rentrez:rentrez]{rentrez}}.
}
