% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_literature.R
\name{search_pubmed}
\alias{search_pubmed}
\title{Search PubMed}
\usage{
search_pubmed(
  term,
  config = NULL,
  retmode = "xml",
  use_history = FALSE,
  retmax = NULL,
  ...
)
}
\arguments{
\item{term}{the search term (character); see reference or
\link[rentrez:entrez_search]{rentrez::entrez_search} for syntax details}

\item{config}{vector configuration options passed to httr::GET}

\item{retmode}{character, one of json (default) or xml. This will make no
difference in most cases.}

\item{use_history}{logical. If TRUE return a web_history object for use in 
later calls to the NCBI}

\item{retmax}{maximum number of PubMed IDs to return (integer; default = 20,
max = 100,000); see reference under 'Optional Parameters – Retrieval'}

\item{...}{character, additional terms to add to the request, see NCBI
documentation linked to in references for a complete list}
}
\description{
Searches PubMed using Entrez Utilities API via \code{\link[rentrez:rentrez]{rentrez}}.
}
\references{
https://www.ncbi.nlm.nih.gov/books/NBK25499/#\emph{chapter4_ESearch}
}
\seealso{
rentrez::entrez_search
}
