% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\name{set_gs_fill}
\alias{set_gs_fill}
\title{Set Google Sheets Range Fill Color}
\usage{
set_gs_fill(ss, sheet, ranges, colors)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:drive_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{(OPTIONAL) The sheet name, as a string. If \code{NULL} (default), the
sheet name will default to "curation-" with today's date appended (formatted
as "\%Y\%m\%d"; see \code{\link[=format.Date]{format.Date()}}).}

\item{ranges}{A character vector of ranges to set the background color for,
as recognized by \code{\link[googlesheets4:range_flood]{googlesheets4::range_flood()}}.}
}
\description{
Sets the fill color for one or more ranges in a Google Sheet.
}
\section{NOTE}{

This function relies on internal functions from the \code{googlesheets4} package
and may break if the package is updated. See examples of
\code{\link[googlesheets4:range_flood]{googlesheets4::range_flood()}} for reference.
}

\seealso{
Other Google Sheets Formatting functions: 
\code{\link{gs_col2rgb}()}
}
\concept{Google Sheets Formatting functions}
\keyword{internal}
