% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_keys.R
\name{set_scopus_keys}
\alias{set_scopus_keys}
\title{Set Keys for Scopus API Access}
\usage{
set_scopus_keys(api_key = NULL, insttoken = NULL)
}
\arguments{
\item{api_key}{Elsevier API key, as a string.}

\item{insttoken}{Elsevier institutional token, as a string.}
}
\value{
Named logical vector stating whether setting operation was
successful for each key provided as input (invisibly).
}
\description{
Sets Elsevier API key and/or institutional token as environment variables
(\code{Elsevier_API} and \code{Elsevier_insttoken}) for access by subsequent Scopus
functions.
}
\seealso{
citedby_scopus
}
