% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation.R
\name{spreadsheet_range}
\alias{spreadsheet_range}
\title{Calculate a Spreadsheet Range}
\usage{
spreadsheet_range(.data, .col, sheet = NULL, rows = NULL, n_header = 1)
}
\arguments{
\item{.data}{A tibble.}

\item{.col}{The column to use for the range, as a string.}

\item{sheet}{(OPTIONAL) The sheet name, as a string. If \code{NULL} (default), the
sheet name will default to "curation-" with today's date appended (formatted
as "\%Y\%m\%d"; see \code{\link[=format.Date]{format.Date()}}).}

\item{rows}{(OPTIONAL) The rows to use for the range, either as a continous
integer vector or as a string (i.e. "1:10"). If \code{NULL} (default), the entire
column will be used.}

\item{n_header}{The number of header rows to skip (default: \code{1}).}
}
\description{
Calculate a range for a spreadsheet program (Google Sheets or Excel).
}
\keyword{internal}
