% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_www_duplicate}
\alias{standardize_www_duplicate}
\title{Standardize "www" in URLs}
\usage{
standardize_www_duplicate(url, how = "remove")
}
\arguments{
\item{url}{One or more URLs, as a character vector.}

\item{how}{How "www" should be handled. One of "remove" (default) or "add".}
}
\description{
Standardizes "www" in URLs by removing or adding it when URLs of a domain
exhibit mixed usage (some with "www" AND others without).
}
\keyword{internal}
