% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_DO}
\alias{theme_DO}
\title{DO Theme for Stats Plots}
\usage{
theme_DO(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
The default ggplot2 theme used for statistical plots uploaded to
disease-ontology.org. This \emph{only} manages the base plot style and does \emph{not}
incorporate DO's color scheme.
}
\section{Background}{

For more information refer to ggplot2's
\link[ggplot2:theme_grey]{theme documentation}.
}

