% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-methods.R
\name{tidy.esummary_list}
\alias{tidy.esummary_list}
\title{Tidy an \code{esummary_list} object}
\usage{
\method{tidy}{esummary_list}(x, ...)
}
\arguments{
\item{x}{An \code{esummary_list} object created by \link[rentrez:entrez_summary]{rentrez::entrez_summary}
(for single inputs \code{always_return_list} must be TRUE) and it's derivatives
(\link{pubmed_summary}).}

\item{...}{Additional arguments. Not used.}
}
\value{
A tibble, where \code{uid} is the esummary input identifier
}
\description{
Tidy converts an \code{esummary_list} into a \link[tibble:tibble]{tibble}.
}
