% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_ga_tbl}
\alias{tidy_ga_tbl}
\title{Tidy Google Analytics Tables (INTERNAL)}
\usage{
tidy_ga_tbl(ga_tbl, keep_total = FALSE)
}
\arguments{
\item{ga_tbl}{A Google analytics table loaded with \code{\link[=read_ga]{read_ga()}}, that needs
to be tidied.}

\item{keep_total}{Whether to keep the row with totals that appears at the
bottom of the table, as a logical scalar (default: \code{FALSE}).}
}
\description{
Tidies Google Analytics tables loaded with \code{\link[=read_ga]{read_ga()}}.
}
\keyword{internal}
