% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_pub_records}
\alias{tidy_pub_records}
\title{Tidy Publication Records}
\usage{
tidy_pub_records(x, ...)
}
\arguments{
\item{x}{A publication record object, as produced by \code{citedby_pubmed()},
\code{citedby_scopus()} and \code{pubmed_summary()}.}

\item{...}{Ignored; included for extensibility.}
}
\description{
Tidy publication record results into a tibble. Limited data is retained. If
full data retention is desired, consider using \code{as_tibble()} instead (expect
list columns and untidy column names).
}
