% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_pubmed.R
\name{tidy_pubmed_summary}
\alias{tidy_pubmed_summary}
\title{Convert PubMed summary to tibble}
\usage{
tidy_pubmed_summary(pm_summary, addl_items = NULL)
}
\arguments{
\item{pm_summary}{result from \code{rentrez::entrez_summary(db = "pubmed", ...)}}

\item{addl_items}{character vector listing "items" (\emph{i.e.} data variables)
to include in tidy dataset beyond the default: uid, pubdate, title,
authors, fulljournalname}
}
\description{
Convert list result generated by \code{\link[rentrez:entrez_summary]{rentrez::entrez_summary()}} from the
PubMed database into a tibble.
}
\section{NOTE}{

This function does not currently work on \code{esummary_list} objects.
It needs to be implemented as a generic.
}

