% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_helpers.R
\name{to_character}
\alias{to_character}
\title{Convert to Character}
\usage{
to_character(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Provides all character conversion methods of \code{\link[base:as.character]{base::as.character()}} and
additional methods for lists and data.frames. \code{to_character()} was created to
enable \code{\link[=collapse_to_string]{collapse_to_string()}} to handle many data types while avoiding unintended
conversions to character, which would have occurred if these methods were
added to \code{as.character()}.
}
\keyword{internal}
