% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{to_range}
\alias{to_range}
\title{Convert Vectors to Range String}
\usage{
to_range(
  x,
  int_fn = NULL,
  ...,
  sep = c(",", "-"),
  start_rm = NULL,
  end_rm = NULL
)
}
\arguments{
\item{x}{A numeric vector of \href{is_whole_number()}{whole numbers}.}

\item{...}{Arguments passed on to \code{int_fn}.}

\item{sep}{The separators to use between ranges (default: ',') and within
a range (default: '-'), as a length-2 character vector.}

\item{start_rm}{A regular expression to remove from \code{x} values at the
\emph{beginning} of a range.
#' @param end_rm A regular expression to remove from \code{x} values at the
\emph{end} of a range.}

\item{int_fun}{A function (or tidyverse-style formula) to convert \code{x} into an
integer vector, used \emph{ONLY for creating ranges}; the original value will
appear in the range except where modified by \code{start_rm} and/or \code{end_rm}.
\code{int_fun} is required when \code{x} is not a numeric vector.}
}
\description{
Converts vectors to a string of ranges. All vector types are accepted and the
original values will appear in the final output range, but where input is not
a numeric vector of \href{is_whole_number()}{whole numbers}, a function to convert
the values to integers must be provided (\code{int_fn}) for the purpose of
identifying the range(s).
}
\section{Notes}{

\itemize{
\item \code{NA} values are always dropped.
\item \code{to_range()} was inspired by answers at
https://stackoverflow.com/q/16911773/6938922, most heavily by speendo
(CC-BY-SA 3.0, accessed 2022-07-01). This is the fastest approach with
few inputs but is significantly slower than other answers for large inputs.
The internal approach will likely be modified in the future, but arguments
and output will remain the same.
}
}

\examples{
x <- c(1:2, 8:6, 4, -1:-2, 20:37, 4, 40, 43, 45)
to_range(x)


# `NA` values are dropped
y <- c(1:4, NA, 5, 7:10)
to_range(x)


# Use `int_fn` when `x` is not a numeric vector (tidyverse-style formulas
#    accepted)
x_char <- as.character(x)
to_range(x, int_fn = as.integer)
to_range(x, int_fn = ~ as.integer(.x))


# `int_fn` allows non-numeric ranges to be created
txt <- paste0(x, "txt")
to_int <- function(x, y) as.integer(stringr::str_remove(x, "txt"))
to_range(txt, to_int, y = "txt")


# text can be selectively removed from the values at the beginning of ranges
#     (`start_rm`) or end of ranges (`end_rm`)
to_range(txt, to_int, start_rm = "txt")
to_range(txt, to_int, end_rm = "txt")

}
