% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{to_uri}
\alias{to_uri}
\title{Convert CURIE to URI}
\usage{
to_uri(x)
}
\arguments{
\item{x}{CURIE(s), as a character vector.}
}
\description{
Converts CURIE(s) to URI(s). Angle brackets will not be added.
}
\section{Note}{

Performs no CURIE validation but relies on a properly formatted CURIE for
conversion (e.g. prefix:reference, according to the
\href{https://www.w3.org/TR/2010/NOTE-curie-20101216/#s_syntax}{CURIE standard}).
Any inputs not matching a prefix of \link{ns_prefix} will be returned without
modification.
}

\examples{
.curie <- c("rdfs:comment", "dc:date", "terms:license", "owl:deprecated",
            "oboInOwl:id", "UBERON:0000002", "DOID:0001816",
            "doid:DO_AGR_slim")
to_uri(.curie)

#returns non-CURIE or unknown namespace prefixes unmodified
to_curie(c("SYMP:0000000", "not a CURIE", "bioregistry.collection:0000001"))

}
\seealso{
Other identifier converters:
\code{\link[=to_curie]{to_curie()}}
}
\concept{identifier converters}
