% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{today_datestamp}
\alias{today_datestamp}
\title{Today's Date Stamp}
\usage{
today_datestamp()
}
\description{
Returns today's date in a format appropriate for use as a date stamp.
Format = YMD (Year Month Day) without separators (e.g. "20210728" for
July 28, 2021).
}
