% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{truncate_authors}
\alias{truncate_authors}
\title{Truncate PubMed Author List (internal)}
\usage{
truncate_authors(pubmed_df)
}
\arguments{
\item{pubmed_df}{A tidy data frame, as produced by \code{\link[=tidy]{tidy()}} on a PubMed
\code{esummary_list}}
}
\description{
Truncate Authors for any PubMed article to 120 and list the number of
additional authors in the subsequent Author position.
}
\section{NOTE}{

The number of authors retained was determined based on JSON string conversion
length and Google Sheets cell character limit. The maximum string conversion
of authors to json is ~ 400 / author (mean ~ 290). Google Sheets character
limit / cell = 50,000. 120 authors is a conservative limit that's unlikely
to cause problems (and is plenty long).
}

