% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try.R
\name{try_robots_txt}
\alias{try_robots_txt}
\title{Try robots.txt (INTERNAL)}
\usage{
try_robots_txt(domain)
}
\arguments{
\item{domain}{The URL of a domain to try, as a string.}
}
\description{
Try to request the robots.txt from a specified domain, capturing R
errors/warnings if they occur and noting HTTP status.
}
\section{NOTE}{

This function performs no checks to make sure the URL is for the overall
domain/subdomain where robots.txt would be found and will likely return an
R/HTTP error with any non-domain URL.
}

\examples{
rt <- try_robots_txt("disease-ontology.org")
rt

}
\keyword{internal}
