% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{update_import_counts_html}
\alias{update_import_counts_html}
\title{Update HTML for Website Pages with Counts}
\usage{
update_import_counts_html(DO_repo, tag, svn_repo)
}
\arguments{
\item{DO_repo}{A \code{pyDOID.repo.DOrepo} object (see \code{\link[=DOrepo]{DOrepo()}}).}

\item{tag}{The repo tag to extract data from, as a string.}

\item{svn_repo}{The local path to the DO website svn directory, as a string.
The correct directory will include a Dockerfile and the
'disease_ontology' directory.}
}
\value{
Direct count updates in the imports table on DO svn repo's Imports page and
the "User" data from the Google Sheet invisibly.
}
\description{
Directly updates counts of imports listed in the table on the
disease-ontology.org "DO Imports" page using data from a specified release.
Changes to the Imports html file should be reviewed and, if correct,
committed to the svn repo.
}
