% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{update_website_contributors}
\alias{update_website_contributors}
\title{Update 'Registry of Contributor' Tables on Web Pages}
\usage{
update_website_contributors(contrib_path, table_id)
}
\arguments{
\item{contrib_path}{The file path to the "Contributors" page HTML file, as a
string.}

\item{table_id}{The id of the table to update in the HTML file, as a string.}
}
\value{
The page's updated HTML, invisibly.
}
\description{
Updates data in the 'Registry of Contributor' tables of disease-ontology.org
to match the curated contributor data in the
\href{https://docs.google.com/spreadsheets/d/1kD7rgOWO2uVUwKYoKFSLBEpv1WZFf-GDhEusAq_H5sM/}{DO_contributors}
google sheet.
}
