% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{update_website_use_cases}
\alias{update_website_use_cases}
\title{Update Table on Use Cases Page}
\usage{
update_website_use_cases(use_cases_path, table_id = "use-cases")
}
\arguments{
\item{use_cases_path}{The file path to the "Use Cases" page HTML file, as a
string.}

\item{table_id}{The id of the table to update in the HTML file, as a string.}
}
\value{
The page's update HTML, invisibly.
}
\description{
Updates the data in the table on the "Use Cases" page of disease-ontology.org
to match the curated use cases in the
\href{https://docs.google.com/spreadsheets/d/1wG-d0wt-9YbwhQTaelxqRzbm4qnu11WDM2rv3THy5mY/?gid=1972219724#gid=1972219724}{DO_uses}
google sheet.
}
