% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr_to_scalar.R
\name{vctr_to_string}
\alias{vctr_to_string}
\title{Collapse vector to string}
\usage{
vctr_to_string(x, delim = "; ")
}
\arguments{
\item{x}{vector}

\item{delim}{delimiter to place between vector elements}
}
\description{
Concatenates values in a vector into a single string.
}
\seealso{
Other vector-to-scalar functions: 
\code{\link{unique_if_invariant}()}
}
\concept{vector-to-scalar functions}
