% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtuoso.R
\name{virtuoso_query}
\alias{virtuoso_query}
\title{Query a Virtuoso Graph}
\usage{
virtuoso_query(con, query, uri)
}
\arguments{
\item{con}{a ODBC connection to Virtuoso, from \code{\link[virtuoso:vos_connect]{vos_connect()}}}

\item{query}{A SPARQL query OR the path to a SPARQL query file (.rq/.sparql),
as a string.}

\item{uri}{The graph URI to be queried enclosed in angle brackets, as a
string (example: "<graph_name>"). If not specified via this argument or
in the query directly, all graphs in thebdatabase will be queried
(default). \code{uri} is used \emph{ONLY} when \code{query} is a file and will \emph{ONLY}
limit the first query to the specified graph. For greater control, edit
the SPARQL query directly and use \code{\link[virtuoso:vos_query]{virtuoso::vos_query()}}.}
}
\description{
Provides a thin wrapper around \code{\link[virtuoso:vos_query]{virtuoso::vos_query()}} so that a user can:
\enumerate{
\item Use a SPARQL file (.rq/.sparql).
\item Specify a graph in the database without altering the query in the file.
}
}
