% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap_onscreen}
\alias{wrap_onscreen}
\title{Wrap an On-screen Message}
\usage{
wrap_onscreen(msg, exdent = 2, ...)
}
\arguments{
\item{msg}{The message to wrap, as a string.}

\item{exdent}{a non-negative integer specifying the indentation of
    subsequent lines in paragraphs.}

\item{...}{
  Arguments passed on to \code{\link[base:strwrap]{base::strwrap}}
  \describe{
    \item{\code{x}}{a character vector, or an object which can be converted to a
    character vector by \code{\link[base]{as.character}}.}
    \item{\code{width}}{a positive integer giving the target column for wrapping
    lines in the output.}
    \item{\code{indent}}{a non-negative integer giving the indentation of the
    first line in a paragraph.}
    \item{\code{prefix,initial}}{a character string to be used as prefix for
    each line except the first, for which \code{initial} is used.}
    \item{\code{simplify}}{a logical.  If \code{TRUE}, the result is a single
    character vector of line text; otherwise, it is a list of the same
    length as \code{x} the elements of which are character vectors of
    line text obtained from the corresponding element of \code{x}.
    (Hence, the result in the former case is obtained by unlisting that
    of the latter.)}
  }}
}
\description{
Wraps an on-screen message, such as those used in messages, warnings, and
errors.
}
\examples{
short_msg <- "This is a short message."
long_msg <- "This is a long message to demonstrate how wrap_onscreen() works. It's basically intended to be a multi-line paragraph but, as you can see if you view this variable, it is really just a very long string. We hope you enjoy!"

# no effect on short messages
message(wrap_onscreen(short_msg))

# wrapping of longer messages, has default exdent
message(wrap_onscreen(long_msg))

# wrapping without exdent
message(wrap_onscreen(long_msg, exdent = 0))

}
