% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_access}
\alias{write_access}
\title{Tests for Write Access}
\usage{
write_access(names)
}
\arguments{
\item{names}{character vector containing file names.
    Tilde-expansion will be done: see \code{\link[base]{path.expand}}.}
}
\value{
Logical vector of length equal to \code{names}. \strong{NOTE} that this
\emph{differs} from \code{\link[base:file.access]{base::file.access()}} which returns an integer vector.
}
\description{
Acts as a minimal wrapper around \link[base:file.access]{file.access(mode = 2)}
to test whether R has permission to write to files and/or directories.
}
\keyword{internal}
