% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_gs}
\alias{write_gs}
\alias{write_gs.omim_inventory}
\alias{write_gs.data.frame}
\alias{write_gs.curation_template}
\title{Write Data to a Google Sheet}
\usage{
write_gs(data, ss, sheet = NULL, hyperlink_curie = NULL, ...)

\method{write_gs}{omim_inventory}(
  data,
  ss,
  sheet = "omim_inventory-\%Y\%m\%d",
  hyperlink_curie = c("omim", "doid"),
  ...
)

\method{write_gs}{data.frame}(data, ss, sheet = "data-\%Y\%m\%d", hyperlink_curie = NULL, ...)

\method{write_gs}{curation_template}(data, ss = NULL, sheet = "curation-\%Y\%m\%d", ...)
}
\arguments{
\item{data}{A data.frame, possibly with a defined method.}

\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:drive_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{The name to use for the sheet to write into (i.e. the tab name
in a worksheet). If a date format recognized by \code{\link[=format.Date]{format.Date()}} is
included in the string, today's date will be added in the specified format
and location.

For each method, the default is a method-specific term appended with a date
format lacking separators (e.g. 20250101 for 2025-01-01).

\strong{\emph{WARNING:}} If a sheet with the same name exists in the Google Sheet file
it will be overwritten.}

\item{hyperlink_curie}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> The
columns with CURIEs to convert to hyperlinks when written in Google Sheets.}

\item{...}{Arguments passed on to methods.}
}
\value{
The input \code{ss}, as an instance of \link[googlesheets4:sheets_id]{googlesheets4::sheets_id}.
}
\description{
Specialized methods for writing data created by \code{DO.utils} to a specified
Google Sheet.
}
\section{Deprecated arguments}{

\code{sheet_nm} and \code{datestamp} have been deprecated in favor of \code{sheet} which
supports more flexible naming, including names without dates.
}

