% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_duplicated}
\alias{all_duplicated}
\title{Identify all duplicates}
\usage{
all_duplicated(x, ...)
}
\arguments{
\item{x}{a vector or a data frame or an array or \code{NULL}.}

\item{...}{arguments for particular methods.}
}
\description{
Built on \code{\link[base:duplicated]{base::duplicated()}} but, unlike \code{base::duplicated()},
identifies all duplicates \emph{including} the first occurrence.
}
