% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{all_duplicated}
\alias{all_duplicated}
\title{Identify all duplicates}
\usage{
all_duplicated(x, ...)
}
\arguments{
\item{x}{a (generalized, see \code{\link[base]{is.vector}}) vector, a data
    frame,  an array or \code{NULL}.}

\item{...}{arguments for particular methods.}
}
\description{
Built on \code{\link[base:duplicated]{base::duplicated()}} but, unlike \code{base::duplicated()},
identifies all duplicates \emph{including} the first occurrence.
}
\seealso{
Other predicates: 
\code{\link{char_val_predicates}},
\code{\link{iff_all_vals}()},
\code{\link{is_curie}()},
\code{\link{is_hex_color}()},
\code{\link{is_invariant}()},
\code{\link{is_uri}()},
\code{\link{lgl_predicates}},
\code{\link{num_val_predicates}},
\code{\link{obo_ID_predicates}}
}
\concept{predicates}
