% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{append_to_url}
\alias{append_to_url}
\title{Append to URL}
\usage{
append_to_url(x, url)
}
\arguments{
\item{x}{value to append}

\item{url}{a URL or the internal name of a URL used in this package (see
\link{get_url} for possible names)}
}
\description{
Append a value to a URL.
}
\section{Note}{

No URL validation is performed.
}

