% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_to_url}
\alias{append_to_url}
\title{Append to URL}
\usage{
append_to_url(x, url, sep = "", preserve_NA = TRUE)
}
\arguments{
\item{x}{Value(s) to append, as a character vector.}

\item{url}{A URL or a URL name recognized by this package, as a string;
see \link{get_url} for recognized names.}

\item{sep}{The separator to use between \code{url} and \code{x}, as a string
(default = ""). If \code{url} ends in \code{sep}, an additional \code{sep} will not be
added.}

\item{preserve_NA}{Whether to preserve \code{NA} in output, as a boolean. \code{FALSE}
will result in \code{NA} being appended to the end of \code{url} (almost certainly not
desired).}
}
\description{
Append a value to a URL.
}
\section{Note}{

No URL validation is performed.
}

