% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble-methods.R
\name{as_tibble.esummary_list}
\alias{as_tibble.esummary_list}
\title{Convert an \code{esummary_list} to a Tibble}
\usage{
\method{as_tibble}{esummary_list}(x, ...)
}
\arguments{
\item{x}{An \code{esummary_list} object created by \link[rentrez:entrez_summary]{rentrez::entrez_summary}
(for single inputs \code{always_return_list} must be TRUE) and it's derivatives
(\link{pubmed_summary}).}

\item{...}{Additional arguments. Not used.}
}
\value{
A tibble, where \code{uid} is the esummary input identifier
}
\description{
Converts an \code{esummary_list} into a \link[tibble:tibble]{tibble}.
}
