% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperlink.R
\name{build_hyperlink}
\alias{build_hyperlink}
\title{Build Hyperlinks}
\usage{
build_hyperlink(x, url, as, ..., sep = "", text = x, preserve = "text")
}
\arguments{
\item{x}{Value(s) to append, as a character vector.}

\item{url}{One or more URLs or URL names recognized by this package, as a
character vector. If only one value is provided, it will be recycled;
otherwise the length of \code{url} and \code{x} must match. See \code{\link[=get_url]{get_url()}} for
recognized base URL names.}

\item{as}{What format to use for the hyperlink, as a string; one of "gs"
(Google Sheet), "xlsx" (Excel), or "html".}

\item{...}{\emph{(Only for \code{as = "html"})} One or more name-value pairs of html
\verb{<a>} \href{https://www.w3schools.com/tags/tag_a.asp}{attributes}.}

\item{sep}{One or more separators to use between \code{url} and \code{x}, as a
character vector. If only one value is provided (e.g. default = ""), it
will be recycled; otherwise the length of \code{sep} and \code{x} must match.
If any \code{url} ends in the corresponding \code{sep}, an additional \code{sep} will
not be added.}

\item{text}{\emph{(OPTIONAL)} The text to display for each link, as a character
vector. The default uses \code{x} as the text. If \code{NULL}, the full URL will serve
as the text.  If a string, the value will be used for the text of each
hyperlink.}

\item{preserve}{The value to return when \code{url} is \code{NA}, as a string. One of
"url" or "text" (default). Note that the default for \code{build_hyperlink()}
is opposite the default of \code{\link[=format_hyperlink]{format_hyperlink()}} because \code{text} is provided by
default.}
}
\description{
Builds hyperlinks for Google Sheets, Excel, or HTML by appending a value to
the end of a base url.
}
\examples{
build_hyperlink(
    x = "DiseaseOntology",
    url = "github",
    as = "html",
    text = "A hyperlink!"
)

# create CURIE links by passing local identifiers as `x` and prefixes as `url`
build_hyperlink(
    x = c("4", "D004194"),
    url = c("DOID", "MESH"),
    as = "gs",
    text = c("DOID:4", "MESH:D004194")
)

# provide internal URL names or direct URLs to append to
# BE SURE to use `preserve = 'url'` when text is `NA`.
build_hyperlink(
    x = c("4", "fakeID"),
    url = c("DOID", "https://madeup.url.com"),
    as = "gs",
    text = c("DOID:4", NA),
    sep = c("_", "/"),
    preserve = "url"
)

}
\seealso{
Functions used internally: \code{\link[=append_to_url]{append_to_url()}} and
\code{\link[=format_hyperlink]{format_hyperlink()}}.
}
\concept{Hyperlink functions}
