% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{char_val_predicates}
\alias{char_val_predicates}
\alias{is_blank}
\alias{is_missing}
\title{Character value predicates}
\usage{
is_blank(x)

is_missing(x)
}
\arguments{
\item{x}{vector to be tested}
}
\description{
These value predicates are designed to identify common values that appear
in character vectors.
}
\details{
\itemize{
\item \code{is_blank()} identifies "" or whitespace of any length
\item \code{is_missing()} identifies NA's and blanks
}
}
\seealso{
Other value predicates: 
\code{\link{is_invariant}()},
\code{\link{num_val_predicates}}

Other predicates: 
\code{\link{all_duplicated}()},
\code{\link{iff_all_vals}()},
\code{\link{is_curie}()},
\code{\link{is_invariant}()},
\code{\link{is_uri}()},
\code{\link{lgl_predicates}},
\code{\link{num_val_predicates}},
\code{\link{obo_ID_predicates}}
}
\concept{predicates}
\concept{value predicates}
