% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citedby.R
\name{citedby_pubmed}
\alias{citedby_pubmed}
\title{Gets Cited By List from PubMed}
\usage{
citedby_pubmed(
  id = NULL,
  web_history = NULL,
  by_id = FALSE,
  config = NULL,
  ...
)
}
\arguments{
\item{id}{vector with unique ID(s) for records in database \code{db}.}

\item{web_history}{a web_history object}

\item{by_id}{logical If FALSE (default) return a single 
\code{elink} objects containing links for all of the provided \code{id}s. 
Alternatively, if TRUE return a list of \code{elink} objects, one for each 
ID in \code{id}.}

\item{config}{vector configuration options passed to httr::GET}

\item{...}{character Additional terms to add to the request, see NCBI
documentation linked to in references for a complete list}
}
\description{
Lists the publications citing those specified by \code{web_history} or \code{id},
potentially split \code{by_id}. \code{pubmed_citedby} can handle cases where the
number of input IDs is >200.
}
