% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citedby.R
\name{citedby_pubmed}
\alias{citedby_pubmed}
\title{Get Cited By List from PubMed}
\usage{
citedby_pubmed(
  id = NULL,
  web_history = NULL,
  by_id = FALSE,
  no_result = "warning",
  save_raw = NULL,
  overwrite = FALSE,
  config = NULL,
  version = "2.0",
  retmode = "xml",
  ...
)
}
\arguments{
\item{id}{PubMed IDs for which to retrieve "cited by", as a character vector.}

\item{web_history}{An Entrez History object to be used instead of \code{id}, as
generated by the modified \code{\link[rentrez:entrez_post]{rentrez::entrez_post()}} fork at
https://github.com/allenbaron/rentrez (see NOTE); should be used when
"cited by" information is desired for >200 IDs.}

\item{by_id}{Whether to split "cited by" publications by the ID they
cite, as a boolean (default: \code{TRUE}); \code{FALSE} returns a unified list of
\emph{unique} cited by publications \emph{without} reference to the ID they cite.}

\item{no_result}{The type of condition that should be signaled when no PubMed
results exist in a response; one of "error" (default), "warning",
"message" or "none".}

\item{save_raw}{The \code{.rda} file path for saving the unprocessed citation
response, as a string. Not saved if \code{NULL} (default).}

\item{overwrite}{Whether an existing unprocessed citation response should be
overwritten; one of \code{TRUE} or \code{FALSE} (default).}

\item{config}{Configuration options passed to \code{\link[httr:GET]{httr::GET()}}.}

\item{version}{Argument included here for flexibility,
but best left with the defaults as set; see \code{\link[rentrez:entrez_summary]{rentrez::entrez_summary()}}
for details.}

\item{retmode}{"xml" (default) or "json"; "xml" is preferred because of a
higher response limit. This default is opposite the
\code{\link[rentrez:entrez_summary]{rentrez::entrez_summary()}} default.}

\item{...}{Additional terms to add to the request, see NCBI documentation on
\href{https://www.ncbi.nlm.nih.gov/books/NBK25499/#_chapter4_ESearch_}{ESearch}
for a complete list.}
}
\description{
List PubMed publications citing those specified by \code{id}, optionally split
\code{by_id}.
}
\seealso{
Other citedby_functions: 
\code{\link{citedby_scopus}()}
}
\concept{citedby_functions}
