% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citedby.R
\name{citedby_scopus}
\alias{citedby_scopus}
\title{Get Cited By List from Scopus}
\usage{
citedby_scopus(
  title,
  by_id = FALSE,
  id = NULL,
  api_key = NULL,
  view = "STANDARD",
  start = 0,
  count = NULL,
  max_count = 20000,
  headers = NULL,
  wait_time = 0,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{title}{Scopus Search API REFTITLE() query; see
https://dev.elsevier.com/sc_search_tips.html}

\item{by_id}{logical; where FALSE (default) prescribes a single query
(multiple titles are separated by OR) and TRUE prescribes a separate query
for each title}

\item{id}{(optional) vector of unique IDs; \emph{ignored if} \code{by_id = FALSE}}

\item{api_key}{API Key for Elsevier}

\item{view}{type of view to give, see
\url{https://api.elsevier.com/documentation/ScopusSearchAPI.wadl}}

\item{start}{where should the records start gathering}

\item{count}{number of records to retrieve (below 200 for STANDARD,
below 25 for COMPLETE views, see
\url{https://dev.elsevier.com/api_key_settings.html}).}

\item{max_count}{Maximum count of records to be returned.}

\item{headers}{additional headers to be added to
\code{\link{add_headers}}}

\item{wait_time}{The time in seconds to wait across consecutive
requests of a single search (when records > 25)}

\item{verbose}{Print diagnostic messages}

\item{...}{Arguments to be passed to the query list for
\code{\link{GET}}}
}
\value{
A list; if \code{by_id = FALSE}, the list result from the Scopus Search
API as produced by \link[rscopus:scopus_search]{rscopus::scopus_search}; if \code{by_id = TRUE} a list of
\code{rscopus::scopus_search} list results labelled numerically or with \code{id}, if
specified.
}
\description{
List Scopus publications citing those specified by \code{title} or \code{eid},
potentially split \code{by_id}.
}
