% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citedby.R
\name{citedby_scopus}
\alias{citedby_scopus}
\title{Get Cited By List from Scopus}
\usage{
citedby_scopus(
  title,
  by_id = FALSE,
  id = NULL,
  api_key = NULL,
  view = "STANDARD",
  start = 0,
  count = NULL,
  max_count = 20000,
  headers = NULL,
  wait_time = 0,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{title}{Publication title(s) to be used in a Scopus Search API
REFTITLE() query, as a character vector; see
https://dev.elsevier.com/sc_search_tips.html.}

\item{by_id}{Whether to split "cited by" publications by the ID they
cite, as a boolean (default: \code{TRUE}); \code{FALSE} returns a unified list of
\emph{unique} cited by publications \emph{without} reference to the ID they cite.}

\item{id}{Unique IDs used to identify \code{title} input, as a character vector of
the same length as \code{title}; \emph{REQUIRED} if \code{by_id = TRUE}, otherwise
ignored.}

\item{api_key}{API Key for Elsevier}

\item{view}{type of view to give, see
\url{https://api.elsevier.com/documentation/ScopusSearchAPI.wadl}}

\item{start}{where should the records start gathering}

\item{count}{number of records to retrieve (below 200 for STANDARD,
below 25 for COMPLETE views, see
\url{https://dev.elsevier.com/api_key_settings.html}).}

\item{max_count}{Maximum count of records to be returned.}

\item{headers}{Additional headers to be added by \code{\link[httr:add_headers]{httr::add_headers()}}.}

\item{wait_time}{The time in seconds to wait across consecutive
requests of a single search (when records > 25)}

\item{verbose}{Print diagnostic messages}

\item{...}{Named arguments to be passed to \code{\link[httr:GET]{httr::GET()}}. Available arguments
are listed in the
\href{https://dev.elsevier.com/documentation/ScopusSearchAPI.wadl}{Scopus Search Documentation: API Specification}.}
}
\value{
If \code{by_id = FALSE}, the list result from the Scopus Search API (as
produced by \link[rscopus:scopus_search]{rscopus::scopus_search}). If \code{by_id = TRUE}, an \code{id} named
list of \code{scopus_search} result lists.
}
\description{
List Scopus publications citing those specified by \code{title}, optionally split
\code{by_id}.
}
\seealso{
Other citedby: 
\code{\link{citedby_pubmed}()}
}
\concept{citedby}
