% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_col.R
\name{collapse_col}
\alias{collapse_col}
\title{Collapse Column}
\usage{
collapse_col(df, .col, delim = "; ")
}
\arguments{
\item{df}{a data.frame}

\item{.col}{the name of the column in the data.frame to collapse}

\item{delim}{delimiter to place between vector elements}
}
\description{
Collapse Column
}
