% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_col}
\alias{collapse_col}
\title{Collapse Column(s)}
\usage{
collapse_col(df, .cols, delim = "|", method = "unique", na.rm = FALSE)
}
\arguments{
\item{df}{a data.frame}

\item{.cols}{the name of the column in the data.frame to collapse}

\item{delim}{A delimiter to place between vector elements (default: "|").}

\item{na.rm}{A logical scalar indicating whether \code{NA} values should be
removed (default: \code{FALSE}).}
}
\value{
A data.frame with the specified columns collapsed. Also \strong{NOTE} the
following:
\enumerate{
\item Collapsed columns \strong{will} be converted to \code{character}.
\item Rows will be reordered by the unique combination of columns \emph{not}
collapsed (due to \code{\link[dplyr:group_by]{dplyr::group_by()}} use).
}
}
\description{
Collapses values in the column(s) specified, concatenating unique values in
those columns for each record (where a record is defined as a unique
row including all columns \emph{NOT} specified in \code{.cols}).
}
\examples{
cc_df <- tibble::tibble(
     x = c(1, 2, 3, 3, 4, 4, 4),
     y = c("a", "a", "b", "b", "c", "c", "e"),
     z = c("Z", "Y", "X", "X", "W", "V", "U")
)
cc_df

# completely duplicated rows (3-4) are collapsed with any .cols specified
collapse_col(cc_df, y)

collapse_col(cc_df, x)
collapse_col(cc_df, z)
collapse_col(cc_df, c(x, z))

}
