% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_col.R
\name{collapse_col}
\alias{collapse_col}
\title{Collapse Column(s)}
\usage{
collapse_col(df, .cols, delim = "; ")
}
\arguments{
\item{df}{a data.frame}

\item{delim}{delimiter to place between vector elements}

\item{.col}{the name of the column in the data.frame to collapse}
}
\description{
Collapses values in the column(s) specified, concatenating unique values in
those columns for each record (where a record is defined as a unique
row including all columns \emph{NOT} specified in \code{.cols}).
}
\examples{
cc_df <- tibble::tibble(
     x = c(1, 2, 3, 3, 4, 4, 4),
     y = c("a", "a", "b", "b", "c", "c", "e"),
     z = c("Z", "Y", "X", "X", "W", "V", "U")
)
cc_df

# completely duplicated rows (3-4) are collapsed with any .cols specified
collapse_col(cc_df, y)

collapse_col(cc_df, x)
collapse_col(cc_df, z)
collapse_col(cc_df, c(x, z))

}
