% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confine_list.R
\name{confine_list}
\alias{confine_list}
\alias{release_list}
\title{Convert (Nested) List to/from Character Vector}
\usage{
confine_list(.list)

release_list(.json_chr)
}
\arguments{
\item{.list}{a (nested) list}

\item{.json_chr}{a character vector containing JSON representation of list
elements below the top level}
}
\description{
These functions convert a (nested) list to or from a character vector of the
same length, preserving or restoring the list structure and values with JSON
strings. They are primarily designed to preserve the structure of simple
nested lists (no internal data.frames, matrices, or arrays), as would result
from an API call converted to a list without any conversion
(e.g. flattening).
}
\section{NOTE}{

These could potentially be made to handle more diverse lists by including
options that control how data.frames & matrices are managed, ensuring
to/from procedures are linked and occur in predictable ways (see
arguments for \href{jsonlite::fromJSON}{jsonlite::to/fromJSON} for details).
}

