% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alliance.R
\name{count_alliance_records}
\alias{count_alliance_records}
\title{Count Alliance Records}
\usage{
count_alliance_records(alliance_tbl, by_type = TRUE, pivot = TRUE)
}
\arguments{
\item{alliance_tbl}{a dataframe derived from Alliance data (usually a
\href{https://www.alliancegenome.org/downloads}{downloaded .tsv file})}

\item{by_type}{logical indicating whether to count by object type}

\item{pivot}{logical indicating whether to pivot values to type columns;
ignored if type = FALSE}
}
\value{
A summary tibble.
}
\description{
Counts records in data from the Alliance of Genome Resources by model
organism database (MOD) and, optionally, by object type. A record, as used
here, is a row in the data (after removing likely duplicates, see NOTE).
}
\section{NOTE}{

For disease-related data, some exact duplicates (reason unknown) and records
that differ by seemingly unimportant information (e.g. only the date differs)
have existed. These types of duplicates are removed prior to record counts.
}

