% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation.R
\name{curation_template}
\alias{curation_template}
\alias{curation_template.NULL}
\title{Create a Curation Template}
\usage{
curation_template(.data = NULL, ss = NULL, sheet = NULL, ...)

\method{curation_template}{`NULL`}(.data = NULL, ss = NULL, sheet = NULL, ..., nrow = 50)
}
\arguments{
\item{.data}{Data to add to the curation sheet. If \code{NULL} (default), an empty
curation sheet will be created.}

\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:drive_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{(OPTIONAL) The sheet name, as a string. If \code{NULL} (default), the
sheet name will default to "curation-" with today's date appended (formatted
as "\%Y\%m\%d"; see \code{\link[=format.Date]{format.Date()}}).}

\item{...}{Additional arguments passed to methods.}

\item{nrow}{The number of rows to create in the curation template when
\code{.data = NULL} (default: \code{50}).}
}
\value{
The Google Sheet info (\code{ss}), as a \link[googlesheets4:sheets_id]{googlesheets4::sheets_id}.
}
\description{
Create a curation template in a Google Sheet, optionally including data.
}
