% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.utils-package.R
\docType{package}
\name{DO.utils-package}
\alias{DO.utils}
\alias{DO.utils-package}
\title{Human Disease Ontology (DO) support package}
\description{
The DO.utils packages provides functions that assist in three main areas:
\enumerate{
\item Managing DO tools/builds
\item Assessing DO's impact on science (bibliometrics/scientometrics)
\item Making R programming simpler/faster
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://diseaseontology.github.io/DO.utils/}
}

}
\keyword{internal}
