% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.utils-package.R
\docType{package}
\name{DO.utils-package}
\alias{DO.utils}
\alias{DO.utils-package}
\title{DO.utils: Public Resource Utilities Designed by/for the Disease Ontology}
\description{
Generally useful tools to assess use of public resources in scientific literature ("cited by", search, etc.) and to simplify common R tasks, along with tools specifically designed for analysis and improvement of the Human Disease Ontology (DO; https://disease-ontology.org).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://diseaseontology.github.io/DO.utils/}
}

}
\keyword{internal}
