% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyDOID.R
\name{DOrepo}
\alias{DOrepo}
\title{Instantiate a DO Repository}
\usage{
DOrepo(path)
}
\arguments{
\item{path}{The local path to the
\href{https://github.com/DiseaseOntology/HumanDiseaseOntology}{HumanDiseaseOntology}
git repository, as a string.}
}
\description{
Instantiate a \code{DOrepo} object from the
\href{https://github.com/allenbaron/pyDOID}{pyDOID} package (powered by
\href{https://github.com/rstudio/reticulate}{reticulate}).
}
\section{Classes & Methods}{

The \code{DOrepo} class inherits the methods & subclasses of the
\href{https://gitpython.readthedocs.io/en/stable/tutorial.html#meet-the-repo-type}{GitPython.Repo}
class, along with the following:
\itemize{
\item \verb{DOrepo$tag_iterate(function, start = [str], end = [str])}: Iterate
through DO releases (tags) from \code{start} to \code{end} executing a \code{function} at
each. \code{start} and \code{end} are \emph{optional}; if unspecified the first or last
release/tag of the repo will be assumed.
\item \code{DOrepo$capture_head()}: Capture the current \code{head} of the repo; useful
for restoring state after git checkout(s). \emph{USE}: If assigned to variable
\code{x}, restore state with \code{x$checkout()}.
\item \code{DOrepo$doid} and \code{DOrepo$doid_merged}: Access the doid.owl and
doid-merged.owl files of the repository. These inherit from the
\code{pyDOID.owl.xml} class, see \code{\link[=owl_xml]{owl_xml()}} for methods. \emph{NOTE} that, unlike
\code{\link[=owl_xml]{owl_xml()}}, doid.owl and doid-merged.owl are not loaded into memory upon
instantiation of \code{DOrepo}, but are loaded automatically when the first
\verb{$query()} is executed or manually via \verb{$load()}.
}
}

\seealso{
Other pyDOID classes: 
\code{\link{owl_xml}()}
}
\concept{pyDOID classes}
