% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alliance.R
\name{download_alliance_tsv}
\alias{download_alliance_tsv}
\title{Download Alliance .tsv.gz File}
\usage{
download_alliance_tsv(dest_dir, url = NULL)
}
\arguments{
\item{dest_dir}{path to directory where file will be saved}

\item{url}{URL to Alliance file; if not provided, will be requested at console}
}
\value{
Path to saved file.
}
\description{
Downloads a URL-specified .tsv.gz file from the Alliance of Genome
Resources. Files can be found at
\url{https://www.alliancegenome.org/downloads}. Right-click on the "tsv" link
of a desired file and select "Copy Link" to get the file URL.
}
\details{
A date stamp indicating download date is added to the base file name.
}
\section{Recommendation}{

Although it's possbile to directly read a file from the URL, downloading it
promotes reproducibility and ensures future access if needed.
}

