% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_file}
\alias{download_file}
\title{Download File(s) from the Internet}
\usage{
download_file(url, dest_file, on_failure = "warn", ...)
}
\arguments{
\item{url}{A character vector naming the URL of resource(s) to be downloaded.}

\item{dest_file}{A character vector with the file path(s) where downloaded
file(s) will be saved.}

\item{on_failure}{A string indicating how to handle download failure.}

\item{...}{Additional arguments passed on to \code{\link[=download.file]{download.file()}}.}
}
\value{
A character vector of paths to destination files that downloaded
successfully along with a warning indicating which failed when
\code{abort = FALSE} (default).
}
\description{
\code{download_file()} is a vectorized version of \code{\link[utils:download.file]{utils::download.file()}} that
does not rely on specifying the "libcurl" method for vectorization.
}
