% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\docType{class}
\name{download_status-class}
\alias{download_status-class}
\alias{download_status}
\title{File Download Status (RefClass)}
\description{
A reference class to track file download status.
}
\section{Fields}{

\describe{
\item{\code{successful}}{A character vector for paths of successfully downloaded
files.}

\item{\code{failed}}{A character vector for paths of files that failed to download}

\item{\code{fail_status}}{An integer vector with an exit code from
\code{\link[utils:download.file]{utils::download.file()}} for each file in \code{failed}.}
}}

\section{Methods}{

\describe{
\item{\code{check(status, url, dest_file, abort = FALSE)}}{Check download status of file with choice to abort on failure.}

\item{\code{return(w_failed = FALSE)}}{Return successful file paths and, optionally, failed URLs.}

\item{\code{warn()}}{Warn about failed downloads.}
}}

