% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_validation.R
\name{extract_doid_url}
\alias{extract_doid_url}
\title{Extract URLs in DO (INTERNAL)}
\usage{
extract_doid_url(
  doid_edit,
  w_raw_match = FALSE,
  quiet = FALSE,
  show_url_string = FALSE
)
}
\arguments{
\item{doid_edit}{The contents of the doid-edit.owl file, as a character
vector (as provided by \code{\link[=read_doid_edit]{read_doid_edit()}}).}

\item{w_raw_match}{Whether to include the full line of doid-edit.owl where
each URL was extracted from, as a boolean (default: \code{FALSE}).}

\item{quiet}{Whether warnings for multiple DOIDs associated with a single
URL should be silenced, as a boolean (default: \code{FALSE}).}

\item{show_url_string}{Whether the complete URL string, including prefix
should be returned, as a boolean (default: \code{FALSE}).}
}
\value{
A tibble of DOIDs and their associated URLs.
}
\description{
Extract URLs from the doid-edit.owl file of the Human Disease Ontology.
}
