% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axiom.R
\name{extract_eq_axiom}
\alias{extract_eq_axiom}
\title{Extract Equivalent Class Axioms}
\usage{
extract_eq_axiom(DO_repo)
}
\arguments{
\item{DO_repo}{The local path to the \code{HumanDiseaseOntology} repo, as a
string.}
}
\value{
Equivalent class axioms in OWL functional format, as a character vector.
}
\description{
Extract \code{owl:equivalentClass} axioms from the doid-edit.owl file.
}
