% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pmid.R
\name{extract_pmid.elink}
\alias{extract_pmid.elink}
\title{Extract PubMed ID from elink object}
\usage{
\method{extract_pmid}{elink}(x, linkname = NULL, quietly = FALSE, ...)
}
\arguments{
\item{x}{An \code{elink} object, generated by \link[rentrez:entrez_link]{rentrez::entrez_link} and it's
derivatives (e.g. \link{citedby_pmid}).}

\item{linkname}{linkname as specified for \href{https://eutils.ncbi.nlm.nih.gov/entrez/query/static/entrezlinks.html}{Entrez API};
only necessary if more than one \code{linkname} is returned from PubMed.}

\item{quietly}{Suppress PubMed linkname message when \code{linkname} is not
specified and multiple results exist in an \code{elink} object.}

\item{...}{Unused, included for generic consistency only.}
}
\description{
Extract PubMed ID from an \code{elink} object.
}
