% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pmid.R
\name{extract_pmid.elink_list}
\alias{extract_pmid.elink_list}
\title{Extract PubMed ID from elink_list object}
\usage{
\method{extract_pmid}{elink_list}(x, ...)
}
\arguments{
\item{x}{An \code{elink_list} object, generated by \link[rentrez:entrez_link]{rentrez::entrez_link} and it's
derivatives (e.g. \link{citedby_pubmed}) when \code{by_id = TRUE} .}

\item{...}{Additional arguments passed on to \link{extract_pmid.elink} method.}
}
\description{
Extract PubMed ID from an \code{elink_list} object.
}
