% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pmid.R
\name{extract_pmid}
\alias{extract_pmid}
\title{Extract PubMed ID}
\usage{
extract_pmid(x, ...)
}
\arguments{
\item{x}{rentrez API result}

\item{...}{Arguments passed on to methods.}
}
\description{
\code{extract_pmid} is a generic function that extracts a vector of PubMed IDs from
Entrez Utilities API results obtained via the \code{\link[rentrez:rentrez]{rentrez}}
package.
}
