% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_url.R
\name{extract_resp_url}
\alias{extract_resp_url}
\title{Extract a Response URL (INTERNAL)}
\usage{
extract_resp_url(x, .which = "last")
}
\arguments{
\item{x}{An \link[httr:response]{httr::response} object.}

\item{.which}{The URL to extract, as a string. Available options include:
\itemize{
\item "first": the initial request URL.
\item "last": the final URL visited (last redirect).
}}
}
\description{
Extract a URL from a response generated by \href{httr::httr}{httr}.
}
\keyword{internal}
