% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axiom.R
\name{extract_subclass_axiom}
\alias{extract_subclass_axiom}
\title{Extract 'Subclass Of' Axioms}
\usage{
extract_subclass_axiom(DO_repo)
}
\arguments{
\item{DO_repo}{The local path to the \code{HumanDiseaseOntology} repo, as a
string.}
}
\value{
'Subclass Of' axioms in OWL functional format, as a character vector.
}
\description{
Extract \code{owl:subClassOf} axioms from the doid-edit.owl file.
}
