% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_subtree}
\alias{extract_subtree}
\title{Extract Subtree (DEPRECATED)}
\usage{
extract_subtree(x, top_node, reload = FALSE)
}
\arguments{
\item{x}{A 'pyDOID.owl.xml' object or the path to an OWL/RDF XML file that
can be instantiated as such an object by \code{\link[=owl_xml]{owl_xml()}}.}

\item{top_node}{The top node of the tree, as a valid DOID (see
\code{\link[=is_valid_doid]{is_valid_doid()}} for valid input formats).}

\item{reload}{Force reload the file into memory, as \code{TRUE} or \code{FALSE}
(default).}
}
\value{
A \link[tibble:tibble]{tibble} with the columns: \code{id}, \code{label}, \code{parent_id},
and \code{parent_label}, with one row for each unique combination for each
subclass below and including \code{top_node}.
}
\description{
Extracts the classes and parents of a DO subtree from a \code{pyDOID.owl.xml}
object.
}
\seealso{
\link[=extract_subclass]{extract_subclass(method = "descendants")} for the
same result with a single \code{class}. \code{\link[=format_subtree]{format_subtree()}} to arrange data in a
tree structure similar to ontology browsers.
}
