% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{fill_subclass}
\alias{fill_subclass}
\title{Fill in Subclasses}
\usage{
fill_subclass(subtree_df, debug = FALSE)
}
\arguments{
\item{subtree_df}{A dataframe from \code{\link[=extract_subtree]{extract_subtree()}}.}

\item{debug}{Whether original and new identifiers should both be returned in
the output dataframe for debugging purposes, \code{TRUE} or \code{FALSE}
(default).}
}
\description{
Fill in subclasses when a parent appears multiple times in a subtree. This
ensures that the subclasses appear each time their parent/superclass does.
}
\keyword{internal}
