% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_doid}
\alias{format_doid}
\title{Format DOIDs}
\usage{
format_doid(x, as = "CURIE", convert_bare = FALSE, validate_input = TRUE)
}
\arguments{
\item{x}{A set of IDs, as a character vector.}

\item{as}{The format to convert the DOIDs to, as a string. All valid formats
are possible options: "CURIE" (default), "URI", "obo_CURIE", "basename".}

\item{convert_bare}{Whether bare numbers should be converted to canonical
DOIDs, \code{TRUE} or \code{FALSE} (default).}

\item{validate_input}{Whether to ensure only valid DOIDs are included in \code{x},
\code{TRUE} (default) or \code{FALSE}. When \code{FALSE}, non-DOID input will be
returned unchanged.}
}
\description{
Convert valid DOIDs and/or bare numbers to a specified DOID format. Input
\emph{may} be tested to ensure it matches a valid DOID format but no attempt is
made to confirm bare numbers or DOIDs match actual diseases in the ontology.
}
\section{Caution}{

Be extra cautious when using \code{format_doid()} with \code{validate_input = FALSE} as
unexpected text conversion may occur.
}

\examples{
x <- c(
    "http://purl.obolibrary.org/obo/DOID_0001816",
    "DOID:4",
    "obo:DOID_14566",
    "DOID_0040001"
)

format_doid(x, as = "URI")
format_doid(x, as = "CURIE")
format_doid(x, as = "obo_CURIE")
format_doid(x, as = "basename")

# bare numbers can be converted to canonical DOIDs, if desired
w_bare <- c(x, "0050117")
format_doid(w_bare, convert_bare = TRUE)

# non-DOIDs can be passed as input, if desired
mixed_input <- c(x, "random_text", "obo:SYMP_0000000", "0050117")
format_doid(mixed_input, validate_input = FALSE)
format_doid(mixed_input, convert_bare = TRUE, validate_input = FALSE)

}
\seealso{
Other format IDs: 
\code{\link{format_obo}()}
}
\concept{format IDs}
