% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_hyperlink}
\alias{format_hyperlink}
\title{Format URLs as Hyperlinks}
\usage{
format_hyperlink(url, txt, as, ..., preserve_NA = TRUE)
}
\arguments{
\item{url}{One or more URLs, as a character vector.}

\item{txt}{The text to display for each link, as a character vector.}

\item{as}{What format to use for the hyperlink, as a string; one of "gs"
(Google Sheet), "xlsx" (Excel), or "html".}

\item{...}{One or more name-value pairs of html \verb{<a>}
\href{https://www.w3schools.com/tags/tag_a.asp}{attributes}.}

\item{preserve_NA}{Whether to preserve \code{NA} in output, as a boolean. \code{FALSE}
will result in hyperlinks built from \code{NA} values in \code{url} (almost
certainly not desired).}
}
\description{
Formats URLs as hyperlinks for Google Sheets, Excel, or html.
}
\section{Excel Note}{

Use the \code{openxlsx} pkg to write data with hyperlinks to Excel.
}

\examples{
format_hyperlink("https://www.google.com", "Google", "gs")
format_hyperlink("https://www.google.com", "Google", "xlsx")
format_hyperlink("https://www.google.com", "Google", "html")

# html with `<a>` attributes
format_hyperlink(
    "https://www.google.com",
    "Google",
    "html",
    target = "_blank",
    rel = "external"
)

}
