% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_impact.R
\name{get_bioc_pkg_stats}
\alias{get_bioc_pkg_stats}
\title{Get Bioconductor Package Statistics}
\usage{
get_bioc_pkg_stats(pkg, pkg_type, yr, delay_rng)
}
\arguments{
\item{pkg}{character vector; name of Bioconductor package(s)}

\item{pkg_type}{character vector of same length as \code{pkg} identifying the type
of Bioconductor package; one of "software", "annotation", "experiment", or
"workflow"}

\item{yr}{character scalar; 4-digit year of stats desired,
default = current year (currently implemented only for single yr)}

\item{delay_rng}{\code{c(min, max)} number of seconds to wait between requests;
default = \code{c(1, 10)}}
}
\description{
Get download statistics for specific Bioconductor packages.
}
