% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{get_url}
\alias{get_url}
\title{Get URL (internal)}
\usage{
get_url(.name)
}
\arguments{
\item{.name}{Internal name of desired URL.}
}
\description{
Get a URL used within this package. List all possible names organized by
type with \code{get_url("names")}.
}
\section{Possible Names (organized by type)}{

\itemize{
\item pub_urls: doi, github, orcid, pubmed, pmc
\item xref_urls: ICD9CM, ICD10CM, KEGG, MESH, NCI, ORDO, OMIM, SNOMEDCT_US, UMLS_CUI
\item web: DO_website
\item data_urls: alliance_disease_tsv
\item ns_prefix: ...includes common OBO Foundry and SPARQL prefixes; use \code{ns_prefix} to see the complete list.
}
}

\section{NOTE}{

\code{get_url()} provides prefixes for disease-ontology.org style link support.
The prefixes are primarily for cross-references (\code{xrefs}) and are generally
better for creating URLs to look up information about a particular entity
online but they \emph{may} overlap with official namespaces in \link{ns_prefix}, which
can also be accessed by \code{get_url()}. Where this occurs, the prefixes can
generally be distinguished by capitalization, with xref prefixes in
uppercase, e.g. "MESH" = "https://meshb.nlm.nih.gov/record/ui?ui=",
and the official namespace in lowercase,
e.g. "mesh" = "https://id.nlm.nih.gov/mesh/". As this
distinction cannot always be guaranteed, care should be taken for any prefix
related to xrefs in the DO.
}

\keyword{internal}
