% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{get_url}
\alias{get_url}
\title{Get URL (internal)}
\usage{
get_url(.name)
}
\arguments{
\item{.name}{internal name of desired URL}
}
\description{
Get a URL used within this package. Available URLs: "doi", "github", "orcid",
"pubmed", "pmc", "pmc_article" (append as prefix to article pmcid for direct
navigation), or "alliance_disease_tsv".
}
\keyword{internal}
