% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{iff_all_vals}
\alias{iff_all_vals}
\title{Test for All Values}
\usage{
iff_all_vals(x, values)
}
\arguments{
\item{x}{A vector.}

\item{values}{The values to ensure exist in \code{x}.}
}
\value{
\code{TRUE} or \code{FALSE}. When \code{FALSE}, \code{missing} and/or \code{extra} attributes will be
included to assist in identifying non-conformity.
}
\description{
Returns \code{TRUE} if, and only if, all \code{values} are present in \code{x} and ONLY
those \code{values} are present in \code{x}.
}
