% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_omim}
\alias{inventory_omim}
\title{Assess whether OMIM Data is in DO}
\usage{
inventory_omim(onto_path, omim_input)
}
\arguments{
\item{onto_path}{The path to an ontology file, as a string.}

\item{omim_input}{An \code{omim_tbl} created by \code{\link[=read_omim]{read_omim()}} or the path to a
.tsv or .csv file (possibly compressed) that can be read by \code{\link[=read_omim]{read_omim()}} and
includes OMIM data to compare against the mappings in the ontology.}
}
\value{
The \code{omim_input} with 4 columns added:
\itemize{
\item \code{exists}: Boolean indicating whether an OMIM ID is present in the DO.
\item \code{mapping_type}: The mapping predicate(s) of this OMIM ID to a disease, if
present. Multiple predicate(s) will be pipe delimited.
\item \code{doid}: The DOID of the disease mapped to this OMIM ID, if present.
\item \code{do_label}: The label of the disease mapped to this OMIM ID, if present.
\item \code{do_dep}: Boolean indicating whether a disease is deprecated or not, if
present.
}

Output will have the class \code{omim_inventory}, a type of class
\code{mapping_inventory}.
}
\description{
Assesses whether OMIM identifiers are present in the Human Disease Ontology
as mappings (either xrefs or skos mappings). Utilizes \code{\link[=robot]{robot()}} for
comparison.
}
\examples{
\dontrun{
inventory_omim(
    onto_path = "~/Ontologies/HumanDiseaseOntology/src/ontology/doid-edit.owl",
    omim_input = "omimps.csv", # e.g. manually copied from https://www.omim.org/phenotypicSeries/PS609060
)
}

}
