% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory_omim}
\alias{inventory_omim}
\title{Assess whether OMIM Data is in DO}
\usage{
inventory_omim(
  onto_path,
  omim_input,
  keep_mim = c("#", "\%"),
  include_hasDbXref = TRUE
)
}
\arguments{
\item{onto_path}{The path to an ontology file, as a string.}

\item{omim_input}{An \code{omim_tbl} created by \code{\link[=read_omim]{read_omim()}} or the path to a
.tsv or .csv file (possibly compressed) that can be read by \code{\link[=read_omim]{read_omim()}} and
includes OMIM data to compare against the mappings in the ontology.

NOTE: If an \code{omim_tbl} is provided, \code{keep_mim} will be ignored.}

\item{keep_mim}{[\strong{OMIM search data only}] The MIM symbols representing
the data types to keep, as a character vector, or \code{NULL} to retain all
(default: \code{"#"} and \code{"\%"}).

The \href{https://www.omim.org/help/faq#1_3}{OMIM} defined MIM symbols are:\tabular{ll}{
   MIM symbol \tab MIM type \cr
   \code{*} \tab gene \cr
   \code{+} \tab gene, includes phenotype \cr
   \verb{#} \tab phenotype \cr
   \verb{\%} \tab phenotype, unknown molecular basis \cr
   \code{^} \tab deprecated \cr
   \code{none} \tab phenotype, suspected/overlap \cr
}}

\item{include_hasDbXref}{Whether \code{oboInOwl:hasDbXref}'s should be included in
tests for "one-to-multiple" mappings, as a boolean (default: \code{TRUE}).
\code{skos:exactMatch} & \code{skos:closeMatch} mappings are always included.}
}
\value{
The \code{omim_input} with 5 additional columns:
\itemize{
\item \code{exists}: Logical indicating whether an OMIM ID is present in the DO.
\item \code{mapping_type}: The mapping predicate(s) of this OMIM ID to a disease, if
present. Multiple predicate(s) between the same OMIM and DOID will be pipe
delimited.
\item \code{doid}: The DOID of the disease mapped to this OMIM ID, if present.
\item \code{do_label}: The label of the disease mapped to this OMIM ID, if present.
\item \code{do_dep}: Logical indicating whether a disease is deprecated or not, if
present.
\item \code{multimaps}: The direction in which an OMIM or DO term maps to multiple
terms in the other resource, as "omim_to_doid", "doid_to_omim", "both_ways"
or \code{NA}.
}

Output will have the class \code{omim_inventory}, a type of class
\code{mapping_inventory}.
}
\description{
Assesses whether OMIM identifiers are present in the Human Disease Ontology
as mappings (either xrefs or skos mappings). Utilizes \code{\link[=robot]{robot()}} for
comparison.
}
\examples{
\dontrun{
# manually copy or download data from https://www.omim.org/phenotypicSeries/PS609060
inventory_omim(
    onto_path = "~/Ontologies/HumanDiseaseOntology/src/ontology/doid-edit.owl",
    omim_input = "omimps.csv",
)
}

}
