% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_predicates.R
\name{is_invariant}
\alias{is_invariant}
\alias{is_invariant.character}
\alias{is_invariant.numeric}
\title{Test if Vector is Invariant}
\usage{
is_invariant(x, na.rm = FALSE, ...)

\method{is_invariant}{character}(x, na.rm = FALSE, ...)

\method{is_invariant}{numeric}(x, na.rm = FALSE, tol = sqrt(.Machine$double.eps), ...)
}
\arguments{
\item{x}{vector to be tested}

\item{na.rm}{logical indicating whether to exclude NA values}

\item{...}{unused; for extensibility}

\item{tol}{double, tolerance to use (for numeric vectors)}
}
\description{
Test if a vector is invariant (\emph{i.e.} all values are equal, within a given
tolerance for numeric vectors).
}
\seealso{
Other type_predicates: 
\code{\link{char_val_predicates}},
\code{\link{lgl_predicates}},
\code{\link{num_val_predicates}}
}
\concept{type_predicates}
