% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_predicates.R
\name{is_invariant}
\alias{is_invariant}
\title{Invariant test}
\usage{
is_invariant(..., na.rm = FALSE)
}
\arguments{
\item{...}{vectors of values}

\item{na.rm}{if \code{TRUE} missing values don't count}
}
\description{
Test if a vector is invariant, i.e. has only one value
}
