% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_invariant}
\alias{is_invariant}
\alias{is_invariant.default}
\alias{is_invariant.numeric}
\alias{is_invariant.list}
\alias{is_invariant.data.frame}
\title{Test if an Object is Invariant}
\usage{
is_invariant(x, ...)

\method{is_invariant}{default}(x, na.rm = FALSE, ...)

\method{is_invariant}{numeric}(x, na.rm = FALSE, tol = sqrt(.Machine$double.eps), ...)

\method{is_invariant}{list}(x, incl_nm = TRUE, ...)

\method{is_invariant}{data.frame}(x, ...)
}
\arguments{
\item{x}{object to be tested}

\item{...}{unused; for extensibility}

\item{na.rm}{logical indicating whether to exclude NA values}

\item{tol}{double, tolerance to use (for numeric vectors)}

\item{incl_nm}{Whether top-level names should be included in determining if a
list is invariant (default: \code{TRUE}).}
}
\description{
Test if an object is invariant (\emph{i.e.} all values are equal, within a given
tolerance for numeric vectors).
}
\seealso{
Other value predicates: 
\code{\link{char_val_predicates}},
\code{\link{num_val_predicates}}

Other predicates: 
\code{\link{all_duplicated}()},
\code{\link{char_val_predicates}},
\code{\link{iff_all_vals}()},
\code{\link{is_curie}()},
\code{\link{is_hex_color}()},
\code{\link{is_uri}()},
\code{\link{lgl_predicates}},
\code{\link{num_val_predicates}},
\code{\link{obo_ID_predicates}}
}
\concept{predicates}
\concept{value predicates}
