% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_uri}
\alias{is_uri}
\title{Test for Valid URIs}
\usage{
is_uri(x, empty_ok = TRUE)
}
\arguments{
\item{x}{A character vector.}

\item{empty_ok}{Whether to allow empty hosts and paths, as a boolean
(default: \code{TRUE}). RFC3986 allows for empty paths & hosts, potentially
making a scheme alone valid and, therefore, the default. However, it is
often desirable to validate that at least one of these is NOT empty, since
a scheme alone is rarely useful in practice.}
}
\value{
A logical vector indicating which values of \code{x} are valid URIs.
}
\description{
This is a vectorized predicate to test if character values are valid URLs,
i.e. have a scheme and at least one of: hostname or path (according to
\href{https://www.rfc-editor.org/rfc/rfc3986#section-3}{RFC3986}).
}
\section{Notes}{

While all URIs are valid CURIEs (see \code{is_curie(def = "w3c")}), not all CURIEs
are valid URIs (e.g. URIs cannot start with \verb{_}).
}

\examples{
.uri <- c(
    # always TRUE
    "http://purl.obolibrary.org/obo/DOID_0001816",
    "https://google.com",
    "mailto:fake.name@blah.com",
    # TRUE, if empty_ok = FALSE
    "file://",
    "mailto:",
    # never TRUE
    "blah",
    ""
)

is_uri(.uri)
is_uri(.uri, empty_ok = FALSE)

}
\seealso{
Other ID predicates: 
\code{\link{is_curie}()},
\code{\link{obo_ID_predicates}}

Other predicates: 
\code{\link{all_duplicated}()},
\code{\link{char_val_predicates}},
\code{\link{iff_all_vals}()},
\code{\link{is_curie}()},
\code{\link{is_hex_color}()},
\code{\link{is_invariant}()},
\code{\link{lgl_predicates}},
\code{\link{num_val_predicates}},
\code{\link{obo_ID_predicates}}
}
\concept{ID predicates}
\concept{predicates}
