% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{length_sort}
\alias{length_sort}
\title{Sort by Character Length}
\usage{
length_sort(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{
  Arguments passed on to \code{\link[base:order]{base::order}}
  \describe{
    \item{\code{decreasing}}{logical.  Should the sort order be increasing or
    decreasing? For the \code{"radix"} method, this can be a vector of
    length equal to the number of arguments in \code{\dots}. For the
    other methods, it must be length one.}
    \item{\code{na.last}}{for controlling the treatment of \code{NA}s.
    If \code{TRUE}, missing values in the data are put last; if
    \code{FALSE}, they are put first; if \code{NA}, they are removed
    (see \sQuote{Note}.)}
    \item{\code{method}}{the method to be used: partial matches are allowed.  The
    default (\code{"auto"}) implies \code{"radix"} for short numeric vectors,
    integer vectors, logical vectors and factors.
    Otherwise, it implies \code{"shell"}.
    For details of methods \code{"shell"}, 
    \code{"quick"}, and \code{"radix"},
    see the help for \code{\link[base]{sort}}.}
  }}
}
\description{
Sort the elements of a vector by character length. Multiple elements of the
the same length in a vector are secondarily sorted by appearance in the
original vector.
}
\examples{
x <- c("ccc", "aaaa", "eee", "b", "DDD")
length_sort(x)
length_sort(x, decreasing = TRUE)

x2 <- c(1:9, NA, 100, 10)
length_sort(x2)
length_sort(x2, decreasing = TRUE)
length_sort(x2, na.last = NA)

}
\seealso{
\code{\link[=length_order]{length_order()}} for sorting data.frames by character length.
}
