% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{make_use_case_html}
\alias{make_use_case_html}
\title{Make HTML for DO Use Case Tables}
\usage{
make_use_case_html(out_dir, .which = "all")
}
\arguments{
\item{out_dir}{The path to the directory where output should be saved, as a
string.}

\item{.which}{The list(s) to generate, as a character vector. One or more of:
"all" (default, to make all lists), "ontology", "resource", or
"methodology".}
}
\value{
One "html" file for each group in \code{.which} named as
"\{datestamp\}-\{.which\}.html".

Also returns the "User" data from the Google Sheet invisibly.

NOTE: The "html" is incomplete and cannot be loaded by browsers,
as is.
}
\description{
Makes the row and cell html code for the various sections/tables of the
disease-ontology.org "Use Cases" page from the DO team's "Uses" google sheet.
This function explicitly avoids including the html code for defining the
table itself to provide for flexibility. The "html" output in the files
specified should be copied and pasted into the disease-ontology.org
"Use Cases" file in the appropriate section/table.
}
