% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{make_user_list_html}
\alias{make_user_list_html}
\title{Make HTML for DO User List (DEPRECATED)}
\usage{
make_user_list_html(file)
}
\arguments{
\item{file}{The file path where the output should be saved, as a string.}
}
\description{
Makes the row and cell html code for the "Users of the Disease Ontology"
section of the collaborators page on disease-ontology.org from the DO team's
"Uses" google sheet. This function explicitly avoids including the html
code for defining the table itself to provide for flexibility.
}
\section{Deprecation Notice}{

The information this was formatting for disease-ontology.org was moved from
the "Collaborators" page to the new "Use Cases" page in mid-2022 and was
split from one section into three, making this function obsolete. Use
\code{\link[=make_use_case_html]{make_use_case_html()}} instead.
}

