% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_citations}
\alias{match_citations}
\title{Citation Matching}
\usage{
match_citations(x, ref, add_col = FALSE, nomatch = NA_integer_)
}
\arguments{
\item{x, ref}{a vector of PMID, PMCID, DOI, or Scopus IDs; or a "citation"
dataframe with 1 or more of these as columns (column names should correspond
to ID type; case-insensitive)}

\item{add_col}{logical; if FALSE (default), returns a vector; if TRUE,
\link[dplyr:mutate]{mutates} the data.frame by adding a \code{cite_match} column;
ignored when \code{x} is a vector}

\item{nomatch}{the value to be returned in the case when no match is
    found.  Note that it is coerced to \code{integer}.}
}
\value{
If \code{x} is a vector or \code{add_col} is FALSE (default), an integer vector of the
same length as \code{x} (if a data.frame, of length equal to the number of rows).
If \code{x} is a data.frame and \code{add_col} is TRUE, a mutated data.frame with a
\code{cite_match} integer column identifying match positions in \code{ref}.
}
\description{
Essentially, \code{\link[base:match]{base::match()}} but tailored to citations. Returns a vector of
the positions of (first) matches of a citation input (first argument)
in a reference citation input (second argument) based on common
publication IDs (PMID, PMCID, DOI).
}
\details{
The citation inputs can be vectors, data.frames, or a mix. For matching,
each input requires at least one vector containing one of the three standard,
consistent publication identifiers (PMID, PMCID, or DOI). Columns in
data.frames are identified by name (case-insensitive, e.g. 'PMID' or 'pmid').

When both inputs are data.frames, \code{match_citations()} will match on all
ID types present in both inputs, returning matches from the highest priority
ID type (priority: PMID > PMCID > DOI > Scopus EID). It also tells the user
what ID columns are identified and the order used in a message.

When no matching citation from \code{ref} can be found, \code{NA} is returned.
}
