% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_fz}
\alias{match_fz}
\title{Fuzzy (Approximate) String Matching}
\usage{
match_fz(x, table, method = "lcs", maxDist = 115, ...)
}
\arguments{
\item{x}{elements to be approximately matched: will be coerced to 
\code{character} unless it is a list consisting of \code{integer} vectors.}

\item{table}{lookup table for matching. Will be coerced to \code{character}
unless it is a list consting of \code{integer} vectors.}

\item{method}{Matching algorithm to use. See \code{\link[stringdist]{stringdist-metrics}}.}

\item{maxDist}{Elements in \code{x} will not be matched with elements of 
\code{table} if their distance is larger than \code{maxDist}. Note that the
maximum distance between strings depends on the method: it should always be
specified.}

\item{...}{arguments passed on to \code{\link[stringdist:amatch]{stringdist::amatch()}}}
}
\value{
A tibble with 3 columns:
\itemize{
\item \code{x}
\item \code{table_match}: the closest match of \code{x}
\item \code{dist}: the distance between x and its closest match (given the method
selected
}
}
\description{
Wraps \code{\link[stringdist:amatch]{stringdist::amatch()}} to perform "fuzzy" (approximate) string
matching while providing more informative output. Instead of an integer
vector of best match positions, this function returns a tibble with the
input, its corresponding best match, and the approximate string distance.
}
\section{NOTES}{

Fuzzy string matching is \emph{SLOW}. Expect this function to take >1 min for
comparisons of more than 500 values for all methods.

For comparison of citation titles specifically, the "lcs" method is faster
than "osa" and seems to work better. Based on light experimentation, a good
setting for \code{maxDist} value for citation titles is between 80-115.
}

