% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_paren_depth}
\alias{max_paren_depth}
\title{Calculate Maximum Parentheses Depth}
\usage{
max_paren_depth(x, unmatched_err = TRUE)
}
\arguments{
\item{x}{A character vector.}

\item{unmatched_err}{Whether to signal an error or return \code{NA} for strings
with unmatched parentheses,}
}
\value{
An integer vector of the same length as the input giving the maximum
depth of nested parentheses in each element. \code{NA} will bStrings with unmatched
parentheses will return \code{NA}
}
\description{
Calculates the maximum depth of nested parentheses in a string or character
vector.
}
\examples{
max_paren_depth(c("no parens", "a (1 deep)", "((a) and (b or (3 deep)))"))

# errs by default
tryCatch(
  max_paren_depths("unmatched ( paren"),
  error = function(e) print(e$message)
)

# allow `NA` output for unmatched parentheses
max_paren_depth("unmatched ( paren", unmatched_err = FALSE)
max_paren_depth(
  c("a (1 deep)", "((a) and (b or (3 deep)))", "unmatched ( paren"),
  unmatched_err = FALSE
)

}
\seealso{
Other general utilities: 
\code{\link{length_sort}()},
\code{\link{sandwich_text}()},
\code{\link{suggest_regex}()}
}
\concept{general utilities}
