% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{multimaps}
\alias{multimaps}
\title{Identify One-to-Multiple Mappings}
\usage{
multimaps(.df, .col1, .type, .col2, include_xrefs = TRUE)
}
\arguments{
\item{.df}{A data.frame with mappings.}

\item{.col1}{The column with \code{subject} mapping CURIEs (i.e. those being
tested; the "one" in the "one-to-multiple" test).}

\item{.type}{The column with the mapping predicate(s). This column should
be formatted as CURIEs but can include multiple delimited predicates.}

\item{.col2}{The column with \code{object} mapping CURIEs (i.e. those being
counted; the "multiple" in the "one-to-multiple" test).}

\item{include_xrefs}{Whether \code{oboInOwl:hasDbXrefs} should be included in
tests for "one-to-multiple" mappings, as a boolean (default: \code{TRUE}).
\code{skos:exactMatch} & \code{skos:closeMatch} mappings are always included.}
}
\description{
Identifies one-to-multiple mappings between two columns of CURIEs in a
data.frame.
}
\keyword{internal}
