% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{multimaps}
\alias{multimaps}
\title{Identify One-to-Multiple Mappings}
\usage{
multimaps(x, pred, y, include_xrefs = TRUE)
}
\arguments{
\item{x}{Vector with \code{subject} of mappings (i.e. those being tested; the
"one" in the "one-to-multiple" test).}

\item{pred}{Vector with predicate(s) of mappings. Predicate(s) should
be formatted as CURIEs but can include multiple delimited predicates.}

\item{y}{Vector with \code{object} of mappings (i.e. those being counted; the
"multiple" in the "one-to-multiple" test).}

\item{include_xrefs}{Whether \code{oboInOwl:hasDbXrefs} should be included in
tests for "one-to-multiple" mappings, as a boolean (default: \code{TRUE}).
\code{skos:exactMatch} & \code{skos:closeMatch} mappings are always included.}
}
\value{
A logical vector specifying the positions in \code{x} that map to
multiple values in \code{y}.
}
\description{
Identifies values in \code{x} that map to multiple values in \code{y} for
\code{skos:exactMatch} or \code{skos:closeMatch} (and optionally also
\code{oboInOwl:hasDbXref}) mapping predicates.
}
\keyword{internal}
