% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_prefix}
\alias{ns_prefix}
\title{Namespace Prefixes}
\format{
A character vector of 268 namespaces,
named by their corresponding prefix (e.g.
obo = "http://purl.obolibrary.org/obo/").
}
\source{
"\code{robot} (v1.9.0) via data-raw/ns_prefix.R; last updated 2022-07-22."
}
\usage{
ns_prefix
}
\description{
Prefixes for namespaces sourced from \code{robot} that cover most uses in
OBO Foundry ontologies, with a few additional prefixes and slight
modifications to match the prefixes as used in the Human Disease Ontology.
}
\keyword{datasets}
