% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_prefix}
\alias{ns_prefix}
\alias{not_obo_prefix}
\alias{obo_prefix}
\alias{obo_ont_prefix}
\alias{obo_prop_prefix}
\title{Namespace Prefixes}
\format{
Character vectors of namespaces named by their corresponding prefix:

\code{ns_prefix}: Superset of all 546 prefix/namespaces listed
below.

\code{not_obo_prefix}: 21 prefix/namespaces
outside but commonly used in the OBO Foundry
(e.g. dc = "http://purl.org/dc/elements/1.1/"),
or in federated SPARQL queries with the DO
(e.g. up = "http://purl.uniprot.org/core/").

\code{obo_prefix}: Superset of all 525 OBO Foundry
prefix/namespaces, including those listed below and, additionally, including
the general \code{obo} prefix.

\code{obo_ont_prefix}: 261 OBO Foundry ontology
\emph{primary} prefix/namespaces
(e.g. AAO = "http://purl.obolibrary.org/obo/AAO_").

\code{obo_prop_prefix}: \code{oboInOwl} (along with the common prefix variant
\code{oio}) and 262 OBO Foundry ontology \emph{property}
namespaces (\emph{these may not all be in actual use})
(e.g. aao = "http://purl.obolibrary.org/obo/aao#")
}
\source{
\verb{ROBOT export-prefixes} (v1.9.8) via \code{data-raw/ns_prefix.R}; last
updated 2025-12-09.
}
\usage{
ns_prefix

not_obo_prefix

obo_prefix

obo_ont_prefix

obo_prop_prefix
}
\description{
Prefixes for namespaces, primarily sourced from
\href{http://robot.obolibrary.org/}{\code{ROBOT}}, that cover most uses in OBO Foundry
ontologies, with slight modification to match the prefixes as used in the
Human Disease Ontology and a few additional prefixes for cross-references and
DO-related SPARQL query support. These prefixes \emph{should} be up-to-date with
the latest OBO Foundry ontologies (https://github.com/ontodev/robot/issues/51).
}
\keyword{datasets}
