% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_predicates.R
\name{num_val_predicates}
\alias{num_val_predicates}
\alias{is_positive}
\alias{is_negative}
\title{Number value predicates}
\usage{
is_positive(x)

is_negative(x)
}
\arguments{
\item{x}{vector to be tested}
}
\description{
These value predicates are designed to identify common values that appear
in numeric vectors.
}
