% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_url.R
\name{parse_try_url}
\alias{parse_try_url}
\title{Parse try_url HTTP response}
\usage{
parse_try_url(resp, include_raw_resp = TRUE)
}
\arguments{
\item{resp}{The HEAD response from \code{try_url()}.}

\item{include_raw_resp}{Whether the full raw response should be included in
the \code{tibble} as a list column, \code{TRUE} (default) or \code{FALSE}.}
}
\value{
\code{tibble} with columns \code{url}, \code{status}, \code{status_code}, and either \code{exception}
if an R exception occurred and the request was not executed or \code{redirect_url},
if the request was executed; also optionally including the full \code{response}.
}
\description{
Parse the HTTP HEAD response obtained from \code{\link[=try_url]{try_url()}}, including any
potential R errors.
}
\keyword{internal}
