% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_url.R
\name{parse_try_url}
\alias{parse_try_url}
\title{Parse try_url HTTP response}
\usage{
parse_try_url(resp, include_resp = TRUE, content = NULL)
}
\arguments{
\item{resp}{The response from \code{try_url()}.}

\item{include_resp}{Whether the full HTTP response should be included in
the \code{tibble} as a list column, \code{TRUE} (default) or \code{FALSE}.}

\item{content}{Arguments used to extract \code{GET} content, as a named list
matching \code{\link[httr:content]{httr::content()}} arguments (\code{as}, \code{type}, \code{encoding}, etc.).}
}
\value{
A \code{tibble} with columns \code{url}, \code{status}, \code{status_code}, \code{redirect_url} and,
if an R exception occurred, \code{exception} providing the exception's message.

\strong{For \code{GET} responses}, additionally includes a \code{content} list-column.

Optionally, includes the full HTTP \code{response} (default).
}
\description{
Parse an HTTP response from \code{\link[=try_url]{try_url()}}, including any potential R errors.
\emph{Currently only works for HEAD and GET requests.}
}
\keyword{internal}
