% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_url}
\alias{parse_url}
\title{Parse URL}
\usage{
parse_url(url, how = "basic")
}
\arguments{
\item{url}{One or more URLs, as a character vector.}

\item{how}{The approach to use for parsing, either "basic" or "complete".}
}
\value{
For "basic", a \link[tibble:tibble]{tibble}, with \code{scheme}, \code{domain}, and \code{path}
columns.

For "complete", a \code{parsed_url_tbl} with a column/list-column for every
component of the URL; in essence, the output of \code{\link[httr:parse_url]{httr::parse_url()}}
formatted as a \link[tibble:tibble]{tibble}.
}
\description{
Parses one or more URL(s) into its constituent components.
}
\section{Provenance}{

This \code{how = "basic"} code is originally from \code{robotstxt:::parse_url()} (MIT
license, \href{https://github.com/ropensci/robotstxt/blob/master/LICENSE}{copyright Peter Meissner}),
with slight modification.
}

\examples{
url <- c(
    "http://user:pass@www.sub.domain.com:80/path/page?query=value#fragment",
    "https://disease-ontology.org/community/publications",
    "http://books.google.com/books?id=Z_z1R9SO8iMC&pg=PA258#v=onepage&q=&f=false"
)

parse_url(url)
parse_url(url, how = "complete")

}
