% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{plot_citedby}
\alias{plot_citedby}
\title{Plot Publications Citing DO by Year}
\usage{
plot_citedby(
  file_in = "data/citedby/DO_citedby.csv",
  dir_out = "graphics/website",
  w = 8,
  h = 5.6
)
}
\arguments{
\item{file_in}{The path to the file containing the list of publications
citing the DO, as a string.}

\item{dir_out}{The directory where the plot .png should be saved to, as a
string. The file name will be "{date}-DO_cited_by_count.png".}

\item{w}{The width of the plot in inches, as an integer.}

\item{h}{The height of the plot in inches, as an integer.}
}
\description{
Plots the count of publications that cite the Human Disease Ontology by
year.
}
\section{Data Preparation}{

To prepare cited by data EXECUTE scripts/citedby_full_procedure.R.
}

