% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_citedby}
\alias{plot_citedby}
\title{Plot Publications Citing DO by Year}
\usage{
plot_citedby(
  out_dir = "graphics/website",
  color_set = c("#C45055", "#934FBB", "#95B1BB", "#83C85F", "#B9964B", "#4C3E45"),
  w = 6,
  h = 3.15
)
}
\arguments{
\item{out_dir}{The directory where the plot \code{"DO_cited_by_count.png"}
should be saved, as a string. If \code{NULL} the plot is not saved to disk.}

\item{color_set}{A set of 6 colors or the prefix of the color set to use from
\link{DO_colors}. Available sets include: "sat", "accent1", "accent2",
and "orange". The default and light versions of the specified color set
will be used.}

\item{w}{The width of the plot in inches, as numeric.}

\item{h}{The height of the plot in inches, as numeric.}
}
\description{
Plots the count of publications that cite the Human Disease Ontology by
year. Data is read from the official
\href{https://docs.google.com/spreadsheets/d/1wG-d0wt-9YbwhQTaelxqRzbm4qnu11WDM2rv3THy5mY/edit#gid=1972219724}{DO_uses}
Google Sheet used for tracking 'cited by' information.
}
\section{Data Preparation}{

To prepare data, execute \code{scripts/citedby_full_procedure.R}.
}

