% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_xref_counts}
\alias{plot_xref_counts}
\title{Plot Xref Counts}
\usage{
plot_xref_counts(
  data_file = "data/DO_release/cross_references.csv",
  out_dir = "graphics/website",
  w = 8,
  h = 5.6
)
}
\arguments{
\item{data_file}{The path to the file containing the latest DO xref counts,
as a string.}

\item{out_dir}{The directory where the plot \code{"DO_xref_count.png"}
should be saved, as a string. If \code{NULL} the plot is not saved to disk.}

\item{w}{The width of the plot in inches, as numeric.}

\item{h}{The height of the plot in inches, as numeric.}
}
\description{
Plots the count of cross-references by source in the Human Disease Ontology.
}
\section{Data Preparation}{

To prepare data, manually copy and paste stats from the Google Sheet
\href{https://docs.google.com/spreadsheets/d/1-ZSUH43MJloR2EsBqHpGeY6IfKG7Gt8KBcU5remnoGI/edit#gid=269344614}{DO_github_release_log}
OR manually copy & paste stats from xref.tsv produced by executing:

\if{html}{\out{<div class="sourceCode">}}\preformatted{robot query --input \{DO_repo\}/src/ontology/doid.owl \\
            --query \{DO_repo\}/src/sparql/build/all-xref-report.rq xref.tsv
}\if{html}{\out{</div>}}
}

