% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pub_id_match}
\alias{pub_id_match}
\title{Prioritized List of Publication IDs & Regex's for Matching}
\format{
A length-4 character vector of named regex's
prioritized as follows:
\describe{ pmid > pmcid > doi > scopus_eid }
}
\usage{
pub_id_match
}
\description{
Prioritized List of Publication IDs & Regex's for Matching
}
\keyword{datasets}
