% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\docType{data}
\name{pub_id_types}
\alias{pub_id_types}
\title{Prioritized List of Publication IDs}
\format{
An object of class \code{character} of length 4.
}
\usage{
pub_id_types
}
\description{
Prioritized List of Publication IDs
}
\keyword{datasets}
