% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{pubmed_summary}
\alias{pubmed_summary}
\title{Get PubMed Summary}
\usage{
pubmed_summary(
  id = NULL,
  web_history = NULL,
  config = NULL,
  version = "2.0",
  retmode = "json",
  always_return_list = TRUE,
  tidy = TRUE,
  ...
)
}
\arguments{
\item{id}{vector with unique ID(s) for records in database \code{db}.
In the case of sequence databases these IDs can take form of an
NCBI accession followed by a version number (eg AF123456.1 or AF123456.2)}

\item{web_history}{A web_history object}

\item{config}{vector configuration options passed to \code{httr::GET}}

\item{version, retmode, always_return_list}{preferred defaults are set but
included here for flexibility; see \code{\link[rentrez:entrez_summary]{rentrez::entrez_summary()}} for details}

\item{tidy}{TRUE to return a \link[tibble:tibble]{tibble} (default), or FALSE
to return an \code{esummary_list}}

\item{...}{character Additional terms to add to the request, see NCBI
documentation linked to in references for a complete list}
}
\description{
Retrieves PubMed Summary information for specified publications. Acts as a
simple wrapper around \link[rentrez:entrez_summary]{rentrez::entrez_summary} to allow larger number of
IDs as input (using \link[rentrez:entrez_post]{rentrez::entrez_post}).
}
