% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyDOID.R
\name{pyDOID}
\alias{pyDOID}
\title{pyDOID}
\description{
\href{https://github.com/allenbaron/pyDOID}{pyDOID} is a python package
developed with the same intent as \code{DO.utils}, to provide methods for
updating, maintaining, organizing, and analyzing the Human Disease Ontology.
\code{DO.utils}, powered by \href{https://github.com/rstudio/reticulate}{reticulate},
wraps \code{pyDOID} for functionality that is currently unavailable in R.
\code{pyDOID} provides two primary classes: the \code{DOrepo} class and the \code{owl.xml}
class. Classes and their methods are accessible via \code{$} notation.
}
\section{DOrepo class}{

The \code{DOrepo} class inherits the methods & subclasses of the
\href{https://gitpython.readthedocs.io/en/stable/tutorial.html#meet-the-repo-type}{GitPython.repo.Repo}
class and adds a few additional classes & methods, see \code{\link[=DOrepo]{DOrepo()}} for
details.
}

\section{owl.xml class}{

The \code{owl.xml} class thinly wraps a small set of python
\href{https://rdflib.readthedocs.io/en/stable/}{rdflib} methods to enable SPARQL
1.1 queries, see \code{\link[=owl_xml]{owl_xml()}} for details.
}

