% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alliance.R
\name{read_alliance}
\alias{read_alliance}
\title{Read Alliance .tsv.gz File}
\usage{
read_alliance(alliance_tsv)
}
\arguments{
\item{alliance_tsv}{path to Alliance .tsv or .tsv.gz file}
}
\value{
A dataframe.
}
\description{
Reads in a .tsv or .tsv.gz file from the Alliance of Genome Resources
as a tibble. It is recommended that Alliance files be downloaded using
\code{\link[=download_alliance_tsv]{download_alliance_tsv()}}.
}
\seealso{
Other Alliance functions: 
\code{\link{count_alliance_records}()},
\code{\link{download_alliance_tsv}()},
\code{\link{save_alliance_counts}()}
}
\concept{Alliance functions}
