% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\name{read_citedby}
\alias{read_citedby}
\title{Read Stored Cited By Info}
\usage{
read_citedby(dataset)
}
\arguments{
\item{dataset}{A string representing one of the datasets stored in Google
Sheets}
}
\description{
Read cited by information stored as Google Sheets.
}
\details{
The datasets currently available are:
\itemize{
\item pubmed: Detailed DO publication cited by list from PubMed (obtained via
Entrez API)
\item scopus: Detailed DO publication cited by list from Scopus (obtained via
Scopus Search API)
\item published_uses: Human-readable summary of merged DO publication cited by
lists, along with any curated use information
}
}
