% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_omim}
\alias{read_omim}
\title{Read OMIM Data}
\usage{
read_omim(file, keep_mim = c("#", "\%"), ...)
}
\arguments{
\item{file}{The path to a file (possibly compressed) with copy/pasted or
manually downloaded from https://omim.org/ (see "Manual Input Requirements"
for details), or downloaded with \code{\link[=download_omim]{download_omim()}}.}

\item{keep_mim}{[\strong{OMIM search data only}] The MIM symbols representing
the data types to keep, as a character vector, or \code{NULL} to retain all
(default: \code{"#"} and \code{"\%"}).

The \href{https://www.omim.org/help/faq#1_3}{OMIM} defined MIM symbols are:\tabular{ll}{
   MIM symbol \tab MIM type \cr
   \code{*} \tab gene \cr
   \code{+} \tab gene, includes phenotype \cr
   \verb{#} \tab phenotype \cr
   \verb{\%} \tab phenotype, unknown molecular basis \cr
   \code{^} \tab deprecated \cr
   \code{none} \tab phenotype, suspected/overlap \cr
}}

\item{...}{
  Arguments passed on to \code{\link[=read_delim_auto]{read_delim_auto}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
An \code{omim_tbl} (tibble) with an \code{omim} column containing OMIM CURIEs
as formatted in DO xrefs, followed by complete OMIM data arranged as seen on
omim.org for OMIM \strong{entries} (where possible). If the omim.org "Download as"
button was used to download the data, the \code{omim_tbl} will be additionally
modified based on the download type:
\itemize{
\item Search list download: Additional \code{omim_search} class and \code{search} column
containing the search used.
\item OMIM phenotypic series titles download: Additional \code{omim_PS_titles} class.
\item OMIM phenotypic series download: Additional \code{omim_PS} class and a row
representing the OMIM phenotypic series itself.
}

Output with columns typical OMIM phenotype entries, including \code{omim_PS}, will
have an additional \code{geno_inheritance} column containing a best guess at
inheritance from the GENO ontology. This simplifies adding inheritance as
logical subClassOf axioms supporting curation.

NOTE: OMIM phenotypic series on https://omim.org/ include the same data as
entries but column are ordered differently.
}
\description{
Reads and formats OMIM data copied or manually downloaded from
https://omim.org/, or downloaded with \code{\link[=download_omim]{download_omim()}} (permission
required), and appends columns to speed up subsequent curation activities.
}
\section{Manual Input Requirements}{

The \code{file} with OMIM data copied or downloaded must include headers at the
top. These data can be left \emph{as copied & pasted from omim.org} even if they
are not formatted correctly, as \code{read_omim()} will process and correct
headers, which includes fixing multi-line or misarranged column headers,
and will trim whitespace.
}

