% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_omim}
\alias{read_omim}
\title{Read Manually Copied OMIMPS data}
\usage{
read_omim(file, ...)
}
\arguments{
\item{file}{The path to a .tsv or .csv file (possibly compressed) with OMIMPS
data.}

\item{...}{
  Arguments passed on to \code{\link[=preprocess_omim_dl]{preprocess_omim_dl}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Properly formats OMIMPS data copied or downloaded from https://omim.org/ and
appends the following columns to speed up curation activities:
\itemize{
\item \code{omim}: properly formatted xref for the DO.
\item \code{geno_inheritance}: best guess at inheritance to add as logical subClassOf
axiom.
\item \code{tidy_label}: phenotype name without question mark
\item \code{provisional}: the status of this phenotype as provisional, as indicated by
the presence of a leading question mark.
}
}
\details{
As part of formatting, column misarrangements are corrected and whitespace
is trimmed.
}
\keyword{internal}
