% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pubmed_txt}
\alias{read_pubmed_txt}
\title{Read in PubMed Citations (from txt file)}
\usage{
read_pubmed_txt(file)
}
\arguments{
\item{file}{Path to .txt file; or another possible input to \code{readLines()}.}
}
\value{
A data.frame with a record number (\code{n}), identifiers (\code{pmid}, \code{pmcid},
\code{doi}), and the full citation (\code{citation}).
}
\description{
Reads PubMed text-format citations spanning multiple lines, usually
obtained by downloading a text file from PubMed as 'Summary (text)'.
}
