% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_keys.R, R/utils-tidyverse.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{set_entrez_key}
\alias{as_tibble}
\alias{replace_na}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rentrez}{\code{\link[rentrez]{set_entrez_key}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidyr}{\code{\link[tidyr]{replace_na}}}
}}

