% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat_doid}
\alias{reformat_doid}
\title{Reformat DOIDs}
\usage{
reformat_doid(x, to = "CURIE", allow_bare = FALSE)
}
\arguments{
\item{x}{A set of DOIDs, as a character vector.}

\item{to}{The format to convert the DOIDs to, as a string. All valid formats
are possible options: "CURIE" (default), "URI", "obo_CURIE", "basename".}

\item{allow_bare}{Whether bare numbers should be allowed as input, \code{TRUE}
or \code{FALSE} (default).}
}
\description{
Convert valid DOIDs and/or bare numbers to a specified DOID format. No
attempt is made to confirm bare numbers or DOIDs match existing DOIDs.
}
\examples{
x <- c(
    "http://purl.obolibrary.org/obo/DOID_0001816",
    "DOID:4",
    "obo:DOID_14566",
    "DOID_0040001"
)

reformat_doid(x, to = "URI")
reformat_doid(x, to = "CURIE")
reformat_doid(x, to = "obo_CURIE")
reformat_doid(x, to = "basename")

w_bare <- c(x, "0001816")
reformat_doid(w_bare, allow_bare = TRUE)

}
