% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-methods.R
\name{replace_na.list}
\alias{replace_na.list}
\title{[DEPRECATED] Replace NAs with specified value}
\usage{
\method{replace_na}{list}(data, replace, ...)
}
\arguments{
\item{data}{A data frame or vector.}

\item{replace}{If \code{data} is a data frame, \code{replace} takes a list of values,
with one value for each column that has \code{NA} values to be replaced.

If \code{data} is a vector, \code{replace} takes a single value. This single value
replaces all of the \code{NA} values in the vector.}

\item{...}{Additional arguments for methods. Currently unused.}
}
\description{
Replace all NAs in lists with specified value. This method of \code{replace_na}
will recurse into nested lists but will skip internal components that are
not lists or simple vectors themselves (e.g. data.frames, matrices, etc).
\strong{NOTE [REQUIRES \verb{tidyr <= 1.1.4}]:} Coercion will occur where vectors in
the list are of different types than \code{replace}, with either the vector or
\code{replace} being coercedaccording to type order:
logical < integer < numeric < complex < character < list.
}
\section{DEPRECATION NOTE}{

The "unsafe" type conversion allowed by \code{replace_na.list()} will not work
with \verb{tidyr v1.2.0}, which uses "safe" type conversion via
\code{vctrs::vec_cast()}. \code{replace_na.list()} will work with \verb{tidyr <= 1.1.4}.
If \code{replace_na.list()} is not frequently needed, it will be removed in
future updates.
}

