% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot.R
\name{robot}
\alias{robot}
\title{Execute Robot Commands}
\usage{
robot(..., .robot_path = NULL)
}
\arguments{
\item{...}{Command(s) passed on to ROBOT, including subcommand(s), either as
a single string or named strings, where the name corresponds to the long
or short form of the option and the value being the option value. Where
options do not use a value, use "". Subcommands should be unnamed.

\strong{Examples:}
\itemize{
\item Subcommands: \code{"query"}, \code{"export"}
\item Common options:
\itemize{
\item \code{"--input doid.owl"} or \code{input = "doid.owl"} (named argument).
\item \code{"-o result.owl"} (i.e. \code{--output}) or \code{o = "result.owl"} (named argument).
\item \code{"--remove-annotations"} (option of \code{annotate}) or
\code{"remove-annotations" = ""} (named argument form).
}
}}

\item{.robot_path}{The path to a ROBOT executable or .jar file, as a string.
When \code{NULL} (default), if a system ROBOT executable is available it will
be used, otherwise an error will be signaled.

\strong{NOTE:} \code{DO.utils} caches the last ROBOT used for future use.}
}
\description{
Light wrapper for OBO Foundry ROBOT program (OBO Foundry). See
\href{http://robot.obolibrary.org/}{ROBOT documentation} for information about
subcommands and arguments.
}
\section{ROBOT Setup}{

\itemize{
\item Requires installation of Java (supported versions are listed under the
"Getting Started" section at http://robot.obolibrary.org/).
\item Can use OBO Foundry ROBOT tool installed on the system path. This
may be achieved with \code{\link[=install_robot]{install_robot()}} or as described at
http://robot.obolibrary.org/.
\item Can also use a standalone robot.jar file (\code{.robot_path} must be specified)
in which case it is executed with \code{java -Xmx10G -jar} (10 GB memory).
\item On the first use of each R session, the ROBOT executable/.jar specified
will be tested. It must succeed prior to any command execution. On success,
\code{robot()} will announce the version & path being used. If at any time a new
\code{.robot_path} is specified, this previous version will be replaced with a new
announcement.
}
}

\section{Citation}{

R.C. Jackson, J.P. Balhoff, E. Douglass, N.L. Harris, C.J. Mungall, and
J.A. Overton. ROBOT: A tool for automating ontology workflows.
BMC Bioinformatics, vol. 20, July 2019.
}

