% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot_wrappers.R
\name{robot_query}
\alias{robot_query}
\title{Execute a SPARQL Query with ROBOT}
\usage{
robot_query(input, query, output, ...)
}
\arguments{
\item{input}{The path to an RDF/OWL file recognized by ROBOT, as a string.}

\item{query}{The text for or path to a valid SPARQL query (\code{ASK}, \code{SELECT},
\code{CONSTRUCT}, or \code{UPDATE}) as a string.}

\item{output}{The path where output will be written, as a string.}

\item{...}{Additional arguments to
\href{http://robot.obolibrary.org/query}{ROBOT query} formatted as described in
\code{\link[=robot]{robot()}}.}
}
\value{
\code{output} invisibly.
}
\description{
Wrapper for \code{robot("query", ...)} that accepts a file \emph{or} text query, and
has more convenient arguments.
}
\seealso{
\code{\link[=robot]{robot()}} for underlying implementation; \code{\link[=tidy_sparql]{tidy_sparql()}} for tidying
tabular output data read into R.
}
