% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot_wrappers.R
\name{robot_query}
\alias{robot_query}
\title{Execute a SPARQL Query with ROBOT}
\usage{
robot_query(input, query, output = NULL, ..., tidy_what = "nothing")
}
\arguments{
\item{input}{The path to an RDF/OWL file recognized by ROBOT, as a string.}

\item{query}{The text for or path to a valid SPARQL query (\code{ASK}, \code{SELECT},
\code{CONSTRUCT}, or \code{UPDATE}) as a string.}

\item{output}{The path where output will be written, as a string, or \code{NULL}
(default) to load data directly. \code{output} is required for \code{UPDATE} and
\code{CONSTRUCT} queries.}

\item{...}{Additional arguments to
\href{http://robot.obolibrary.org/query}{ROBOT query} formatted as described in
\code{\link[=robot]{robot()}}.}

\item{tidy_what}{The elements of a SPARQL-created data.frame to tidy, as a
character vector. One or more of the following:
\itemize{
\item \code{"everything"} to apply all tidy operations (has precedence over
\code{"nothing"}).
\item \code{"header"} to remove leading \verb{?} from header labels.
\item \code{"unnest"} to unnest list columns with \code{\link[=unnest_cross]{unnest_cross()}}.
\item \code{"uri_to_curie"} to convert all URIs recognized by DO.utils to CURIEs with
\code{\link[=to_curie]{to_curie()}}.
\item \code{"lgl_NA_false"} to replace \code{NA} in logical columns with \code{FALSE}.
\item \code{"as_tibble"} to make the output a \link[tibble:tibble]{tibble}.
\item \code{"nothing"} to prevent all tidying.
}}
}
\value{
If \code{output} is specified, the path to the output file with the query result.
Otherwise, the query result (ASK as boolean or SELECT as \code{tibble}).
}
\description{
Wrapper for \code{robot("query", ...)} that accepts a file \emph{or} text query, and
has more convenient arguments.
}
\seealso{
\code{\link[=robot]{robot()}} for underlying implementation.
}
