% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot_query.R
\name{robot_query}
\alias{robot_query}
\title{Execute Robot Query}
\usage{
robot_query(input, rq, save)
}
\arguments{
\item{input}{path to owl file to be queried}

\item{rq}{file with SPARQL query to execute}

\item{save}{path to save output file to}
}
\description{
Wrapper for system robot program that executes the query subcommand
}
\section{NOTE}{

Requires installation of the OBO Foundry tool
\href{http://robot.obolibrary.org/}{ROBOT} in the system path.
}

\section{Citation}{

R.C. Jackson, J.P. Balhoff, E. Douglass, N.L. Harris, C.J. Mungall, and
J.A. Overton. ROBOT: A tool for automating ontology workflows.
BMC Bioinformatics, vol. 20, July 2019.
}

