% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{round_down}
\alias{round_down}
\title{Round Number Down}
\usage{
round_down(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{The number of decimal places, as an integer. Negative
values are allowed (see 'Details').}
}
\description{
Round Number Down
}
\details{
Rounding to a negative number of digits means rounding to a power of ten, so
for example round_down(x, digits = -2) rounds to the nearest hundred.
}
