% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sandwich_text}
\alias{sandwich_text}
\title{Sandwich Text Between Placeholders}
\usage{
sandwich_text(x, placeholder, add_dup = TRUE)
}
\arguments{
\item{x}{A string or character vector.}

\item{placeholder}{One or two placeholders to sandwich each element of \code{x}
between. When two placeholders are provided, \code{x} will be sandwiched
between them with the first at the start and second at the end.
Otherwise, \code{x} will be sandwiched at both start and end by the same
placeholder.}

\item{add_dup}{Whether to add placeholders even if the same character is
already found in that position, as a boolean (default: \code{TRUE}).}
}
\description{
Sandwiches strings between one or two placeholders.
}
\examples{
sandwich_text("a", placeholder = "h")
sandwich_text("a", placeholder = c("b", "h"))
sandwich_text("bah", placeholder = c("b", "h"), add_dup = TRUE)
sandwich_text("bah", placeholder = c("b", "h"), add_dup = FALSE)
sandwich_text("bah", placeholder = "h", add_dup = FALSE)

}
\seealso{
Other general utilities: 
\code{\link{length_sort}()},
\code{\link{max_paren_depth}()},
\code{\link{suggest_regex}()}
}
\concept{general utilities}
