% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sandwich_text}
\alias{sandwich_text}
\title{Sandwich Text Between Placeholders}
\usage{
sandwich_text(x, placeholder, ...)
}
\arguments{
\item{x}{A string or character vector.}

\item{...}{
  Arguments passed on to \code{\link[base:paste]{base::paste0}}
  \describe{
    \item{\code{collapse}}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}
    \item{\code{recycle0}}{\code{\link[base]{logical}} indicating if zero-length
    character arguments should lead to the zero-length
    \code{\link[base]{character}(0)} after the \code{sep}-phase (which turns into
    \code{""} in the \code{collapse}-phase, i.e., when \code{collapse} is
    not \code{NULL}).}
  }}

\item{placholder}{One or two placeholders to sandwich each element of \code{x}
between. When two placeholders are provided, \code{x} will be sandwiched
between them with the first on the left and second on the right.
Otherwise, \code{x} will be sandwiched on both sides by the same placeholder.}
}
\description{
Sandwiches strings between one or two placeholders.
}
\examples{
sandwich_text("a", placeholder = "h")
sandwich_text("a", placedholder = c("b", "h"))

}
\concept{general utilities}
