% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{set_html_attr}
\alias{set_html_attr}
\title{Set HTML Attributes}
\usage{
set_html_attr(..., quote = "\\"")
}
\arguments{
\item{...}{Named character vector(s) or name-value pairs of HTML attributes.
Attributes with values of \code{NULL} or \code{NA} will be dropped. Names should
correspond exactly to desired HTML attributes.}

\item{quote}{The quote with which to surround the attribute values. Use \code{""}
if quotes are not desired (default: \verb{"\\""}, double quote).}
}
\value{
A character vector of HTML attribute strings including necessary
quotes and with a leading space, e.g.
\code{' src="img path" alt="img alt text here"'}.
}
\description{
Sets HTML attributes for a single HTML element.
}
\keyword{internal}
