% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{set_html_attr}
\alias{set_html_attr}
\title{Set HTML Attributes}
\usage{
set_html_attr(attr)
}
\arguments{
\item{attr}{A named character vector of HTML attributes. Attributes with
\code{NULL} or \code{NA} values will be dropped. Names should correspond to HTML
attributes.}
}
\value{
A character vector of HTML attribute strings including necessary
quotes and with a leading space, e.g.
\code{' src="img path" alt="img alt text here"'}.
}
\description{
Sets HTML attributes for a single HTML element.
}
\keyword{internal}
