% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suggest_regex}
\alias{suggest_regex}
\title{Suggest a Regular Expression That Will Match All Input}
\usage{
suggest_regex(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A string with a bracketed character set for each position.
}
\description{
Collects the full set of characters found at each position across all strings
in \code{x} and returns it as a quasi-regular expression. Letter and numbers will
not be condensed to ranges in output, even if the full sets are present at a
position.
}
\examples{
x <- c("DNA", "MHC", "TAP1", "TAP2", "520", "ACD")

suggest_regex(x)

}
