% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_sparql}
\alias{tidy_sparql}
\title{Tidy SPARQL Query}
\usage{
tidy_sparql(query_res)
}
\arguments{
\item{query_res}{The results of a SPARQL query, as a data.frame (usually
produced by \href{owl_xml()}{owl_xml()$query()} or
\href{DOrepo()}{DOrepo()$doid{_merged}?$query()}).}
}
\description{
Tidies SPARQL query results, unnesting list columns and returning results as
a \href{tibble::tibble()}{tibble} instead of a data.frame.
}
\section{Note}{

This function exists because the results are not currently tidied by \code{pyDOID}
(the python package that provides SPARQL query functionality to \code{DO.utils}).
}

