% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_sparql}
\alias{tidy_sparql}
\title{Tidy SPARQL Query}
\usage{
tidy_sparql(query_res, tidy_what = "everything", ...)
}
\arguments{
\item{query_res}{The results of a SPARQL query, as a data.frame (usually
produced by \code{\link[=owl_xml]{owl_xml()}}$query() or similar from \code{\link[=DOrepo]{DOrepo()}}, but can also
be used to tidy results of \link[=robot]{robot("query", ...)} loaded with
\code{readr}).}

\item{tidy_what}{The elements of a SPARQL-created data.frame to tidy, as a
character vector. One or more of the following:
\itemize{
\item \code{"everything"} to apply all tidy operations (has precedence over
\code{"nothing"}).
\item \code{"header"} to remove leading \verb{?} from header labels.
\item \code{"unnest"} to unnest list columns with \code{\link[=unnest_cross]{unnest_cross()}}.
\item \code{"uri_to_curie"} to convert all URIs recognized by DO.utils to CURIEs with
\code{\link[=to_curie]{to_curie()}}.
\item \code{"lgl_NA_FALSE"} to replace \code{NA} in logical columns with \code{FALSE}.
\item \code{"as_tibble"} to make the output a \link[tibble:tibble]{tibble}.
\item \code{"rm_lang_tag"} to remove language tags. Tags will only be removed from
\code{character} class columns, and then only if there is one unique language tag
in the given column.
\item \code{"nothing"} to prevent all tidying.
}}

\item{...}{
  Arguments passed on to \code{\link[=to_curie]{to_curie}}
  \describe{
    \item{\code{strip_angle_brackets}}{Whether to remove all \code{<} and \code{>} from the
input, as a boolean (default: \code{TRUE}). Where this might have undesirable
consequences, e.g. \emph{some} angle brackets need to be removed, perform the
removal beforehand and use \code{strip_angle_brackets = FALSE}.}
  }}
}
\description{
Tidies SPARQL query results according to desired specifications (see
\code{tidy_what} parameter for details).
}
