% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_to_string.R
\name{to_character}
\alias{to_character}
\title{Convert to Character}
\usage{
to_character(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Wrapper for \code{\link[base:character]{base::as.character()}} designed to provide extra methods for converting
lists and data.frames to character vectors. \code{to_character()} is intended to
function as an \emph{internal} helper of \code{\link[=cast_to_string]{cast_to_string()}}.
}
