% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{to_curie}
\alias{to_curie}
\title{Convert URI to CURIE}
\usage{
to_curie(x, strip_angle_brackets = TRUE)
}
\arguments{
\item{x}{URI(s), as a character vector.}

\item{string_angle_brackets}{Whether to remove all \code{<} and \code{>} from the
input, as a boolean (default: \code{TRUE}). Where this might have undesirable
consequences, e.g. \emph{some} angle brackets need to be removed, perform the
removal beforehand and use \code{strip_angle_brackets = FALSE}.}
}
\description{
Converts URI(s) to CURIE(s).
}
\section{Note}{

Performs no URI validation, relying on simple string matching of
namespace-prefix pairs of \link{ns_prefix} for conversion. Any values not
matching one of these will be returned without modification.
}

\examples{
.uri <- c(
    "http://www.w3.org/2000/01/rdf-schema#comment",
    "http://purl.org/dc/elements/1.1/date",
    "http://purl.org/dc/terms/license",
    "http://www.w3.org/2002/07/owl#deprecated",
    "http://www.geneontology.org/formats/oboInOwl#id",
    "http://purl.obolibrary.org/obo/UBERON_0000002",
    "http://purl.obolibrary.org/obo/DOID_0001816",
    "http://purl.obolibrary.org/obo/doid#DO_AGR_slim"
    "<http://www.geneontology.org/formats/oboInOwl#hasDbXref>"
)
to_curie(.uri)

# uses 'obo' namespace when an OBO Foundry ontology namespace isn't available
to_curie(
    c("http://purl.obolibrary.org/obo/SO_0000110",
    "http://purl.obolibrary.org/obo/so#has_origin")
)

#returns non-URI or unknown namespace prefixes unmodified
to_curie(
    c("http://purl.obolibrary.org/obo/SYMP_0000000",
    "not a URI", "https://disease-ontology.org/")
)

}
\seealso{
Other identifier converters: 
\code{\link{to_uri}()}
}
\concept{identifier converters}
