% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_url.R
\name{try_url}
\alias{try_url}
\title{Try a URL}
\usage{
try_url(url, type = "HEAD", config = httr::user_agent(pkg_user_agent), ...)
}
\arguments{
\item{url}{URL to try, as a string.}

\item{type}{The HTTP request type, as a string.}

\item{...}{Arguments passed on to the corresponding \href{httr::httr}{httr}
request function.}
}
\description{
Try to reach a URL with a request of the specified type, capturing R
errors/warnings if they occur.
}
\keyword{internal}
