% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr_to_scalar.R
\name{unique_if_invariant}
\alias{unique_if_invariant}
\alias{unique_if_invariant.character}
\alias{unique_if_invariant.numeric}
\title{Return Unique Value for Invariant Vectors}
\usage{
unique_if_invariant(x, na.rm = FALSE, ...)

\method{unique_if_invariant}{character}(x, na.rm = FALSE, ...)

\method{unique_if_invariant}{numeric}(x, na.rm = FALSE, tol = sqrt(.Machine$double.eps), ...)
}
\arguments{
\item{x}{vector to be tested}

\item{na.rm}{logical indicating whether to exclude NA values}

\item{...}{unused; for extensibility}

\item{tol}{double, tolerance to use (for numeric vectors)}
}
\description{
Returns the unique value from a vector of any length, if and only if, only 1
unique value exists (\emph{i.e.} the vector is invariant), otherwise returns the
original vector.
}
\seealso{
Other vector-to-scalar functions: 
\code{\link{vctr_to_string}()}
}
\concept{vector-to-scalar functions}
