% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_if_invariant.R
\name{unique_if_invariant}
\alias{unique_if_invariant}
\title{Collapse invariant vectors}
\usage{
unique_if_invariant(x)
}
\arguments{
\item{x}{vector}
}
\description{
Returns the unique value from a vector of any length if and only if only 1
unique value exists, otherwise returns the original vector
}
