% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_url.R
\name{validate_url}
\alias{validate_url}
\title{Validate a URL}
\usage{
validate_url(url, include_resp = TRUE, ...)
}
\arguments{
\item{url}{The url to test, as a string.}

\item{include_resp}{Whether the full HTTP response should be included in
the \code{tibble} as a list column, \code{TRUE} (default) or \code{FALSE}.}

\item{...}{Further named parameters, such as \code{query}, \code{path}, etc,
passed on to \code{\link[httr:modify_url]{modify_url()}}. Unnamed parameters will be combined
with \code{\link[httr:config]{config()}}.}
}
\value{
See \code{\link[=parse_try_url]{parse_try_url()}} documentation.
}
\description{
Test a URL with a HEAD request to determine if it is valid.
}
