% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr_to_scalar.R
\name{vctr_to_string}
\alias{vctr_to_string}
\title{Convert Vectors to Strings}
\usage{
vctr_to_string(x, delim = "; ", na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{delim}{A delimiter to place between vector elements (default: "; ").}

\item{na.rm}{A logical scalar indicating whether \code{NA} values should be
removed (default: \code{FALSE}).}
}
\description{
Concatenate values in a vector into a single string.
}
\seealso{
\code{\link[=unique_if_invariant]{unique_if_invariant()}} for an alternative, \emph{conditional} \code{unique}
vector-to-string conversion method
}
