% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr_to_scalar.R
\name{vctr_to_string}
\alias{vctr_to_string}
\alias{unique_to_string}
\title{Convert Vectors to Strings}
\usage{
vctr_to_string(
  x,
  delim = "|",
  na.rm = FALSE,
  sort = FALSE,
  decreasing = FALSE,
  ...
)

unique_to_string(
  x,
  delim = "|",
  na.rm = FALSE,
  sort = FALSE,
  decreasing = FALSE,
  ...
)
}
\arguments{
\item{x}{A vector.}

\item{delim}{A delimiter to place between vector elements (default: "|").}

\item{na.rm}{A logical scalar indicating whether \code{NA} values should be
removed (default: \code{FALSE}).}

\item{sort}{A logical scalar indicating whether values should be sorted
(default: \code{FALSE}).}

\item{decreasing}{logical.  Should the sort be increasing or decreasing?
    Not available for partial sorting.}

\item{...}{arguments to be passed to or from methods or (for the
    default methods and objects without a class) to \code{sort.int}.}
}
\description{
Concatenate values in a vector into a single string.
* \code{vctr_to_string} performs simple concatenation.
* \code{unique_to_string} reduces the vector to unique values prior to
concatenation.
}
\seealso{
\code{\link[=unique_if_invariant]{unique_if_invariant()}} for an alternative, \emph{conditional} \code{unique}
vector-to-string conversion method
}
