% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_graphml}
\alias{write_graphml}
\title{Write a Graph to .graphml File}
\usage{
write_graphml(graph, file)
}
\arguments{
\item{graph}{A \href{tidygraph::tidygraph-package}{tidygraph} or
\href{igraph::igraph-package}{igraph} object.}

\item{file}{The file path to write to, as a string. If '.graphml' extension
is missing, it will be appended.}
}
\value{
Absolute path of .graphml file written.
}
\description{
Writes a graph object (tidygraph/iGraph) to a file as
\href{https://en.wikipedia.org/wiki/GraphML}{GraphML}.
}
