% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_gs}
\alias{write_gs}
\alias{write_gs.omim_inventory}
\title{Write Data to a Google Sheet}
\usage{
write_gs(data, ss, hyperlink_curie = NULL)

\method{write_gs}{omim_inventory}(data, ss, hyperlink_curie = c("omim", "doid"))
}
\arguments{
\item{data}{A specially-classed data.frame with a defined method.}

\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:drive_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}.}

\item{hyperlink_curie}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> The
columns with CURIEs to convert to hyperlinks when written in Google Sheets.}
}
\value{
The
}
\description{
Specialized methods for writing data created by \code{DO.utils} to a specified
Google Sheet.
}
