% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_gs}
\alias{write_gs}
\alias{write_gs.omim_inventory}
\title{Write Data to a Google Sheet}
\usage{
write_gs(data, ss, sheet = NULL, hyperlink_curie = NULL, ...)

\method{write_gs}{omim_inventory}(
  data,
  ss,
  sheet = "omim_inventory-\%Y\%m\%d",
  hyperlink_curie = c("omim", "doid"),
  datestamp = "\%Y\%m\%d",
  ...
)
}
\arguments{
\item{data}{A specially-classed data.frame with a defined method.}

\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:drive_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{The name to use for the sheet to write into, in the sense of
"worksheet" or "tab". If a date format recognized by \code{\link[=format.Date]{format.Date()}} is
included in the string, today's date will be added in the specified format
and location. If \code{NULL}, the sheet name will be the default for the
underlying method.

\strong{\emph{WARNING:}} If a sheet with the same name exists it will be overwritten.}

\item{hyperlink_curie}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> The
columns with CURIEs to convert to hyperlinks when written in Google Sheets.}

\item{...}{Arguments passed on to methods.}

\item{datestamp}{\strong{DEPRECATED} Use \code{sheet} instead.

Previously \code{NULL} or \code{NA} would default to the sheet name 'omim_inventory',
while a format recognized by \code{\link[=format.Date]{format.Date()}} would add an additional date
stamp suffix, separated by '-', to the default sheet name. This behavior is
retained for backward compatibility and invoked when \code{sheet = NULL}.}
}
\value{
The data as written to the Google Sheet, invisibly.
}
\description{
Specialized methods for writing data created by \code{DO.utils} to a specified
Google Sheet.
}
